% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_categorise.R
\name{categorise_headsize}
\alias{categorise_headsize}
\title{Categorise head circumference-for-age z-scores into head
circumference-for-age strata}
\usage{
categorise_headsize(hcaz)
}
\arguments{
\item{hcaz}{A numeric vector of length one or more containing
head circumference-for-age z-scores (HCAZs).}
}
\value{
An object of class factor with the same length as \code{hcaz}, containing
head circumference-for-age classifications. Its levels are
\code{c("microcephaly_severe", "microcephaly", "normal_headcirc", "macrocephaly", "macrocephaly_severe")}.
}
\description{
Categorise head circumference-for-age z-scores into head
circumference-for-age strata
}
\details{
Cut-offs for head size categories are:

\tabular{lll}{
\strong{Category}    \tab \strong{Factor level}  \tab
\strong{Z-score bounds} \cr
Severe microcephaly       \tab \code{"microcephaly_severe"} \tab \code{hcaz} =< -3      \cr
Microcephaly              \tab \code{"microcephaly"}        \tab -3 < \code{hcaz} =< -2 \cr
Normal head circumference \tab \code{"normal_headcirc"}      \tab \code{abs(hcaz)} < 2 \cr
Macrocephaly              \tab \code{"macrocephaly"}        \tab \code{hcaz} >= 2       \cr
Severe macrocephaly       \tab \code{"macrocephaly_severe"} \tab \code{hcaz} >= 3
}
}
\examples{
hcaz <- c(-6.5, -3.5, -2.5, 0, 2.5, 3.5)
categorise_headsize(hcaz)
}
\references{
Victora CG, Schuler-Faccini L, Matijasevich A, Ribeiro E, Pessoa A,
Barros FC. \strong{Microcephaly in Brazil: how to interpret reported numbers?}
\emph{The Lancet} 2016, \emph{387(10019):621-624} \doi{10.1016/S0140-6736(16)00273-7}

Accogli A, Geraldo AF, Piccolo G, Riva A, Scala M, Balagura G, et al.
\strong{Diagnostic Approach to Macrocephaly in Children}. \emph{Frontiers in
Paediatrics} 2022, \emph{9:794069} \doi{10.3389/fped.2021.794069}
}
