% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_categorise.R
\name{categorise_sfga}
\alias{categorise_sfga}
\title{Categorise birthweight centiles into size-for-gestational age strata}
\usage{
categorise_sfga(p, severe = FALSE)
}
\arguments{
\item{p}{A numeric vector with birth weight centiles from the
INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Newborn Size standard. If elements of \code{p} are
not between \code{0} and \code{1}, gigs will warn you. You can customise this
behaviour with the \link[=gigs_options]{GIGS package-level options}.}

\item{severe}{A single logical value specifying whether to categorise SGA
values below the third centile as \code{"SGA(<3)"}. Default = \code{FALSE}.}
}
\value{
An object of class factor with the same length as \code{p}, containing
size-for-GA classifications. If \code{severe = FALSE}, levels are \code{c("SGA", "AGA", "LGA")}. If \code{severe = TRUE}, levels are \code{c("SGA(<3)", "SGA", "AGA", "LGA")}.
}
\description{
Categorise birthweight centiles into size-for-gestational age strata
}
\details{
Cut-offs for size-for-gestational age categorisations are:

\tabular{lll}{
\strong{Category} \tab \strong{Factor level} \tab
\strong{Centile bounds} \cr
Severely small-for-GA \tab \code{"SGA(<3)"} \tab \code{p} < 0.03      \cr
Small-for-GA          \tab \code{"SGA"}     \tab \code{p} < 0.1       \cr
Appropriate-for-GA    \tab \code{"AGA"}     \tab 0.1 =< \code{p} =< 0.9 \cr
Large-for-GA          \tab \code{"LGA"}     \tab \code{p} > 0.9
}
}
\note{
The returned factor will have unused levels if none of your observed
data fit given categories. To drop these, you will need to use
\code{\link[=droplevels]{droplevels()}} or similar.
}
\examples{
p <- c(0.01, 0.07, 0.25, 0.75, 0.93, 0.99)
categorise_sfga(p = p, severe = FALSE)
categorise_sfga(p = p, severe = TRUE)
}
\references{
WHO. \strong{Physical status: the use and interpretation of anthropometry. Report
of a WHO Expert Committee.} \emph{World Health Organisation Technical Report
Series 1995,} \strong{854: 1–452}

Royal College of Obstetricians and Gynaecologists. \strong{The Investigation and
Management of the Small-for-Gestational-Age Fetus: Green-top Guideline No.
31.} \emph{Technical report, Royal College of Obstetricians and Gynaecologists,
London, 2013.}
}
\seealso{
\code{\link[=classify_sfga]{classify_sfga()}}, which wraps this function for easy use in
\code{data.frame}-based analytic pipelines.
}
