% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_categorise.R
\name{categorise_stunting}
\alias{categorise_stunting}
\title{Categorise length/height-for-age z-scores into stunting strata}
\usage{
categorise_stunting(lhaz, outliers = FALSE)
}
\arguments{
\item{lhaz}{A numeric vector of length one or more containing
length/height-for-age z-scores (LHAZs).}

\item{outliers}{A single \code{TRUE} or \code{FALSE} value specifying whether
implausible z-score thresholds should be applied. Default = \code{FALSE}.}
}
\value{
An object of class factor with the same length as the longest input
vector, containing stunting classifications. Its levels are
\code{c("stunting_severe", "stunting", "not_stunting")} if \code{outliers = FALSE} (the default), else \code{c("stunting_severe", "stunting", "not_stunting", "outlier")}.

\tabular{lll}{
\strong{Category} \tab \strong{Factor level} \tab
\strong{Z-score bounds} \cr
Severe stunting \tab \code{"stunting_severe"}     \tab \code{lhaz} ≤ -3         \cr
Stunting        \tab \code{"stunting"}            \tab -3 < \code{lhaz} ≤ -2    \cr
No stunting     \tab \code{"not_stunting"}        \tab \code{lhaz} > -2         \cr
Outlier         \tab \code{"outlier"}             \tab \code{abs(lhaz)} > 6
}
}
\description{
Categorise length/height-for-age z-scores into stunting strata
}
\note{
This function assumes that your measurements were taken according to
WHO guidelines, which stipulate that recumbent length should not be
measured after 730 days. Instead, standing height should be used.
Implausible z-score bounds are sourced from the referenced WHO report, and
classification cut-offs from the DHS manual.
}
\examples{
lhaz <- c(-6.5, -5, -3, 0, 3, 5, 6.5)
categorise_stunting(lhaz, outliers = FALSE)
categorise_stunting(lhaz, outliers = TRUE)
}
\references{
\strong{'Implausible z-score values'} \emph{in} World Health Organization (ed.)
\emph{Recommendations for data collection, analysis and reporting on
anthropometric indicators in children under 5 years old}. Geneva: World
Health Organization and the United Nations Children's Fund UNICEF, (2019).
pp. 64-65.

\strong{'Percentage of children stunted, wasted, and underweight, and mean z-scores
for stunting, wasting and underweight'} \emph{in} \emph{Guide to DHS Statistics DHS-7}
Rockville, Maryland, USA: ICF (2020). pp. 431-435.
\url{https://dhsprogram.com/data/Guide-to-DHS-Statistics/Nutritional_Status.htm}
}
