% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_categorise.R
\name{categorise_svn}
\alias{categorise_svn}
\title{Categorise birthweight centiles and gestational ages into small vulnerable
newborn strata}
\usage{
categorise_svn(p, gest_days)
}
\arguments{
\item{p}{A numeric vector with birth weight centiles from the
INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Newborn Size standard. If elements of \code{p} are
not between \code{0} and \code{1}, gigs will warn you. You can customise this
behaviour with the \link[=gigs_options]{GIGS package-level options}.}

\item{gest_days}{A numeric vector of length one or more with gestational
age(s) in days. Used to determine whether infants are term or preterm (a
gestational age > 259 days means an infant is term).}
}
\value{
An object of class factor with the same length as the longest input
vector, containing small vulnerable newborn classifications. Its levels
are \code{c("Preterm SGA", "Preterm AGA", "Preterm LGA", "Term SGA", "Term AGA", "Term LGA")}.
}
\description{
Categorise birthweight centiles and gestational ages into small vulnerable
newborn strata
}
\details{
Cut-offs for small, vulnerable newborn categorisations are:

\tabular{llll}{
\strong{SVN category} \tab \strong{Factor level} \tab
\strong{Gestational age range} \tab \strong{Centile bounds} \cr
Preterm-SGA \tab \code{"Preterm SGA"} \tab \code{p} < 0.1       \tab
\code{gest_days} < 259  \cr
Preterm-AGA \tab \code{"Preterm AGA"} \tab 0.1 ≤ \code{p} ≤ 0.9 \tab
\code{gest_days} < 259  \cr
Preterm-LGA \tab \code{"Preterm LGA"} \tab \code{p} > 0.9       \tab
\code{gest_days} < 259  \cr
Term-SGA \tab \code{"Term SGA"}    \tab \code{p} < 0.1       \tab
\code{gest_days} ≥ 259  \cr
Term-AGA \tab \code{"Term AGA"}    \tab 0.1 ≤ \code{p} ≤ 0.9 \tab
\code{gest_days} ≥ 259  \cr
Term-LGA \tab \code{"Term LGA"}    \tab \code{p} > 0.9       \tab
\code{gest_days} ≥ 259
}

\emph{Abbreviations:} SGA, small-for-gestational age; AGA,
appropriate-for-gestational age; LGA, large-for-gestational age.
}
\note{
Input vectors are recycled by \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}, and must
adhere to the \link{vctrs} recycling rules.
}
\examples{
p <- c(0.01, 0.07, 0.25, 0.75, 0.93, 0.99)
gest_days <- c(250, 250, 250, 280, 280, 280) # 3 preterm, 3 term
categorise_svn(p = p, gest_days = gest_days)
}
\references{
Lawn JE, Ohuma EO, Bradley E, Idueta LS, Hazel E, Okwaraji YB et al.
\strong{Small babies, big risks: global estimates of prevalence and mortality for
vulnerable newborns to accelerate change and improve counting.} \emph{The Lancet}
2023, \emph{401(10389):1707-1719.} \doi{10.1016/S0140-6736(23)00522-6}
}
\seealso{
\code{\link[=classify_svn]{classify_svn()}}, which wraps this function for easy use in
analytic pipelines. Also check out \code{\link[=categorise_sfga]{categorise_sfga()}}, which this function
uses to stratify \code{p} into size-for-gestational age categories.
}
