% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_classify.R
\name{classify_growth}
\alias{classify_growth}
\title{Classify multiple growth indicators at the same time using international
growth standards}
\usage{
classify_growth(
  .data,
  gest_days,
  age_days,
  sex,
  weight_kg = NULL,
  lenht_cm = NULL,
  headcirc_cm = NULL,
  .analyses = c("sfga", "svn", "stunting", "wasting", "wfa", "headsize"),
  .new = list(sfga = c("birthweight_centile", "sfga", "sfga_severe"), svn =
    c("birthweight_centile", "svn"), stunting = c("lhaz", "stunting",
    "stunting_outliers"), wasting = c("wlz", "wasting", "wasting_outliers"), wfa =
    c("waz", "wfa", "wfa_outliers"), headsize = c("hcaz", "headsize")),
  .verbose = TRUE
)
}
\arguments{
\item{.data}{A \code{data.frame}-like tabular object with one or more rows. Must
be permissible by \code{\link[checkmate:checkDataFrame]{checkmate::assert_data_frame()}}, so you can also supply
a \code{tibble}, \code{data.table}, or similar.}

\item{gest_days}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of gestational age(s) at birth
in days. This column, in conjunction with the column referred to by
\code{age_days}, is used to select which growth standard to use for each
observation.}

\item{age_days}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of age values in days.}

\item{sex}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a case-sensitive character vector of sexes,
either \code{"M"} (male) or \code{"F"} (female). By default, gigs will warn you if
any elements of \code{sex} are not \code{"M"} or \code{"F"}, or are missing (\code{NA}). You
can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}

\item{weight_kg}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of weight values in kg.}

\item{lenht_cm}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of length/height values in cm.}

\item{headcirc_cm}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of head circumference values in
cm.}

\item{.analyses}{A character vector of up to six elements in length
describing which growth analyses you want to run. Use this when you supply
data that can be used to generate multiple growth indicators, but only want
to run a specific set. This argument is case-sensitive, and an error will
be thrown if you supply a string in \code{.analyses} which is either \code{NA} or is
not a member of this argument's default. Default = \code{c("sfga", "svn", "stunting", "wasting", "wfa", "headsize")}.}

\item{.new}{A list with names corresponding to \code{.analyses}, which describes
the names of new columns to be added to \code{.data}. If any elements of the
vectors in \code{.new} are equal to existing elements of \code{colnames(.data)}, the
function will throw an error. Excluding the centiles produced for
size-for-gestational age (\code{"sfga"}) and small vulnerable newborn (\code{"svn"})
classification, all strings in \code{.new} must be unique. Each character vector
in \code{.new} is repaired with \code{vctrs::as_names()}, which will issue messages
if elements in \code{.new} are changed.

\itemize{
\item{\code{"sfga"}:}{  \code{c("birthweight_centile", "sfga", "sfga_severe")}}
\item{\code{"svn"}:}{  \code{c("birthweight_centile", "svn")}}
\item{\code{"stunting"}:}{  \code{c("lhaz", "stunting", "stunting_outliers")}}
\item{\code{"wasting"}:}{  \code{c("wlz", "wasting", "wasting_outliers")}}
\item{\code{"wfa"}:}{  \code{c("waz", "wfa", "wfa_outliers")}}
\item{\code{"headsize"}:}{  \code{c("hcaz", "headsize")}}
}}

\item{.verbose}{A single logical value. When \code{TRUE} (the default),
messages about which analyses were requested versus which were performed
will be printed to the console. Warnings from \code{classify_growth()} will
still be printed even if \code{.verbose} is \code{FALSE}.}
}
\description{
This function permits classification of multiple growth indicators (stunting,
wasting, weight-for-age, and more) at once in \code{data.frame}-like objects.
}
\note{
For size-for-GA and small vulnerable newborn analyses, centiles and
categorisations will only be applied on birthweights, i.e. rows where
the column referred to by \code{weight_kg} has an age between \code{0} and \code{0.5}.
}
\examples{
data <- data.frame(
  agedays = c(0, 100, 100),
  gestage = c(270, 270, 270),
  wtkg = c(2.5, 7.5, 7.5),
  len_cm = c(45, 60, 60),
  head_cm = c(36, 40.2, 40.2),
  sex = c("M", "M", "F")
)

data_classified <- classify_growth(data,
                                   weight_kg = wtkg,
                                   age_days = agedays,
                                   gest_days = gestage,
                                   sex = sex)

data_classified

# Use `.analyses` to set which growth indicators will be computed
data_svn <- classify_growth(data,
                            weight_kg = wtkg,
                            age_days = agedays,
                            gest_days = gestage,
                            sex = sex,
                            .analyses = "svn")

data_svn

# Use `.new` to set new column names
data_svn <- classify_growth(data,
                            weight_kg = wtkg,
                            age_days = agedays,
                            gest_days = gestage,
                            sex = sex,
                            .analyses = "svn",
                            .new = list("svn" = c("ig_nbs_centile",
                                                  "SVN_Category")))
data_svn
}
\references{
WHO. \strong{Physical status: the use and interpretation of anthropometry. Report
of a WHO Expert Committee.} \emph{World Health Organisation Technical Report
Series 1995,} \strong{854: 1–452}

Royal College of Obstetricians and Gynaecologists. \strong{The Investigation and
Management of the Small-for-Gestational-Age Fetus: Green-top Guideline No.
31.} \emph{Technical report, Royal College of Obstetricians and Gynaecologists,
London, 2013.}

Lawn JE, Ohuma EO, Bradley E, Idueta LS, Hazel E, Okwaraji YB et al.
\strong{Small babies, big risks: global estimates of prevalence and mortality for
vulnerable newborns to accelerate change and improve counting.} \emph{The Lancet}
2023, \emph{401(10389):1707-1719.} \doi{10.1016/S0140-6736(23)00522-6}

\strong{'Implausible z-score values'} \emph{in} World Health Organization (ed.)
\emph{Recommendations for data collection, analysis and reporting on
anthropometric indicators in children under 5 years old}. Geneva: World
Health Organization and the United Nations Children's Fund UNICEF, (2019).
pp. 64-65.

\strong{'Percentage of children stunted, wasted, and underweight, and mean z-scores
for stunting, wasting and underweight'} \emph{in} \emph{Guide to DHS Statistics DHS-7}
Rockville, Maryland, USA: ICF (2020). pp. 431-435.
\url{https://dhsprogram.com/data/Guide-to-DHS-Statistics/Nutritional_Status.htm}

Victora CG, Schuler-Faccini L, Matijasevich A, Ribeiro E, Pessoa A,
Barros FC. \strong{Microcephaly in Brazil: how to interpret reported numbers?}
\emph{The Lancet} 2016, \emph{387(10019):621-624} \doi{10.1016/S0140-6736(16)00273-7}

Accogli A, Geraldo AF, Piccolo G, Riva A, Scala M, Balagura G, et al.
\strong{Diagnostic Approach to Macrocephaly in Children}. \emph{Frontiers in
Paediatrics} 2022, \emph{9:794069} \doi{10.3389/fped.2021.794069}
}
