% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_classify.R
\name{classify_headsize}
\alias{classify_headsize}
\title{Classify head size in \code{data.frame}-like objects with GIGS-recommended
growth standards}
\usage{
classify_headsize(
  .data,
  headcirc_cm,
  age_days,
  gest_days,
  sex,
  .new = c("hcaz", "headsize")
)
}
\arguments{
\item{.data}{A \code{data.frame}-like tabular object with one or more rows. Must
be permissible by \code{\link[checkmate:checkDataFrame]{checkmate::assert_data_frame()}}, so you can also supply
a \code{tibble}, \code{data.table}, or similar.}

\item{headcirc_cm}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of head circumference values in
cm.}

\item{age_days}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of age values in days.}

\item{gest_days}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of gestational age(s) at birth
in days. This column, in conjunction with the column referred to by
\code{age_days}, is used to select which growth standard to use for each
observation.}

\item{sex}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a case-sensitive character vector of sexes,
either \code{"M"} (male) or \code{"F"} (female). By default, gigs will warn you if
any elements of \code{sex} are not \code{"M"} or \code{"F"}, or are missing (\code{NA}). You
can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}

\item{.new}{A three-length character vector with names for the output
columns. These inputs will be repaired if necessary using
\code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}, which will print any changes to the console. If
any elements in \code{.new} are the same as elements in \code{colnames(.data)}, the
function will throw an error. Default = \code{c("waz", "wfa", "wfa_outliers")}.}
}
\value{
A tabular object of the same class that was provided as \code{.data},
with three new columns named according to \code{.new}. These columns will be
(from left to right):
\itemize{
\item \code{wlz} - Numeric vector of weight-for-length/height zscores
\item \code{wfa} - Factor of weight-for-age categories without outlier flagging
\item \code{wfa_outliers} - Factor of weight-for-age categories with outlier
flagging
}
}
\description{
Classify head size in \code{data.frame}-like objects with GIGS-recommended
growth standards
}
\details{
Cut-offs for head size categories are:

\tabular{lll}{
\strong{Category}    \tab \strong{Factor level}  \tab
\strong{Z-score bounds} \cr
Severe microcephaly       \tab \code{"microcephaly_severe"} \tab \code{hcaz} ≤ -3      \cr
Microcephaly              \tab \code{"microcephaly"}        \tab -3 < \code{hcaz} ≤ -2 \cr
Normal head circumference \tab \code{"normal_headcirc"}      \tab \code{abs(hcaz)} < 2 \cr
Macrocephaly              \tab \code{"macrocephaly"}        \tab \code{hcaz} ≥ 2       \cr
Severe macrocephaly       \tab \code{"macrocephaly_severe"} \tab \code{hcaz} ≥ 3
}
}
\examples{
data <- data.frame(
  head_cm = c(41, 40, 41, 51),
  age = c(401, 185, 101, 607),
  gestage = c(189, 252, 280, 287),
  sex = c("F", "M", "F", "M")
)

data |>
  classify_headsize(
    headcirc_cm = head_cm,
    age_days = age,
    gest_days = gestage,
    sex = sex
  )
}
\references{
Victora CG, Schuler-Faccini L, Matijasevich A, Ribeiro E, Pessoa A,
Barros FC. \strong{Microcephaly in Brazil: how to interpret reported numbers?}
\emph{The Lancet} 2016, \emph{387(10019):621-624} \doi{10.1016/S0140-6736(16)00273-7}

Accogli A, Geraldo AF, Piccolo G, Riva A, Scala M, Balagura G, et al.
\strong{Diagnostic Approach to Macrocephaly in Children}. \emph{Frontiers in
Paediatrics} 2022, \emph{9:794069} \doi{10.3389/fped.2021.794069}
}
