% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify_sfga}
\alias{classify_sfga}
\title{Classify size for gestational age using the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
weight-for-gestational age standard}
\usage{
classify_sfga(weight_kg, gest_days, sex, severe = FALSE)
}
\arguments{
\item{weight_kg}{Numeric vector of length one or more with weight
value(s) in kg.}

\item{gest_days}{Numeric vector of length one or more with gestational
age(s) at birth in days between \code{168} and \code{300}. By default, gigs will
replace out-of-bounds elements in \code{gest_days} with \code{NA} and warn you. This
behaviour can be customised using the functions in \link{gigs_options}.}

\item{sex}{Character vector of length one or more with sex(es), either \code{"M"}
(male) or \code{"F"} (female). This argument is case-sensitive. By default, gigs
will replace elements of \code{sex} which are not \code{"M"} or \code{"F"} with \code{NA}
and warn you. This behaviour can be customised using the functions in
\link{gigs_options}.}

\item{severe}{A single logical value specifying whether to categorise SGA
values are below the third centile as \code{"SGA(<3)"}. Default = \code{FALSE}.}
}
\value{
Factor the same length as the longest input vector, with
size-for-gestational age classification(s). If \code{severe = FALSE}, levels are
\code{c("SGA", "AGA",  "LGA")}. If \code{severe = TRUE}, levels are \code{c("SGA(<3)", "SGA", "AGA",  "LGA")}.
}
\description{
Size for gestational age categories are split by centile: small-for-GA
(SGA; 10\if{html}{\out{<sup>}}th\if{html}{\out{</sup>}} centile), appropriate-for-GA (AGA; 10\if{html}{\out{<sup>}}th\if{html}{\out{</sup>}}
to 90\if{html}{\out{<sup>}}th\if{html}{\out{</sup>}} centile) and large-for-GA (LGA; >90\if{html}{\out{<sup>}}th\if{html}{\out{</sup>}}
centile). This function also supports classification of severe SGA
(<3\if{html}{\out{<sup>}}rd\if{html}{\out{</sup>}} centile) using the \code{severe} parameter.
}
\note{
Input vectors are recycled by \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}, and must
adhere to the \link{vctrs} recycling rules.
}
\examples{
# Without severe flag, does not differentiate between
# p < 0.03 and p < 0.10
classify_sfga(
  weight_kg = c(2.2, 3.4, 4.2),
  gest_days = 267,
  sex = "M"
)

# With severe = TRUE, highlights p < 0.03
classify_sfga(
  weight_kg = c(2.2, 3.4, 4.2),
  gest_days = 267,
  sex = "M",
  severe = TRUE
)
}
\references{
WHO. \strong{Physical status: the use and interpretation of anthropometry. Report
of a WHO Expert Committee.} \emph{World Health Organisation Technical Report
Series 1995,} \strong{854: 1–452}

Royal College of Obstetricians and Gynaecologists. \strong{The Investigation and
Management of the Small-for-Gestational-Age Fetus: Green-top Guideline No.
31.} \emph{Technical report, Royal College of Obstetricians and Gynaecologists,
London, 2013.}
}
\seealso{
\code{\link[=ig_nbs_wfga_value2centile]{ig_nbs_wfga_value2centile()}}, which this function calls to get
centiles for each observation.
}
