% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify_svn}
\alias{classify_svn}
\title{Classify small vulnerable newborns according to the
INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} weight-for-gestational age standard}
\usage{
classify_svn(weight_kg, gest_days, sex)
}
\arguments{
\item{weight_kg}{Numeric vector of length one or more with weight
value(s) in kg.}

\item{gest_days}{Numeric vector of length one or more with gestational
age(s) at birth in days between \code{168} and \code{300}. By default, gigs will
replace out-of-bounds elements in \code{gest_days} with \code{NA} and warn you. This
behaviour can be customised using the functions in \link{gigs_options}.}

\item{sex}{Character vector of length one or more with sex(es), either \code{"M"}
(male) or \code{"F"} (female). This argument is case-sensitive. By default, gigs
will replace elements of \code{sex} which are not \code{"M"} or \code{"F"} with \code{NA}
and warn you. This behaviour can be customised using the functions in
\link{gigs_options}.}
}
\value{
Factor with small vulnerable newborn classification(s), with levels
\code{c("Preterm SGA", "Preterm AGA", "Preterm LGA", "Term SGA", "Term AGA", "Term LGA")}.
}
\description{
A small vulnerable newborn (SVN) can be preterm (born too soon),
small-for-gestational age (born too small), or within both of these
categories. This function uses weight and gestational age to categorise
newborns according to their SVN type, for use in downstream analyses.
}
\note{
Input vectors are recycled by \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}, and must
adhere to the \link{vctrs} recycling rules.
}
\examples{
classify_svn(
  weight_kg = c(1.5, 2.6, 2.6, 3.5),
  gest_days = c(235, 257, 275, 295),
  sex = c("F", "M", "F", "M")
)
}
\references{
Lawn JE, Ohuma EO, Bradley E, Idueta LS, Hazel E, Okwaraji YB et al.
\strong{Small babies, big risks: global estimates of prevalence and mortality for
vulnerable newborns to accelerate change and improve counting.} \emph{The Lancet}
2023, \emph{401(10389):1707-1719.} \doi{10.1016/S0140-6736(23)00522-6}
}
\seealso{
\code{\link[=classify_sfga]{classify_sfga()}}, which this function uses to stratify newborns
into size-for-gestational age categories.
}
