% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify_wfa}
\alias{classify_wfa}
\title{Classify weight-for-age using WHO or INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} standards}
\usage{
classify_wfa(weight_kg, age_days, gest_days, sex, outliers = FALSE)
}
\arguments{
\item{weight_kg}{Numeric vector of length one or greater with weight
value(s) in kg.}

\item{age_days}{Numeric vector of length one or greater  with age(s) in days
for each child. Should be between \code{0} to \code{1856} days.}

\item{gest_days}{Numeric vector of length one or greater with gestational
age(s) at birth in days.}

\item{sex}{Character vector of length one or more with sex(es), either \code{"M"}
(male) or \code{"F"} (female). This argument is case-sensitive. By default, gigs
will replace elements of \code{sex} which are not \code{"M"} or \code{"F"} with \code{NA}
and warn you. This behaviour can be customised using the functions in
\link{gigs_options}.}

\item{outliers}{A single \code{TRUE} or \code{FALSE} value specifying whether
implausible z-score value thresholds should be applied. Default = \code{FALSE}.}
}
\value{
Factor of weight-for-age classifications with same length as longest
input. If \code{outliers = FALSE}, levels are \code{c("underweight_severe", "underweight", "normal", "overweight")}. If \code{outliers = TRUE}, levels are
\code{c("underweight_severe", "underweight", "normal", "overweight", "outlier")}.
}
\description{
Classify weight-for-age z-scores using the WHO Child Growth Standards or
INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Postnatal Growth standards depending on the
gestational age at birth for the infant. Severely underweight is less than 3
SD below the mean, underweight is less than 2 SD below the mean, and
overweight is > 2 SDs above the mean.
}
\note{
Input vectors will be recycled by \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}.
Implausible z-score bounds are sourced from the referenced WHO report, and
classification cut-offs from the DHS manual.
}
\examples{
classify_wfa(
  weight_kg = c(7.2, 4.5, 9.1, 24),
  age_days = c(401, 185, 101, 607),
  gest_days = 7 * c(27, 36, 40, 41),
  sex = c("F", "M", "F", "M")
)

classify_wfa(
  weight_kg = c(7.2, 4.5, 9.1, 24),
  age_days = c(401, 185, 101, 607),
  gest_days = 7 * c(27, 36, 40, 41),
  sex = c("F", "M", "F", "M"),
  outliers = TRUE
)
}
\references{
\strong{'Implausible z-score values'} \emph{in} World Health Organization (ed.)
\emph{Recommendations for data collection, analysis and reporting on
anthropometric indicators in children under 5 years old}. Geneva: World
Health Organization and the United Nations Children's Fund UNICEF, (2019).
pp. 64-65.

\strong{'Percentage of children stunted, wasted, and underweight, and mean z-scores
for stunting, wasting and underweight'} \emph{in} \emph{Guide to DHS Statistics DHS-7}
Rockville, Maryland, USA: ICF (2020). pp. 431-435.
\url{https://dhsprogram.com/data/Guide-to-DHS-Statistics/Nutritional_Status.htm}
}
