% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_compute.R
\name{compute_sfga}
\alias{compute_sfga}
\title{Compute size-for-GA classifications using multiple vectors}
\usage{
compute_sfga(weight_kg, gest_days, sex, severe = FALSE)
}
\arguments{
\item{weight_kg}{Numeric vector of length one or more with weight value(s) in
kg.}

\item{gest_days}{Numeric vector of length one or more with gestational ages
in days. Classifications can only be computed when \code{gest_days} is between
\code{168} and \code{300}, as these are the limits of the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
Newborn Size standards (including the Very Preterm standards). By default,
gigs will warn you about elements of \code{gest_days} which are either outside
these bounds, are \code{NA}, or are \code{Inf}. You can customise this behaviour
using the \link[=gigs_options]{GIGS package-level options}.}

\item{sex}{Character vector of length one or more with sex(es), either \code{"M"}
(male) or \code{"F"} (female). This argument is case-sensitive. By default, gigs
will replace elements of \code{sex} which are not \code{"M"} or \code{"F"} with \code{NA}
and warn you. You can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}

\item{severe}{A single logical value specifying whether to categorise SGA
values below the third centile (\code{0.03}) as \code{"SGA(<3)"}. Default = \code{FALSE}.}
}
\value{
An object of class factor with the same length as the longest input
vector, containing size-for-GA classifications. If \code{severe = FALSE}, levels
are \code{c("SGA", "AGA", "LGA")}. If \code{severe = TRUE}, levels are \code{c("SGA(<3)", "SGA", "AGA", "LGA")}.
}
\description{
Compute size-for-GA classifications using multiple vectors
}
\note{
Input vectors are recycled by \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}, and must
adhere to the \link{vctrs} recycling rules. It is assumed that all weights
provided to this function are birth weights recorded <12 hours after an
infant's birth.
}
\examples{
# By default, does not differentiate between p < 0.03 and p < 0.10
compute_sfga(
  weight_kg = c(2.2, 3.3, 4.2),
  gest_days = 267:269,
  sex = c("M", "F", "M")
)

# With severe = TRUE, highlights p < 0.03
compute_sfga(
  weight_kg = c(2.2, 3.3, 4.2),
  gest_days = 267:269,
  sex = c("M", "F", "M"),
  severe = TRUE
)
}
\references{
WHO. \strong{Physical status: the use and interpretation of anthropometry. Report
of a WHO Expert Committee.} \emph{World Health Organisation Technical Report
Series 1995,} \strong{854: 1–452}

Royal College of Obstetricians and Gynaecologists. \strong{The Investigation and
Management of the Small-for-Gestational-Age Fetus: Green-top Guideline No.
31.} \emph{Technical report, Royal College of Obstetricians and Gynaecologists,
London, 2013.}
}
