% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_compute.R
\name{compute_stunting}
\alias{compute_stunting}
\title{Compute stunting classifications using multiple vectors}
\usage{
compute_stunting(lenht_cm, age_days, gest_days, sex, outliers = FALSE)
}
\arguments{
\item{lenht_cm}{Numeric vector of length one or more with length/height
measurement(s) in cm.}

\item{age_days}{Numeric vector of length one or more with age(s) in days
for each child. Should be between \code{0} to \code{1856} days. By default, gigs will
replace out-of-bounds elements in \code{age_days} with \code{NA} and warn you. This
behaviour can be customised using the functions in \link{gigs_options}.}

\item{gest_days}{Numeric vector with gestational age(s) at birth in days.}

\item{sex}{Character vector of length one or more with sex(es), either \code{"M"}
(male) or \code{"F"} (female). This argument is case-sensitive. By default, gigs
will replace elements of \code{sex} which are not \code{"M"} or \code{"F"} with \code{NA}
and warn you. You can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}

\item{outliers}{A single \code{TRUE} or \code{FALSE} value specifying whether
implausible z-score thresholds should be applied. Default = \code{FALSE}.}
}
\value{
An object of class factor with the same length as the longest input
vector, containing stunting classifications. Its levels are
\code{c("stunting_severe", "stunting", "not_stunting")} if \code{outliers = FALSE} (the default), else \code{c("stunting_severe", "stunting", "not_stunting", "outlier")}.

\tabular{lll}{
\strong{Category} \tab \strong{Factor level} \tab
\strong{Z-score bounds} \cr
Severe stunting \tab \code{"stunting_severe"}     \tab \code{lhaz} ≤ -3         \cr
Stunting        \tab \code{"stunting"}            \tab -3 < \code{lhaz} ≤ -2    \cr
No stunting     \tab \code{"not_stunting"}        \tab \code{lhaz} > -2         \cr
Outlier         \tab \code{"outlier"}             \tab \code{abs(lhaz)} > 6
}
}
\description{
Compute stunting classifications using multiple vectors
}
\note{
This function assumes that your measurements were taken according to
WHO guidelines, which stipulate that recumbent length should not be
measured after 730 days. Instead, standing height should be used.
Implausible z-score bounds are sourced from the referenced WHO report, and
classification cut-offs from the DHS manual.
}
\examples{
# The first observation uses the INTERGROWTH-21st Postnatal Growth standards;
# the next two use the WHO Child Growth Standards.
compute_stunting(
  lenht_cm = c(52.2, 75.4, 63.1),
  age_days = c(357, 375, 250),
  gest_days = c(196, 287, 266),
  sex = c("M", "M", "F")
)

# And with outlier flagging:
compute_stunting(
  lenht_cm = c(52.2, 75.4, 63.1),
  age_days = c(357, 375, 250),
  gest_days = c(196, 287, 266),
  sex = c("M", "M", "F"),
  outliers = TRUE
)
}
\references{
\strong{'Implausible z-score values'} \emph{in} World Health Organization (ed.)
\emph{Recommendations for data collection, analysis and reporting on
anthropometric indicators in children under 5 years old}. Geneva: World
Health Organization and the United Nations Children's Fund UNICEF, (2019).
pp. 64-65.

\strong{'Percentage of children stunted, wasted, and underweight, and mean z-scores
for stunting, wasting and underweight'} \emph{in} \emph{Guide to DHS Statistics DHS-7}
Rockville, Maryland, USA: ICF (2020). pp. 431-435.
\url{https://dhsprogram.com/data/Guide-to-DHS-Statistics/Nutritional_Status.htm}
}
