% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_compute.R
\name{compute_wasting}
\alias{compute_wasting}
\title{Compute wasting classifications using multiple vectors}
\usage{
compute_wasting(
  weight_kg,
  lenht_cm,
  age_days,
  gest_days,
  sex,
  outliers = FALSE
)
}
\arguments{
\item{weight_kg}{Numeric vector of length one or more with weight
measurement(s) in kg.}

\item{lenht_cm}{Numeric vector of length one or more with length/height
measurement(s) in cm.}

\item{age_days}{Numeric vector of length one or more with age(s) in days
for each child. Should be between \code{0} to \code{1856} days. By default, gigs will
replace out-of-bounds elements in \code{age_days} with \code{NA} and warn you. This
behaviour can be customised using the functions in \link{gigs_options}.}

\item{gest_days}{Numeric vector with gestational age(s) at birth in days.}

\item{sex}{Character vector of length one or more with sex(es), either \code{"M"}
(male) or \code{"F"} (female). This argument is case-sensitive. By default, gigs
will replace elements of \code{sex} which are not \code{"M"} or \code{"F"} with \code{NA}
and warn you. You can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}

\item{outliers}{A single \code{TRUE} or \code{FALSE} value specifying whether
implausible z-score thresholds should be applied. Default = \code{FALSE}.}
}
\value{
An object of class factor with the same length as \code{wlz}, containing
wasting classifications. Its levels are \code{c("wasting_severe", "wasting", "not_wasting", "overweight")} if \code{outliers = FALSE} (the default), else
\code{c("wasting_severe", "wasting", "not_wasting", "overweight", "outlier")}.

\tabular{lll}{
\strong{Category} \tab \strong{Factor level} \tab
\strong{Z-score bounds} \cr
Severe wasting \tab \code{"wasting_severe"}       \tab \code{wlz} ≤ -3      \cr
Wasting        \tab \code{"wasting"}              \tab -3 < \code{wlz} ≤ -2 \cr
No wasting     \tab \code{"not_wasting"}          \tab \code{abs(wlz)} < 2  \cr
Overweight     \tab \code{"overweight"}           \tab \code{wlz} ≥ 2       \cr
Outlier        \tab \code{"outlier"}              \tab \code{abs(wlz)} > 5
}
}
\description{
Compute wasting classifications using multiple vectors
}
\examples{
compute_wasting(
  weight_kg = c(5.75, 2.18, 3.00, 6.75),
  lenht_cm = c(67.7, 46.6, 50.0, 80.1),
  age_days = c(251, 197, 225, 243),
  gest_days = c(251, 197, 225, 243),
  sex =  c("F", "M", "F", "M")
)

# And with outlier flagging:
compute_wasting(
  weight_kg = c(5.75, 2.18, 3.00, 6.75),
  lenht_cm = c(67.7, 46.6, 50.0, 80.1),
  gest_days = c(251, 197, 225, 243),
  age_days = c(251, 197, 225, 243),
  sex =  c("F", "M", "F", "M"),
  outliers = TRUE
)
}
\references{
\strong{'Implausible z-score values'} \emph{in} World Health Organization (ed.)
\emph{Recommendations for data collection, analysis and reporting on
anthropometric indicators in children under 5 years old}. Geneva: World
Health Organization and the United Nations Children's Fund UNICEF, (2019).
pp. 64-65.

\strong{'Percentage of children stunted, wasted, and underweight, and mean z-scores
for stunting, wasting and underweight'} \emph{in} \emph{Guide to DHS Statistics DHS-7}
Rockville, Maryland, USA: ICF (2020). pp. 431-435.
\url{https://dhsprogram.com/data/Guide-to-DHS-Statistics/Nutritional_Status.htm}
}
