% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/gigs-package.R
\docType{package}
\name{gigs-package}
\alias{gigs-package}
\alias{gigs}
\title{gigs: Assess Growth in Infants and Newborns}
\description{
Convert between anthropometric measures and z-scores/percentiles using WHO
and INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} growth standards. This includes functions for
classification of newborn and infant growth, including size for gestational
age, stunting, wasting, and weight-for-age.

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Convert between anthropometric measures and z-scores/percentiles in multiple growth standards, and classify infant growth accordingly. With a simple interface to growth standards from the World Health Organisation and International Fetal and Newborn Growth Consortium for the 21st Century, gigs makes growth assessment easy and reproducible for clinicians, researchers and policy-makers.
}
\examples{
\dontrun{
help(package = "gigs")
browseURL("https://www.github.com/lshtm-gigs/gigs")
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lshtm-gigs/gigs/}
  \item \url{https://lshtm-gigs.github.io/gigs/}
  \item Report bugs at \url{https://github.com/lshtm-gigs/gigs/issues}
}

}
\author{
\strong{Maintainer}: Simon Parker \email{simon.parker@lshtm.ac.uk} (\href{https://orcid.org/0009-0003-8214-4496}{ORCID})

Other contributors:
\itemize{
  \item Eric Ohuma (\href{https://orcid.org/0000-0002-3116-2593}{ORCID}) [data contributor, reviewer]
  \item Bill & Melinda Gates Foundation [funder, copyright holder]
}

}
\keyword{internal}
