% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gigs_zscoring.R
\name{gigs_waz}
\alias{gigs_waz}
\alias{gigs_lhaz}
\alias{gigs_wlz}
\alias{gigs_hcaz}
\title{Calculate z-scores for anthropometric measures according to GIGS guidance}
\usage{
gigs_waz(weight_kg, age_days, gest_days, sex, id = NULL)

gigs_lhaz(lenht_cm, age_days, gest_days, sex, id = NULL)

gigs_wlz(weight_kg, lenht_cm, age_days, gest_days, sex, id = NULL)

gigs_hcaz(headcirc_cm, age_days, gest_days, sex, id = NULL)
}
\arguments{
\item{weight_kg}{A numeric vector of length one or more with weight in kg.}

\item{age_days}{A numeric vector of length one or more with ages in days.}

\item{gest_days}{A numeric vector of length one or more with gestational ages
in days.}

\item{sex}{Character vector of length one or more with sex(es), either \code{"M"}
(male) or \code{"F"} (female). This argument is case-sensitive. By default, gigs
will replace elements of \code{sex} which are not \code{"M"} or \code{"F"} with \code{NA}
and warn you. You can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}

\item{id}{A factor of length one or more with an ID for each observation,
either ordered or unordered, containing no missing (\code{NA}) values. When
supplied, \code{id} is used to ensure that only the earliest measurement for
each individual is used as a birth measure. Leave this argument as \code{NULL}
if all your data comes from the same individual. Default = \code{NULL}.}

\item{lenht_cm}{A numeric vector of length one or more with length or height
in cm. Should be recumbent length when \code{age_days < 731}, and standing
height when \code{age_days >= 731}.}

\item{headcirc_cm}{Numeric vector of length one or more with head
circumference in cm.}
}
\value{
A numeric vector of z-scores, derived using the appropriate growth
standard for each element-wise combination of \code{gest_days} and \code{age_days}.
For \code{gigs_wlz()}, all birth WLZs will be missing (\code{NA}), as there is no
INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{<sup>}} Newborn standard for weight-for-length.
}
\description{
These functions calculate z-scores for weight-for-age (WAZs),
length/height-for-age (LHAZs), weight-for-length/height (WLZs), or
head circumference-for-age (HCAZs).

The growth standard used for each observation differs based on the
gestational and post-menstrual age of the infant being analysed. The
different procedures are based on the advised use of each growth standard,
where a measurement is considered a 'birth measurement' if it is the first
measurement taken from an infant and was taken <3 days after birth:
\itemize{
\item \code{gest_days} < 24 weeks (168 days):
\itemize{
\item Birth: No standard available
\item Postnatal: IG-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Postnatal Growth standards from 27 to 64
weeks' PMA
\item Postnatal: Uncorrected WHO standard after 64 weeks' PMA
}
\item \code{gest_days} >= 24 weeks (168 days):
\itemize{
\item Birth: IG-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Newborn Size standards at birth
(including Very Preterm Newborn Size standards if born before 33
weeks).
\item Postnatal: IG-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Postnatal Growth standards from 27 to 64
weeks' PMA
\item Postnatal: Uncorrected WHO Child Growth standards after 64 weeks' PMA
}
\item \code{gest_days} ≥ 37 weeks (259 days):
\itemize{
\item Birth: IG-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Newborn Size standards
\item Postnatal: Uncorrected WHO Child Growth standards
}
\item \code{gest_days} ≥ 43 weeks (301 days):
\itemize{
\item Birth: No standard available
\item Postnatal: Uncorrected WHO Child Growth standards
}
}

For \code{gigs_wlz()}, two WHO standards are available:
\itemize{
\item The weight-for-length standard, which is applied when \code{age_days < 731}.
\item The weight-for-height standard, which is applied when \code{age_days >= 731}.
}
}
\note{
These functions expect vectors which are recyclable with
\code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}.
}
\examples{
gest_days <- c(rep(35 * 7, 3), rep(35 * 7, 3))
age_days <- c(0, 100, 500, 2, 100, 500)
sex <- rep.int(c("M", "F"), c(3, 3))
wt_kg <- c(3, 6, 9, 3, 6, 9)
len_cm <- rep.int(c(52.2, 60.4, 75), 2)
head_cm <- rep.int(c(30, 40, 49), 2)
ids <- factor(rep.int(c("A", "B"), c(3, 3)))

# Weight-for-age z-score (WAZ)
waz <- gigs_waz(weight_kg = wt_kg,
                age_days = age_days,
                gest_days = gest_days,
                sex = sex,
                id = ids)
print(waz)

# Note - if you don't specify 'id' you'll get different results!
waz_no_id <- gigs_waz(weight_kg = wt_kg,
                      age_days = age_days,
                      gest_days = gest_days,
                      sex = sex)
print(waz == waz_no_id)

# Length/height-for-age z-score (LHAZ)
lhaz <- gigs_lhaz(lenht_cm = len_cm,
                  age_days = age_days,
                  gest_days = gest_days,
                  sex = sex,
                  id = ids)
print(lhaz)

# Weight-for-length/height z-score (WLZ)
#   Note: There's no at-birth standards for weight-for-length, so elements 1
#   and 4 have `NA` values
wlz <- gigs_wlz(weight_kg = wt_kg,
                lenht_cm = len_cm,
                age_days = age_days,
                gest_days = gest_days,
                sex = sex,
                id = ids)
print(wlz)

# Head circumference-for-age z-score (HCAZ)
hcaz <- gigs_hcaz(headcirc_cm = head_cm,
                  age_days = age_days,
                  gest_days = gest_days,
                  sex = sex,
                  id = ids)
print(hcaz)
}
