% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ig_fet}
\alias{ig_fet}
\title{INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Fetal Growth Standards growth curve data}
\source{
\url{INTERGROWTH-21<sup>st</sup> Fetal Growth standards}
}
\description{
A set of nested lists containing tables with reference values at different
z-scores/centiles for valid gestational ages in days. The list is ordered
by acronym first, then by z-score/centile.
}
\examples{
names(gigs::ig_fet)
head(gigs::ig_fet$hcfga$zscores)
}
\references{
Papageorghiou AT, Ohuma EO, Altman DG, Todros T, Cheikh Ismail L, Lambert A
et al. \strong{International standards for fetal growth based on serial ultrasound
measurements: the Fetal Growth Longitudinal Study of the INTERGROWTH-21st
Project.} \emph{Lancet} 2014, \strong{384(9946):869-79.}
\doi{0.1016/S0140-6736(14)61490-2}
}
\keyword{data}
