% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_fet.R
\name{ig_fet_estimate_fetal_weight}
\alias{ig_fet_estimate_fetal_weight}
\title{Estimate fetal weight in grams using the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
predictive equation}
\usage{
ig_fet_estimate_fetal_weight(abdocirc_cm, headcirc_cm)
}
\arguments{
\item{abdocirc_cm}{Numeric vector with abdominal circumference value(s) in
cm. Should have length 1 or same length as \code{headcirc_mm}.}

\item{headcirc_cm}{Numeric vector with head circumference value(s) in cm.
Should have length 1 or same length as \code{abdocirc_mm}.}
}
\description{
Estimate fetal weight in grams using the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
predictive equation
}
\references{
Stirnemann J, Villar J, Salomon LJ, Ohuma EO, Lamber A, Victoria CG et al.
\strong{International Estimated Fetal Weight Standards of the INTERGROWTH-21st
Project.} \emph{Ultrasound Obstet Gynecol} 2016, \strong{49:478-486}
\doi{10.1002/uog.17347}
}
