% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_fet_estimation.R
\name{ig_fet_estimate_ga}
\alias{ig_fet_estimate_ga}
\title{Estimate gestational age using INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} predictive
equations}
\usage{
ig_fet_estimate_ga(crl_mm = NULL, headcirc_mm = NULL, femurlen_mm = NULL)
}
\arguments{
\item{crl_mm}{Numeric vector with crown-rump length value(s) in mm taken
from an ultrasound scan obtained within 56 to 97 days since the last
menstrual period. If not \code{NULL} (the default), should be of length 1 or the
same length as other inputs. The function will return NA for all elements
of \code{crl_mm} which are not between 15 and 95 mm and do not have
\code{headcirc_mm} or \code{abdocirc_mm} values.}

\item{headcirc_mm}{Numeric vector with head circumference value(s) in cm
taken from an ultrasound scan obtained within 98 to 181 days since the last
menstrual period. If not \code{NULL} (the default), should have length 1 or
same length as other inputs.}

\item{femurlen_mm}{Numeric vector with femur length value(s) in cm. If not
\code{NULL} (the default), should have length 1 or same length as other
arguments inputs.}
}
\value{
A numeric vector with estimated gestational ages in days. This
vector will have a length equal to the outputs from
\code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}.
}
\description{
Estimate gestational age using INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} predictive
equations
}
\note{
Vector inputs to this function are recycled using
\code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}. The function will attempt to estimate the
gestational age (GA) in three steps, from highest to lowest accuracy.
First, using crown-rump length measurements obtained in early pregnancy,
then with both head circumference and femur length measurements, and
finally with head circumference alone.
}
\examples{
# Estimate gestational age in days using crown-rump length (most accurate)
ig_fet_estimate_ga(crl_mm = 25)

# Estimate gestational age in days using head circumference and femur length
# (second-most accurate)
ig_fet_estimate_ga(headcirc_mm = 250, femurlen_mm = 55)

# Estimate gestational age in days using head circumference only (least
# accurate)
ig_fet_estimate_ga(headcirc_mm = 250)

# The function defaults to CRL if available
ig_fet_estimate_ga(crl_mm = 25, headcirc_mm = 250, femurlen_mm = 55)

# Inputs are recycled using [vctrs::vec_recycle_common]
ig_fet_estimate_ga(headcirc_mm = 249:251, femurlen_mm = 55)
}
\references{
Papageorghiou AT, Kennedy SH, Salomon LJ, Ohuma EO, Cheikh Ismail L, Barros
FC et al. \strong{International standards for early fetal size and pregnancy dating
based on ultrasound measurement of crown-rump length in the first trimester
of pregnancy.} \emph{Ultrasound Obstet Gynecol} 2014, \strong{44(6):641-48}
\doi{10.1002/uog.13448}

Papageorghiou AT, Kemp B, Stones W, Ohuma EO, Kennedy SH, Purwar M et al.
\strong{Ultrasound-based gestational age estimation in late pregnancy.}
\emph{Ultrasound Obstet Gynecol} 2016. \strong{48(6):719-26} \doi{10.1002/uog.15894}
}
