% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_fet.R
\name{ig_fet_value2zscore}
\alias{ig_fet_value2zscore}
\alias{ig_fet_hcfga_value2zscore}
\alias{ig_fet_bpdfga_value2zscore}
\alias{ig_fet_acfga_value2zscore}
\alias{ig_fet_flfga_value2zscore}
\alias{ig_fet_ofdfga_value2zscore}
\alias{ig_fet_efwfga_value2zscore}
\alias{ig_fet_value2centile}
\alias{ig_fet_hcfga_value2centile}
\alias{ig_fet_bpdfga_value2centile}
\alias{ig_fet_acfga_value2centile}
\alias{ig_fet_flfga_value2centile}
\alias{ig_fet_ofdfga_value2centile}
\alias{ig_fet_efwfga_value2centile}
\title{Convert values to z-scores/centiles in the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Fetal
Growth standards}
\usage{
ig_fet_value2zscore(y, gest_days, acronym)

ig_fet_hcfga_value2zscore(headcirc_mm, gest_days)

ig_fet_bpdfga_value2zscore(bpd_mm, gest_days)

ig_fet_acfga_value2zscore(abdocirc_mm, gest_days)

ig_fet_flfga_value2zscore(femurlen_mm, gest_days)

ig_fet_ofdfga_value2zscore(ofd_mm, gest_days)

ig_fet_efwfga_value2zscore(efw_g, gest_days)

ig_fet_value2centile(headcirc_mm, gest_days, acronym)

ig_fet_hcfga_value2centile(headcirc_mm, gest_days)

ig_fet_bpdfga_value2centile(bpd_mm, gest_days)

ig_fet_acfga_value2centile(abdocirc_mm, gest_days)

ig_fet_flfga_value2centile(femurlen_mm, gest_days)

ig_fet_ofdfga_value2centile(ofd_mm, gest_days)

ig_fet_efwfga_value2centile(efw_g, gest_days)
}
\arguments{
\item{y}{Numeric vector of measurement(s) to convert to centiles/z-scores.
Units depend on which \code{acronym}(s) are in use.}

\item{gest_days}{Numeric vector of gestational age(s) in days. Elements
should be between \code{98} to \code{280} for all standards other than estimated
fetal weight-for-GA (\code{"efwfga"}), which takes values betweem \code{154} and
\code{280}. Will return NA for each observation where GA is not within bounds.}

\item{acronym}{Acronym(s) denoting an INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Fetal
Growth standard. Each value should be one of \code{"hcfga"} (head
circumference-for-GA), \code{"bpdfga"} (biparietal diameter-for-age), \code{"acfga"}
(abdominal circumference-for-GA), \code{"flfga"} (femur length-for-GA),
\code{"ofdfga"} (occipito-frontal diameter-for-GA), or \code{"efwfga"} (estimated
fetal length-for-GA). If none of these, the function will return NA.}
}
\description{
Convert values to z-scores/centiles in the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Fetal
Growth standards
}
\examples{
# TODO: update with a proper example
print("temp")
}
\references{
Papageorghiou AT, Ohuma EO, Altman DG, Todros T, Cheikh Ismail L, Lambert A
et al. \strong{International standards for fetal growth based on serial ultrasound
measurements: the Fetal Growth Longitudinal Study of the INTERGROWTH-21st
Project.} \emph{Lancet} 2014, \strong{384(9946):869-79.}
\doi{0.1016/S0140-6736(14)61490-2}

Stirnemann J, Villar J, Salomon LJ, Ohuma EO, Lamber A, Victoria CG et al.
\strong{International Estimated Fetal Weight Standards of the INTERGROWTH-21st
Project.} \emph{Ultrasound Obstet Gynecol} 2016, \strong{49:478-486}
\doi{10.1002/uog.17347}
}
