% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ig_nbs}
\alias{ig_nbs}
\title{INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Newborn Standards (including very preterm) tables}
\source{
\href{https://intergrowth21.tghn.org/very-preterm-size-birth/#vp1}{INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Newborn Size in Very Preterm Infants}\cr
\href{https://intergrowth21.tghn.org/newborn-size-birth/#ns1}{INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Newborn Size Standards}\cr
\href{https://www.nature.com/articles/pr201752}{INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Newborn Size Standards - Body Composition}
}
\description{
A set of nested lists containing tables with reference values at different
z-scores/percentiles for valid gestational ages in days. The list is ordered
by acronym first, then by sex and finally by z-score/percentile.
}
\note{
The tables in this package are combined versions of the tables published by
Villar \emph{et al.} (2014) and Villar \emph{et al.} (2016), so they cover \code{133} to
\code{300} days' gestational age. The body composition tables (\code{ffmfga}, \code{bfpfga},
and \code{fmfga}) cover a smaller gestational age span, ranging from only \code{266} to
\code{294} days' (38 to 42 weeks') gestational age.
}
\examples{
names(gigs::ig_nbs)
head(gigs::ig_nbs$wfga$male$zscores)
}
\references{
Villar J, Giuliani F, Fenton TR, Ohuma EO, Ismail LC, Kennedy SH et al.
\strong{INTERGROWTH-21st very preterm size at birth reference charts.} \emph{Lancet}
2016, \strong{387(10021):844-45.} \doi{10.1016/S0140-6736(16)00384-6}

Villar J, Cheikh Ismail L, Victora CG, Ohuma EO, Bertino E, Altman DG, et al.
\strong{International standards for newborn weight, length, and head circumference
by gestational age and sex: the Newborn Cross-Sectional Study of the
INTERGROWTH-21st Project.} \emph{Lancet} 2014, \strong{384(9946):857-68.}
\doi{10.1016/S0140-6736(14)60932-6}

Villar J, Puglia FA, Fenton TR, Ismal LC, Staines-Urias E, Giuliani F, et al.
\strong{Body composition at birth and its relationship with neonatal anthropometric
ratios: the newborn body composition study of the INTERGROWTH-21st
project.} \emph{Pediatric Research} 2017, \strong{82:305-316.}
\doi{10.1038/pr.2017.52}
}
\keyword{data}
