% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_nbs.R
\name{ig_nbs_bodycomp}
\alias{ig_nbs_bodycomp}
\title{INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} body composition equation parameters}
\usage{
ig_nbs_bodycomp(sex, acronym)
}
\arguments{
\item{sex}{Sex(es), either \code{"M"} (male) or \code{"F"} (female).}

\item{acronym}{Acronym(s) denoting the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} NBS body
composition standard to use. Must be one of \code{"fmfga"}, \code{"bfpfga"}, or
\code{"ffmfga"}.}
}
\value{
Body composition equation parameters for each provided
age/sex/acronym combination.
}
\description{
INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} body composition equation parameters
}
\note{
These parameters are not included in the referenced publication, but
the associated supplementary materials. We used centile tables S1, S2 and S3
and linear models to derive the equations. As a result, z-scores/percentiles
derived from these parameters differ slightly from the Villar \emph{et al.}'s
published values.
}
\references{
Villar J, Puglia FA, Fenton TR, Ismal LC, Staines-Urias E, Giuliani F, et al.
\strong{Body composition at birth and its relationship with neonatal anthropometric
ratios: the newborn body composition study of the INTERGROWTH-21st
project.} \emph{Pediatric Research} 2017, \strong{82:305-316.}
\doi{10.1038/pr.2017.52}
}
\keyword{internal}
