% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_nbs.R
\name{ig_nbs_centile2value}
\alias{ig_nbs_centile2value}
\alias{ig_nbs_wfga_centile2value}
\alias{ig_nbs_lfga_centile2value}
\alias{ig_nbs_hcfga_centile2value}
\alias{ig_nbs_wlrfga_centile2value}
\alias{ig_nbs_fmfga_centile2value}
\alias{ig_nbs_bfpfga_centile2value}
\alias{ig_nbs_ffmfga_centile2value}
\alias{ig_nbs_zscore2value}
\alias{ig_nbs_wfga_zscore2value}
\alias{ig_nbs_lfga_zscore2value}
\alias{ig_nbs_hcfga_zscore2value}
\alias{ig_nbs_wlrfga_zscore2value}
\alias{ig_nbs_fmfga_zscore2value}
\alias{ig_nbs_bfpfga_zscore2value}
\alias{ig_nbs_ffmfga_zscore2value}
\title{Convert z-scores/centiles to values in the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
Newborn Size Standards}
\usage{
ig_nbs_centile2value(p, gest_days, sex, acronym)

ig_nbs_wfga_centile2value(p, gest_days, sex)

ig_nbs_lfga_centile2value(p, gest_days, sex)

ig_nbs_hcfga_centile2value(p, gest_days, sex)

ig_nbs_wlrfga_centile2value(p, gest_days, sex)

ig_nbs_fmfga_centile2value(p, gest_days, sex)

ig_nbs_bfpfga_centile2value(p, gest_days, sex)

ig_nbs_ffmfga_centile2value(p, gest_days, sex)

ig_nbs_zscore2value(z, gest_days, sex, acronym)

ig_nbs_wfga_zscore2value(z, gest_days, sex)

ig_nbs_lfga_zscore2value(z, gest_days, sex)

ig_nbs_hcfga_zscore2value(z, gest_days, sex)

ig_nbs_wlrfga_zscore2value(z, gest_days, sex)

ig_nbs_fmfga_zscore2value(z, gest_days, sex)

ig_nbs_bfpfga_zscore2value(z, gest_days, sex)

ig_nbs_ffmfga_zscore2value(z, gest_days, sex)
}
\arguments{
\item{p, z}{Numeric vector of length one or more with centiles/z-scores to
convert to values. For \code{p}, if an element of \code{p} is not between \code{0} and
\code{1}, gigs will replace it with \code{NA} and warn you informatively. This
behaviour can be customised using the functions in \link{gigs_options}.}

\item{gest_days}{Numeric vector of length one or more with gestational ages
in days. Elements should be between certain values depending on the Newborn
Size standard in use (defined by \code{acronym}). These are:
\itemize{
\item Between 168 and 300 days for \code{"wfga"}, \code{"lfga"}, \code{"hcfga"}, and
\code{"wlrfga"}.
\item Between 154 and 280 days for \code{"fmfga"}, \code{"bfpfga"}, and \code{"ffmfga"}.
}

By default, gigs will replace elements in \code{gest_days} that are out of
bounds for the growth standard in use with \code{NA} and warn you. This
behaviour can be customised using the functions in \link{gigs_options}.}

\item{sex}{Character vector of length one or more with sex(es), either \code{"M"}
(male) or \code{"F"} (female). This argument is case-sensitive. By default, gigs
will replace elements of \code{sex} which are not \code{"M"} or \code{"F"} with \code{NA}
and warn you. This behaviour can be customised using the functions in
\link{gigs_options}.}

\item{acronym}{Character vector of length one or more denoting the
INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Newborn Size standard to use for each
observation. Each element should be one of:
\itemize{
\item \code{"wfga"} (weight-for-GA)
\item \code{"lfga"} (length-for-GA)
\item \code{"hcfga"} (head circumference-for-GA)
\item \code{"wlrfga"} (weight/length ratio-for-GA)
\item \code{"fmfga"} (fat mass-for-GA)
\item \code{"bfpfga"} (body fat \%-for-GA)
\item \code{"ffmfga"} (fat-free mass-for-GA)
}

This argument is case-sensitive. By default, gigs will replace elements in
\code{acronym} which are not one of the above values with \code{NA} and warn you.
This behaviour can be customised using the functions in \link{gigs_options}. If
all elements in \code{acronym} are not one of the above values, gigs will throw
an error.}
}
\value{
Numeric vector of expected measurements with length equal to the
longest input vector.
}
\description{
Convert z-scores/centiles to values in the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
Newborn Size Standards
}
\note{
Input vectors are recycled by \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}, and must
adhere to the \link{vctrs} recycling rules.
}
\examples{
# Convert centiles to values
ig_nbs_centile2value(p = c(0.25, 0.5, 0.75),
                     gest_days = 274:276,
                     sex = c("F", "F", "F"),
                     acronym = "wfga") |>
  round(digits = 2)

# Or z-scores to values
ig_nbs_zscore2value(z = qnorm(c(0.25, 0.5, 0.75)),
                    gest_days = 280,
                    sex = "M",
                    acronym = "wfga") |>
  round(digits = 2)

# Specify which standard to use with the acronym parameter...
ig_nbs_zscore2value(z = qnorm(c(0.25, 0.5, 0.75)),
                    gest_days = 280,
                    sex = "M",
                    acronym = "lfga") |>
  round(digits = 2)

# ... or by using a standard-specific function
ig_nbs_lfga_zscore2value(z = -2:2,
                         gest_days = 278:282,
                         sex = "M") |>
  round(digits = 2)

# Inputs are recycled to the input of the longest length
ig_nbs_lfga_zscore2value(z = seq(0.1, 0.9, by = 0.2),
                         gest_days = 280,
                         sex = "M") |>
  round(digits = 2)

# Bad inputs will not stop the function but will instead produce `NA`s in the
# output - by default gigs will issue useful warnings
ig_nbs_hcfga_zscore2value(z = -0.5,
                          gest_days = c(140, 182, 224, 266),
                          sex = c("M", NA, "M", "M")) |>
  round(digits = 2)
}
\references{
Villar J, Cheikh Ismail L, Victora CG, Ohuma EO, Bertino E, Altman DG, et al.
\strong{International standards for newborn weight, length, and head circumference
by gestational age and sex: the Newborn Cross-Sectional Study of the
INTERGROWTH-21st Project.} \emph{Lancet} 2014, \strong{384(9946):857-68.}
\doi{10.1016/S0140-6736(14)60932-6}

Villar J, Giuliani F, Fenton TR, Ohuma EO, Ismail LC, Kennedy SH et al.
\strong{INTERGROWTH-21st very preterm size at birth reference charts.} \emph{Lancet}
2016, \strong{387(10021):844-45.} \doi{10.1016/S0140-6736(16)00384-6}

Villar J, Puglia FA, Fenton TR, Ismal LC, Staines-Urias E, Giuliani F, et al.
\strong{Body composition at birth and its relationship with neonatal anthropometric
ratios: the newborn body composition study of the INTERGROWTH-21st project.}
\emph{Pediatric Research} 2017, \strong{82:305-316.} \doi{10.1038/pr.2017.52}
}
