% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ig_nbs_coeffs}
\alias{ig_nbs_coeffs}
\title{INTERGROWTH-21st Newborn Size Standards GAMLSS coefficients}
\source{
Provided by Dr Eric Ohuma, produced for the provided reference.
}
\description{
A set of nested lists containing mu, sigma, nu and tau values across gestational ages for either sex. The list is
ordered by acronym then sex.
}
\examples{
names(gigs::ig_nbs_coeffs)
head(gigs::ig_nbs_coeffs$wfga$male)
}
\references{
Villar J, Cheikh Ismail L, Victora CG, Ohuma EO, Bertino E, Altman DG, et al. \strong{International
standards for newborn weight, length, and head circumference by gestational age and sex: the Newborn Cross-Sectional
Study of the INTERGROWTH-21st Project.} \emph{Lancet} 2014, \strong{384(9946):857-68.}
\doi{10.1016/S0140-6736(14)60932-6}
}
\keyword{data}
