% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_nbs.R
\name{ig_nbs_percentile2value}
\alias{ig_nbs_percentile2value}
\alias{ig_nbs_wfga_percentile2value}
\alias{ig_nbs_lfga_percentile2value}
\alias{ig_nbs_hcfga_percentile2value}
\alias{ig_nbs_wlrfga_percentile2value}
\alias{ig_nbs_fmfga_percentile2value}
\alias{ig_nbs_bfpfga_percentile2value}
\alias{ig_nbs_ffmfga_percentile2value}
\alias{ig_nbs_zscore2value}
\alias{ig_nbs_wfga_zscore2value}
\alias{ig_nbs_lfga_zscore2value}
\alias{ig_nbs_hcfga_zscore2value}
\alias{ig_nbs_wlrfga_zscore2value}
\alias{ig_nbs_fmfga_zscore2value}
\alias{ig_nbs_bfpfga_zscore2value}
\alias{ig_nbs_ffmfga_zscore2value}
\title{Convert z-scores/percentiles to INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} newborn size
values}
\usage{
ig_nbs_percentile2value(p, gest_age, sex, acronym)

ig_nbs_wfga_percentile2value(p, gest_age, sex)

ig_nbs_lfga_percentile2value(p, gest_age, sex)

ig_nbs_hcfga_percentile2value(p, gest_age, sex)

ig_nbs_wlrfga_percentile2value(p, gest_age, sex)

ig_nbs_fmfga_percentile2value(p, gest_age, sex)

ig_nbs_bfpfga_percentile2value(p, gest_age, sex)

ig_nbs_ffmfga_percentile2value(p, gest_age, sex)

ig_nbs_zscore2value(z, gest_age, sex, acronym)

ig_nbs_wfga_zscore2value(z, gest_age, sex)

ig_nbs_lfga_zscore2value(z, gest_age, sex)

ig_nbs_hcfga_zscore2value(z, gest_age, sex)

ig_nbs_wlrfga_zscore2value(z, gest_age, sex)

ig_nbs_fmfga_zscore2value(z, gest_age, sex)

ig_nbs_bfpfga_zscore2value(z, gest_age, sex)

ig_nbs_ffmfga_zscore2value(z, gest_age, sex)
}
\arguments{
\item{p, z}{Percentile(s)/z-score(s) to convert to a value/values.}

\item{gest_age}{Gestational age(s) in days. Must be between \code{168} and \code{300}
for weight/length/head circumference-for-gestational age, or from \code{266} to
\code{294} for body composition equations.}

\item{sex}{Sex(es), either \code{"M"} (male) or \code{"F"} (female).}

\item{acronym}{Acronym(s) denoting the INTERGROWTH NBS standard to use. Must
be one of \code{"wfga"}, \code{"lfga"}, \code{"hcfga"}, \code{"wlrfga"}, \code{"fmfga"}, \code{"bfpfga"},
or \code{"ffmfga"}.}
}
\value{
Expected measurements for each combination of z-score/percentile,
gestational age, sex, and acronym provided to the function.
}
\description{
Convert z-scores/percentiles to INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} newborn size
values
}
\note{
Input vectors will be recycled to the length of the longest vector.
}
\examples{
# Convert percentiles to values
p <- 0.25 # 25th percentile
ig_nbs_percentile2value(p = p, gest_age = 280, sex = "M", acronym = "wfga") |>
  round(digits = 2)

# Or z-scores to values
z <- qnorm(p)
ig_nbs_zscore2value(z = z, gest_age = 280, sex = "M", acronym = "wfga") |>
  round(digits = 2)

# Specify which standard to use with the acronym parameter...
ig_nbs_zscore2value(z = z, gest_age = 280, sex = "M", acronym = "lfga") |>
  round(digits = 2)

# ... or by using a standard-specific function
ig_nbs_lfga_zscore2value(z = z, gest_age = 280, sex = "M") |>
  round(digits = 2)

# Inputs are recycled to the input of the longest length
ig_nbs_lfga_zscore2value(z = seq(0.1, 0.9, by = 0.2),
                         gest_age = 280,
                         sex = "M") |>
  round(digits = 2)

# Bad inputs will not stop the function but will instead return NA - here 140
# days for gest_age is outside the bounds of the INTERGROWTH-21st newborn
# size standards
ig_nbs_hcfga_zscore2value(z = z,
                          gest_age = c(140, 182, 224, 266),
                          sex = "F") |>
  round(digits = 2)
}
\references{
Villar J, Cheikh Ismail L, Victora CG, Ohuma EO, Bertino E, Altman DG, et al.
\strong{International standards for newborn weight, length, and head circumference
by gestational age and sex: the Newborn Cross-Sectional Study of the
INTERGROWTH-21st Project.} \emph{Lancet} 2014, \strong{384(9946):857-68.}
\doi{10.1016/S0140-6736(14)60932-6}

Villar J, Giuliani F, Fenton TR, Ohuma EO, Ismail LC, Kennedy SH et al.
\strong{INTERGROWTH-21st very preterm size at birth reference charts.} \emph{Lancet}
2016, \strong{387(10021):844-45.} \doi{10.1016/S0140-6736(16)00384-6}

Villar J, Puglia FA, Fenton TR, Ismal LC, Staines-Urias E, Giuliani F, et al.
\strong{Body composition at birth and its relationship with neonatal anthropometric
ratios: the newborn body composition study of the INTERGROWTH-21st
project.} \emph{Pediatric Research} 2017, \strong{82:305-316.}
\doi{10.1038/pr.2017.52}
}
