% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_nbs.R
\name{ig_nbs_value2centile}
\alias{ig_nbs_value2centile}
\alias{ig_nbs_wfga_value2centile}
\alias{ig_nbs_lfga_value2centile}
\alias{ig_nbs_hcfga_value2centile}
\alias{ig_nbs_wlrfga_value2centile}
\alias{ig_nbs_fmfga_value2centile}
\alias{ig_nbs_bfpfga_value2centile}
\alias{ig_nbs_ffmfga_value2centile}
\alias{ig_nbs_value2zscore}
\alias{ig_nbs_wfga_value2zscore}
\alias{ig_nbs_lfga_value2zscore}
\alias{ig_nbs_hcfga_value2zscore}
\alias{ig_nbs_wlrfga_value2zscore}
\alias{ig_nbs_fmfga_value2zscore}
\alias{ig_nbs_bfpfga_value2zscore}
\alias{ig_nbs_ffmfga_value2zscore}
\title{Convert values to z-scores/centiles in the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
Newborn Size Standards}
\usage{
ig_nbs_value2centile(y, gest_days, sex, acronym)

ig_nbs_wfga_value2centile(weight_kg, gest_days, sex)

ig_nbs_lfga_value2centile(length_cm, gest_days, sex)

ig_nbs_hcfga_value2centile(headcirc_cm, gest_days, sex)

ig_nbs_wlrfga_value2centile(wei_len_ratio, gest_days, sex)

ig_nbs_fmfga_value2centile(fat_mass_g, gest_days, sex)

ig_nbs_bfpfga_value2centile(body_fat_perc, gest_days, sex)

ig_nbs_ffmfga_value2centile(fatfree_mass_g, gest_days, sex)

ig_nbs_value2zscore(y, gest_days, sex, acronym)

ig_nbs_wfga_value2zscore(weight_kg, gest_days, sex)

ig_nbs_lfga_value2zscore(length_cm, gest_days, sex)

ig_nbs_hcfga_value2zscore(headcirc_cm, gest_days, sex)

ig_nbs_wlrfga_value2zscore(wei_len_ratio, gest_days, sex)

ig_nbs_fmfga_value2zscore(fat_mass_g, gest_days, sex)

ig_nbs_bfpfga_value2zscore(body_fat_perc, gest_days, sex)

ig_nbs_ffmfga_value2zscore(fatfree_mass_g, gest_days, sex)
}
\arguments{
\item{y}{Numeric vector of length one or more with anthropometric
measurement(s) to convert to centiles/z-scores. Units depend on which
\code{acronym}(s) are in use.}

\item{gest_days}{Numeric vector of length one or more with gestational ages
in days. Elements should be between certain values depending on the Newborn
Size standard in use (defined by \code{acronym}). These are:
\itemize{
\item Between 168 and 300 days for \code{"wfga"}, \code{"lfga"}, \code{"hcfga"}, and
\code{"wlrfga"}.
\item Between 266 and 280 days for \code{"fmfga"}, \code{"bfpfga"}, and \code{"ffmfga"}.
}

By default, gigs will replace elements in \code{gest_days} that are out of
bounds for the growth standard in use with \code{NA} and warn you. You can
customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}

\item{sex}{Character vector of length one or more with sex(es), either \code{"M"}
(male) or \code{"F"} (female). This argument is case-sensitive. By default, gigs
will replace elements of \code{sex} which are not \code{"M"} or \code{"F"} with \code{NA}
and warn you. You can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}

\item{acronym}{A single string denoting the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
Newborn Size standard to use. Should be one of:
\itemize{
\item \code{"wfga"} (weight-for-GA)
\item \code{"lfga"} (length-for-GA)
\item \code{"hcfga"} (head circumference-for-GA)
\item \code{"wlrfga"} (weight/length ratio-for-GA)
\item \code{"fmfga"} (fat mass-for-GA)
\item \code{"bfpfga"} (body fat \%-for-GA)
\item \code{"ffmfga"} (fat-free mass-for-GA)
}

This argument is case-sensitive. If \code{acronym} is not one of the values
listed above, the function will throw an error.}

\item{weight_kg}{Numeric vector of length one or more with birth weight(s) in
kg.}

\item{length_cm}{Numeric vector of length one or more with birth length(s) in
cm.}

\item{headcirc_cm}{Numeric vector of length one or more with birth head
circumference(s) in cm.}

\item{wei_len_ratio}{Numeric vector of length one or more with weight-length
ratio(s) in kg per m.}

\item{fat_mass_g}{Numeric vector of length one or more with fat mass(es) in
g.}

\item{body_fat_perc}{Numeric vector of length one or more with body fat
percentage(s).}

\item{fatfree_mass_g}{Numeric vector of length one or more with fat-free
mass(es) in g.}
}
\value{
Numeric vector of z-scores/centiles with length equal to the longest
input vector.
}
\description{
Convert values to z-scores/centiles in the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
Newborn Size Standards
}
\note{
Input vectors other than \code{acronym} are recycled by
\code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}, and must adhere to the \link{vctrs} recycling
rules.
}
\examples{
# Convert values to centiles
ig_nbs_value2centile(y = c(2.90, 3.17, 3.46),
                     gest_days = 274:276,
                     sex = c("F", "F", "F"),
                     acronym = "wfga") |>
  round(digits = 2)

# Or values to z-scores
ig_nbs_value2zscore(y = c(2.90, 3.17, 3.46),
                    gest_days = 274:276,
                    sex = c("F", "F", "F"),
                    acronym = "wfga") |>
  round(digits = 2)

# Specify which standard to use with the acronym parameter...
ig_nbs_value2zscore(y = 48.84,
                    gest_days = 280,
                    sex = "M",
                    acronym = "lfga") |>
  round(digits = 2)

# ... or by using a standard-specific function
ig_nbs_lfga_value2zscore(length_cm = 48.84,
                         gest_days = 280,
                         sex = "M") |>
  round(digits = 2)

# Inputs are recycled to the input of the longest length
ig_nbs_wlrfga_value2zscore(wei_len_ratio = c(7.37, 6.47, 6.12, 6.86),
                           gest_days = 280,
                           sex = "M") |>
  round(digits = 2)

# Bad inputs will not stop the function but will instead produce `NA`s in the
# output - by default gigs will issue useful warnings
ig_nbs_hcfga_value2centile(headcirc_cm = 34.0,
                           gest_days = c(262, 264, 266, 301),
                           sex = c(NA, "M", "M", "M")) |>
  round(digits = 2)
}
\references{
Villar J, Cheikh Ismail L, Victora CG, Ohuma EO, Bertino E, Altman DG, et al.
\strong{International standards for newborn weight, length, and head circumference
by gestational age and sex: the Newborn Cross-Sectional Study of the
INTERGROWTH-21st Project.} \emph{Lancet} 2014, \strong{384(9946):857-68.}
\doi{10.1016/S0140-6736(14)60932-6}

Villar J, Giuliani F, Fenton TR, Ohuma EO, Ismail LC, Kennedy SH et al.
\strong{INTERGROWTH-21st very preterm size at birth reference charts.} \emph{Lancet}
2016, \strong{387(10021):844-45.} \doi{10.1016/S0140-6736(16)00384-6}

Villar J, Puglia FA, Fenton TR, Ismal LC, Staines-Urias E, Giuliani F, et al.
\strong{Body composition at birth and its relationship with neonatal anthropometric
ratios: the newborn body composition study of the INTERGROWTH-21st project.}
\emph{Pediatric Research} 2017, \strong{82:305-316.} \doi{10.1038/pr.2017.52}
}
