% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ig_png}
\alias{ig_png}
\title{INTERGROWTH-21st Post-natal Growth Standards tables}
\source{
\href{https://intergrowth21.tghn.org/postnatal-growth-preterm-infants/#pg1}{INTERGROWTH-21st Post-natal Growth of Preterm Infants}
}
\description{
A set of nested lists containing tables with reference values at different z-scores/percentiles for valid
postmenstrual ages. The list is ordered by acronym first, then by sex and finally z-score/percentile.
}
\examples{
names(gigs::ig_png)
head(gigs::ig_png$wfa$male$zscores)
}
\references{
Villar J, Cheikh Ismail L, Victora CG, Ohuma EO, Bertino E, Altman DG, et al. \strong{International
standards for newborn weight, length, and head circumference by gestational age and sex: the Newborn Cross-Sectional
Study of the INTERGROWTH-21st Project.} \emph{Lancet} 2014, \strong{384(9946):857-68.}
\doi{10.1016/S0140-6736(14)60932-6}
}
\keyword{data}
