% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_png.R
\name{ig_png_equations}
\alias{ig_png_equations}
\title{INTERGROWTH-21st equations for postnatal size for age in preterm infants}
\usage{
ig_png_equations(pma_weeks, sex, acronym)
}
\arguments{
\item{pma_weeks}{Post-menstrual age(s) in exact weeks. Must be between \code{27}
and \code{64}.}

\item{sex}{Sex(es), either \code{"M"} (male) or \code{"F"} (female).}

\item{acronym}{Acronym(s) denoting an INTERGROWTH-21st standard for
postnatal growth in preterm infants. Should be one of \code{"wfa"}, \code{"lhfa"},
\code{"hcfa"}.}
}
\value{
A table with median(s) and standard deviation(s) for each
\code{age}/\code{sex}/\code{acronym} combination provided to the function.
}
\description{
Estimates median and standard deviation for different measures of postnatal
growth in preterm infants.
}
\note{
The weight-for-age and length-for-age standards are logarithmic, so
require slightly different treatment to use in z-score conversions. In
contrast, head circumference for gestational age returns the median and
standard deviation with no logarithm applied.
}
\references{
Villar J, Giuliani F, Bhutta ZA, Bertino E, Ohuma EO, Ismail LC et al.
\strong{Postnatal growth standards for preterm infants: the Preterm Postnatal
Follow-up Study of the INTERGROWTH-21st Project.} \emph{Lancet Glob Health} 2015,
\emph{3(11):e681-e691.} \doi{10.1016/S2214-109X(15)00163-1}
}
\keyword{internal}
