% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_png.R
\name{ig_png_zscore2value}
\alias{ig_png_zscore2value}
\alias{ig_png_wfa_zscore2value}
\alias{ig_png_lfa_zscore2value}
\alias{ig_png_hcfa_zscore2value}
\alias{ig_png_wfl_zscore2value}
\alias{ig_png_centile2value}
\alias{ig_png_wfa_centile2value}
\alias{ig_png_lfa_centile2value}
\alias{ig_png_hcfa_centile2value}
\alias{ig_png_wfl_centile2value}
\title{Convert z-scores/centiles to values in the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
Postnatal Growth Standards for preterm infants}
\usage{
ig_png_zscore2value(z, x, sex, acronym)

ig_png_wfa_zscore2value(z, pma_weeks, sex)

ig_png_lfa_zscore2value(z, pma_weeks, sex)

ig_png_hcfa_zscore2value(z, pma_weeks, sex)

ig_png_wfl_zscore2value(z, length_cm, sex)

ig_png_centile2value(p, x, sex, acronym)

ig_png_wfa_centile2value(p, pma_weeks, sex)

ig_png_lfa_centile2value(p, pma_weeks, sex)

ig_png_hcfa_centile2value(p, pma_weeks, sex)

ig_png_wfl_centile2value(p, length_cm, sex)
}
\arguments{
\item{z, p}{Numeric vector of length one or more with centiles/z-scores to
convert to values. For \code{p}, gigs will warn you if elements of \code{p} are not
between \code{0} and \code{1}. You can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}

\item{x, pma_weeks, length_cm}{Numeric vector of length one or more with x
values. Elements of \code{x} or its standard-specific equivalents (\code{pma_weeks},
\code{length_cm}) should have specific units and be between certain values
depending on the standard in use (defined by \code{acronym}). These are:
\itemize{
\item Between 27 and 64 weeks for \code{"wfa"}, \code{"lfa"}, and \code{"hcfa"}.
\item Between 35 and 65 cm for \code{"wfl"}.
}

By default, gigs will replace out-of-bounds elements in \code{x} with \code{NA} and
warn you. You can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}

\item{sex}{Character vector of length one or more with sex(es), either \code{"M"}
(male) or \code{"F"} (female). This argument is case-sensitive. By default, gigs
will replace elements of \code{sex} which are not \code{"M"} or \code{"F"} with \code{NA}
and warn you. You can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}

\item{acronym}{Single-length character variable denoting the
INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Postnatal Growth standard(s) in use. Should be
one of:
\itemize{
\item \code{"wfa"} (weight-for-age)
\item \code{"lfa"} (length-for-age)
\item \code{"hcfa"} (head circumference-for-age)
\item \code{"wfl"} (weight-for-length)
}

This argument is case-sensitive. If \code{acronym} is not one of the values
listed above the function will throw an error.}
}
\value{
Numeric vector of expected measurements with length equal to the
longest input vector.
}
\description{
Convert z-scores/centiles to values in the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
Postnatal Growth Standards for preterm infants
}
\note{
Input vectors other than \code{acronym} are recycled by
\code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}, and must adhere to the \link{vctrs} recycling
rules.
}
\examples{
# Convert centiles to values
ig_png_centile2value(p = c(0.25, 0.5, 0.75),
                     x = 54:56,
                     sex = c("M", "F", "M"),
                     acronym = "wfa") |>
  round(digits = 2)

# Or z-scores to values
ig_png_zscore2value(z = qnorm(c(0.25, 0.5, 0.75)),
                    x = 54:56,
                    sex = c("M", "F", "M"),
                    acronym = "wfa") |>
  round(digits = 2)

# Specify which standard to use with the acronym parameter...
ig_png_zscore2value(z = qnorm(c(0.25, 0.5, 0.75)),
                    x = 54:56,
                    sex = c("M", "F", "M"),
                    acronym = "lfa") |>
  round(digits = 2)

# ... or by using a standard-specific function
ig_png_lfa_zscore2value(z = qnorm(c(0.25, 0.5, 0.75)),
                        pma_weeks = 54:56,
                        sex = c("M", "F", "M")) |>
  round(digits = 2)

# Inputs are recycled to the input of the longest length
ig_png_lfa_zscore2value(z = seq(-0.5, 0.5, by = 0.2),
                        pma_weeks = 40,
                        sex = "M") |>
  round(digits = 2)

# Bad inputs will not stop the function but will instead produce `NA`s in the
# output - by default gigs will issue useful warnings
ig_png_hcfa_zscore2value(z = 0,
                         pma_weeks = c(25, 27, 46, 64),
                         sex = c("M", "M", NA, "M")) |>
  round(digits = 2)
}
\references{
Villar J, Giuliani F, Bhutta ZA, Bertino E, Ohuma EO, Ismail LC et al.
\strong{Postnatal growth standards for preterm infants: the Preterm Postnatal
Follow-up Study of the INTERGROWTH-21st Project.} \emph{Lancet Glob Health} 2015,
\emph{3(11):e681-e691.} \doi{10.1016/S2214-109X(15)00163-1}
}
