% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_vpns.R
\name{ig_vpns_zscore2value}
\alias{ig_vpns_zscore2value}
\title{Convert z-scores to values in the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Newborn Size
Standards for Very Preterm Infants}
\usage{
ig_vpns_zscore2value(z, gest_age, sex, acronym)
}
\arguments{
\item{z}{Z-score(s) to convert to a value/values.}

\item{gest_age}{Gestational age in days. Must be between \code{168} and \code{230}.
weeks.}

\item{sex}{Sex(es), either \code{"M"} (male) or \code{"F"} (female).}

\item{acronym}{Acronym(s) denoting the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} VPNS
standard to use. Must be one of \code{"wfga"}, "lfga"\verb{, or }"hcfga"`.}
}
\value{
Expected measurements for each combination of z-score, gestational
age, sex, and acronym provided to the function.
}
\description{
Convert z-scores to values in the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Newborn Size
Standards for Very Preterm Infants
}
\references{
Villar J, Giuliani F, Fenton TR, Ohuma EO, Ismail LC, Kennedy SH et al.
\strong{INTERGROWTH-21st very preterm size at birth reference charts.} \emph{Lancet}
2016, \strong{387(10021):844-45.} \doi{10.1016/S0140-6736(16)00384-6}
}
\keyword{internal}
