% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_params.R
\name{inrange}
\alias{inrange}
\title{Check if x is within upper and lower bounds of a vector}
\usage{
inrange(x, vec)
}
\description{
Check if x is within upper and lower bounds of a vector
}
\keyword{internal}
