% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{interpolate_coeffs}
\alias{interpolate_coeffs}
\title{Linearly interpolate between LMS or MSNT coefficients}
\usage{
interpolate_coeffs(coeff_tbl_long, xvar, sex, acronym)
}
\arguments{
\item{coeff_tbl_long}{A table of reference LMS/MSNT coefficients, from within
\code{gigs::who_gs_coeffs} or \code{gigs:::ig_nbs_coeffs}}

\item{xvar}{A value of x which is not found in a coefficient table but is
between two values in that coefficient table}

\item{sex}{A character denoting male (\code{"M"}) or female (\code{"F"})}
}
\value{
Data frame containing LMS/MSNT values which have been sourced from
interpolated between existing LMS/MSNT values.
}
\description{
Linearly interpolate between LMS or MSNT coefficients
}
\note{
All inputs should be length one. The function will also fail if
\code{coeff_tbl_long} does not contain named LMS/MSNT values.
}
\keyword{internal}
