% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mean_if_sex_undefined}
\alias{mean_if_sex_undefined}
\title{Return male/female mean if sex is undefined}
\usage{
mean_if_sex_undefined(fn, arg1, x_arg, acronym)
}
\arguments{
\item{fn}{Conversion function to call}

\item{arg1}{z, p or y values to pass to \code{fn}}

\item{x_arg}{X values (usually age values) to pass to \code{fn}}

\item{acronym}{Acronym values to pass to \code{fn}}
}
\description{
Return male/female mean if sex is undefined
}
\keyword{internal}
