% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rep_to_longest}
\alias{rep_to_longest}
\title{Extend vectors in list to the length of the longest vector}
\usage{
rep_to_longest(vec_list)
}
\arguments{
\item{vec_list}{List of vectors to extend out to the length of the longest
vector in that list.}
}
\value{
List of vectors recycled to length of longest vector in input list.
}
\description{
Extend vectors in list to the length of the longest vector
}
\keyword{internal}
