% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{report_units}
\alias{report_units}
\title{Check required units for a conversion using gigs}
\usage{
report_units(family, acronym)
}
\arguments{
\item{family}{A single string denoting the family of growth standards in use.
Must be one of \code{"ig_fet"}, \code{"ig_nbs"}, \code{"ig_png"}, or \code{"who_gs"}, or the
function will throw an error. This argument is case-sensitive.}

\item{acronym}{A single string denoting the specific growth standard in use.
The allowed values of \code{acronym} depend on the value of \code{family}, and are
listed in the documentation for the \code{x} argument. Incompatible \code{family} and
\code{acronym} values will cause errors to be thrown. This argument is
case-sensitive.}
}
\value{
Returns a message describing which units are required for conversion
in the desired growth standard.
}
\description{
Check required units for a conversion using gigs
}
\examples{
# Get units for working with the IG-21st Newborn Size Standard for
# weight-for-GA
report_units("ig_nbs", "wfga")

# Get units for working with the IG-21st Fetal Standard for
# gestational weight gain
report_units("ig_fet", "gwgfga")
}
