% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{round2}
\alias{round2}
\title{Round, but round 0.5 up in all cases}
\usage{
round2(x, digits)
}
\arguments{
\item{x}{Value(s) to round}

\item{digits}{Number of digits to round to}
}
\value{
Values of \code{x} rounded to \code{digits} number of digits.
}
\description{
Round, but round 0.5 up in all cases
}
\note{
Taken from https://stackoverflow.com/questions/12688717/round-up-from-5
}
\keyword{internal}
