% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{shared_value2zscore_returns}
\alias{shared_value2zscore_returns}
\title{Shared value to z-score/centile roxygen parameters}
\value{
Numeric vector of z-scores/centiles with length equal to the longest
input vector.
}
\description{
Much of the roxygen content in \verb{[gigs]} is shared between different
functions. This roxygen block is designed to make inheriting these parameters
more simple.
}
\keyword{internal}
