% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{shared_zscore2value_returns}
\alias{shared_zscore2value_returns}
\title{Shared z-score/centile to value roxygen parameters}
\value{
Numeric vector of expected measurements with length equal to the
longest input vector.
}
\description{
Much of the roxygen content in \verb{[gigs]} is shared between different
functions. This roxygen block is designed to make inheriting these parameters
more simple.
}
\keyword{internal}
