% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{value2zscore}
\alias{value2zscore}
\alias{value2centile}
\title{Convert anthropometric measures to z-scores/centiles using international
growth standards}
\usage{
value2zscore(y, x, sex = NULL, family, acronym)

value2centile(y, x, sex = NULL, family, acronym)
}
\arguments{
\item{y}{Numeric vector of length one or more with anthropometric
measurement(s) to convert to centiles/z-scores. Units depend on which
combination of \code{family} and \code{acronym} are in use, which you can check with
\code{\link[=report_units]{report_units()}}. By default, GIGS will warn you if elements in \code{y} are
missing (\code{NA}) or undefined (\code{NaN}/\code{Inf}), and replace the latter with
\code{NA}. You can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}

\item{x}{Numeric vector of x values, which take on different units depending
on the growth standard in use. To know which units to use, run
\code{report_units(family, acronym)}. Limits for each combination are as
follows:
\itemize{
\item WHO Growth Standards (family = \code{"who_gs"})
\itemize{
\item Between 0 and 1856 days for \code{"wfa"}, \code{"bfa"}, \code{"lhfa"}, and \code{"hcfa"}.
\item Between 45 and 110 cm for \code{"wfl"}.
\item Between 65 and 120 cm for \code{"wfh"}.
\item Between 91 and 1856 days for \code{"acfa"}, \code{"ssfa"}, \code{"tsfa"}.
}
\item INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Fetal Standards (family = \code{"ig_fet"})
\itemize{
\item Between 98 and 280 days for \code{"hcfga"}, \code{"bpdfga"}, \code{"acfga"},
\code{"flfga"}, \code{"ofdfga"}, and\if{html}{\out{<br>}} \code{"tcdfga"}.
\item Between 154 and 280 days for \code{"efwfga"}.
\item Between 112 and 294 days for \code{"sfhfga"}.
\item Between 58 and 105 days for \code{"crlfga"}.
\item Between 19 and 95 mm for \code{"gafcrl"}.
\item Between 105 and 280 days for \code{"gwgfga"}.
\item Between 168 and 280 days for \code{"pifga"}, \code{"rifga"}, and \code{"sdrfga"} (the
INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Fetal Doppler standards).
\item Between 12 and 55 mm for\code{"tcdfga"}.
\item Between 105 and 252 days for \code{"poffga"}, \code{"sffga"}, \code{"avfga"}, \code{"pvfga"},
and \code{"cmfga"} (the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Fetal Brain Development
standards).
\item Between 126 and 287 days for \code{"hefwfga"} (the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
standard for Hadlock-based estimated fetal weight).
}
\item INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Newborn Size Standards (family = \code{"ig_nbs"})
\itemize{
\item Between 168 and 300 days for \code{"wfga"}, \code{"lfga"}, \code{"hcfga"}, and
\code{"wlrfga"}.
\item Between 266 and 280 days for \code{"fmfga"}, \code{"bfpfga"}, and \code{"ffmfga"}.
}
\item INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Postnatal Growth Standards (family =
\code{"ig_png"})
\itemize{
\item Between 27 and 64 weeks for \code{"wfa"}, \code{"lfa"}, and \code{"hcfa"}.
\item Between 35 and 65 cm for \code{"wfl"}.
}
}

By default, gigs will replace out-of-bounds elements in \code{x} with \code{NA} and
warn you. It will also warn you about elements in \code{x} which are missing
(\code{NA}) or undefined (\code{NaN}/\code{Inf}). You can customise this behaviour using
the \link[=gigs_options]{GIGS package-level options}.}

\item{sex}{Character vector of length one or more with sex(es), either \code{"M"}
(male) or \code{"F"} (female). This argument should be left as \code{NULL} (the
default) if \code{family == "ig_fet"}. This argument is case-sensitive. By
default, gigs will replace elements of \code{sex} which are not \code{"M"} or \code{"F"}
with \code{NA} and warn you. You can customise this  behaviour using the \link[=gigs_options]{GIGS package-level options}.}

\item{family}{A single string denoting the family of growth standards in use.
Must be one of \code{"ig_fet"}, \code{"ig_nbs"}, \code{"ig_png"}, or \code{"who_gs"}, or the
function will throw an error. This argument is case-sensitive.}

\item{acronym}{A single string denoting the specific growth standard in use.
The allowed values of \code{acronym} depend on the value of \code{family}, and are
listed in the documentation for the \code{x} argument. Incompatible \code{family} and
\code{acronym} values will cause errors to be thrown. This argument is
case-sensitive.}
}
\value{
Numeric vector of z-scores/centiles with length equal to the longest
input vector.
}
\description{
Convert anthropometric measures to z-scores/centiles using international
growth standards
}
\note{
Inputs other than \code{acronym} and \code{family} will be recycled by
\code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}, and must adhere to the \link{vctrs} recycling
rules.
}
\examples{
# Convert in the weight-for-age standard from the WHO Child Growth Standards
value2zscore(3, 0, "M", family = "who_gs", acronym = "wfa")

# Or obtain centiles...
value2centile(3, 0, "M", family = "who_gs", acronym = "wfa")

# Inputs will be recycled if necessary
value2centile(3, 0, c("M", "F"), family = "ig_nbs", acronym = "wfga")
}
\seealso{
\code{\link[=report_units]{report_units()}} will tell you the units needed for \code{y} and \code{x}, for
any combination of \code{family} and \code{acronym}.
}
