% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{who_gs}
\alias{who_gs}
\title{WHO Growth Standards tables}
\source{
\href{https://www.who.int/tools/child-growth-standards/standards}{WHO Child Growth Standards}
}
\description{
A set of nested lists containing tables with reference values at different z-scores/percentiles for valid
postmenstrual ages. The list is ordered by acronym first, then by sex and finally z-score/percentile.
}
\examples{
names(gigs::who_gs)
head(gigs::who_gs$wfa$male$zscores)
}
\references{
de Onis M, Garza C, Victora CG, Onyango AW, Frongillo EA, Martines J. \strong{The WHO Multicentre Growth Reference
Study: planning, study design, and methodology} \emph{Food Nutr Bull.} 2004, \strong{25(1 Suppl):S15-26.}
doi: \href{https://journals.sagepub.com/doi/10.1177/15648265040251S104}{10.1177/15648265040251s104}
}
\keyword{data}
