% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{who_gs_coeffs}
\alias{who_gs_coeffs}
\title{WHO Child Growth Standards LMS coefficients}
\source{
\href{https://www.who.int/tools/child-growth-standards/standards}{WHO Child Growth Standards}
}
\description{
A set of nested lists containing tables with LMS values for each sex in
different combinations of age/length-height/BMI. The list is ordered by
acronym first, then by sex.
}
\examples{
names(gigs::who_gs_coeffs)
head(gigs::who_gs_coeffs$lhfa$male)
}
\references{
World Health Organisation. \strong{WHO child growth standards:
length/height-for-age, weight-for-age, weight-for-length, weight-for-height
and body mass index-for-age: methods and development.} \emph{Technical report,
WHO, Geneva}, 2006.

World Health Organisation. \strong{WHO child growth standards: head
circumference-for-age, arm circumference-for-age, triceps skinfold-for-age
and subscapular skinfold-for-age: methods and development.} \emph{Technical
report, WHO, Geneva}, 2007.
}
\keyword{data}
