% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_gs.R
\name{who_gs_lms}
\alias{who_gs_lms}
\title{Get LMS values for WHO Growth Standards models}
\usage{
who_gs_lms(x, sex, acronym)
}
\arguments{
\item{x}{X value(s) at which to retrieve LMS values. Must be within bounds of available x values for given acronym.}

\item{sex}{Sex(es), either \code{"M"} (male) or \code{"F"} (female).}

\item{acronym}{Valid acronym for WHO Growth Standards datasets: one of \code{"wfa"} (weight-for-age), \code{"bfa"}
(bmi-for-age), \code{"lhfa"} (length/height-for-age), \code{"wfl"} (weight-for-length), or \code{"wfh"}
(weight-for-height), \code{"hcfa"} (head circumference-for-age), \code{"acfa"} (arm circumference-for-age),
\code{"ssfa"} (subscapular skinfold-for-age), or \code{"tsfa"} (triceps skinfold-for-age),}
}
\value{
A dataframe with lambda, mu and sigma values for the WHO growth standard(s) specified by the \code{acronym}
parameter, for the supplied \code{x} and \code{sex} values.
}
\description{
Get LMS values for WHO Growth Standards models
}
\references{
de Onis M, Garza C, Victora CG, Onyango AW, Frongillo EA, Martines J. \strong{The WHO Multicentre Growth Reference
Study: planning, study design, and methodology} \emph{Food Nutr Bull.} 2004, \strong{25(1 Suppl):S15-26.}
doi: \href{https://journals.sagepub.com/doi/10.1177/15648265040251S104}{10.1177/15648265040251s104}
}
\keyword{internal}
