% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_gs.R
\name{who_gs_lms2sd}
\alias{who_gs_lms2sd}
\title{Get standard deviations for WHO Growth Standards data from LMS values}
\usage{
who_gs_lms2sd(l, m, s, n_sd)
}
\arguments{
\item{l}{Lambda value as provided by \link{who_gs_lms}}

\item{m}{Mu value as provided by \link{who_gs_lms}}

\item{s}{Sigma value as provided by \link{who_gs_lms}}

\item{n_sd}{Number of SD from the median to be computed}
}
\description{
Get standard deviations for WHO Growth Standards data from LMS values
}
\references{
de Onis M \emph{et al.} (2006) \strong{'COMPUTATION OF CENTILES AND Z-SCORES FOR LENGTH/HEIGHT-FOR-AGE,
WEIGHT-FOR-AGE, WEIGHT-FOR-LENGTH, WEIGHT-FOR-HEIGHT AND BMI-FOR-AGE'} \emph{in} World Health Organization (ed.)
\emph{WHO Child Growth Standards: length/height-for-age, weight-for-age, weight-for-length, weight-for-height and
body mass index-for-age: methods and development}. Geneva: World Health Organization, pp. 301-304.
}
\keyword{internal}
