% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_gs.R
\name{who_gs_value2zscore}
\alias{who_gs_value2zscore}
\alias{who_gs_wfa_value2zscore}
\alias{who_gs_bfa_value2zscore}
\alias{who_gs_lhfa_value2zscore}
\alias{who_gs_wfl_value2zscore}
\alias{who_gs_wfh_value2zscore}
\alias{who_gs_hcfa_value2zscore}
\alias{who_gs_acfa_value2zscore}
\alias{who_gs_ssfa_value2zscore}
\alias{who_gs_tsfa_value2zscore}
\alias{who_gs_value2centile}
\alias{who_gs_wfa_value2centile}
\alias{who_gs_bfa_value2centile}
\alias{who_gs_lhfa_value2centile}
\alias{who_gs_wfl_value2centile}
\alias{who_gs_wfh_value2centile}
\alias{who_gs_hcfa_value2centile}
\alias{who_gs_acfa_value2centile}
\alias{who_gs_ssfa_value2centile}
\alias{who_gs_tsfa_value2centile}
\title{Convert values to z-scores/centiles in the WHO Child Growth Standards}
\usage{
who_gs_value2zscore(y, x, sex, acronym)

who_gs_wfa_value2zscore(weight_kg, age_days, sex)

who_gs_bfa_value2zscore(bmi, age_days, sex)

who_gs_lhfa_value2zscore(lenht_cm, age_days, sex)

who_gs_wfl_value2zscore(weight_kg, length_cm, sex)

who_gs_wfh_value2zscore(weight_kg, height_cm, sex)

who_gs_hcfa_value2zscore(headcirc_cm, age_days, sex)

who_gs_acfa_value2zscore(armcirc_cm, age_days, sex)

who_gs_ssfa_value2zscore(subscap_sf_mm, age_days, sex)

who_gs_tsfa_value2zscore(triceps_sf_mm, age_days, sex)

who_gs_value2centile(y, x, sex, acronym)

who_gs_wfa_value2centile(weight_kg, age_days, sex)

who_gs_bfa_value2centile(bmi, age_days, sex)

who_gs_lhfa_value2centile(lenht_cm, age_days, sex)

who_gs_wfl_value2centile(weight_kg, length_cm, sex)

who_gs_wfh_value2centile(weight_kg, height_cm, sex)

who_gs_hcfa_value2centile(headcirc_cm, age_days, sex)

who_gs_acfa_value2centile(armcirc_cm, age_days, sex)

who_gs_ssfa_value2centile(subscap_sf_mm, age_days, sex)

who_gs_tsfa_value2centile(triceps_sf_mm, age_days, sex)
}
\arguments{
\item{y}{Numeric vector of measurement(s) to convert to centiles/z-scores.
Units depend on which \code{acronym}(s) are in use.}

\item{x, age_days, length_cm, height_cm}{Numeric vector with \code{x} values at which
to convert inputs. Must be within bounds of available \code{x} values for given
acronym. The standard-specific versions of each function specify \code{x}:
either \code{age_days} (age since birth in days), \code{length_cm} (recumbent length
measurement(s) in cm) or \code{height_cm} (standing height measurement(s) in
cm).}

\item{sex}{Character vector of sex(es), either \code{"M"} (male) or \code{"F"}
(female).}

\item{acronym}{Valid acronym(s) for the WHO Child Growth standards: either
\code{"wfa"} (weight-for-age), \code{"bfa"} (BMI-for-age), \code{"lhfa"}
(length/height-for-age), \code{"wfl"} (weight-for-length), \code{"wfh"}
(weight-for-height), \code{"hcfa"} (head circumference-for-age), \code{"acfa"}
(arm circumference-for-age), \code{"ssfa"} (subscapular skinfold-for-age), or
\code{"tsfa"} (triceps skinfold-for-age).}

\item{weight_kg}{Numeric vector of weight measurement(s) in kg.}

\item{bmi}{Numeric vector of body mass index measurement(s) in
kg/m\if{html}{\out{<sup>}}2\if{html}{\out{</sup>}}.}

\item{lenht_cm}{Numeric vector of length/height measurement(s) in cm.}

\item{headcirc_cm}{Numeric vector of head circumference measurement(s) in cm.}

\item{armcirc_cm}{Numeric vector of arm circumference measurement(s) in cm.}

\item{subscap_sf_mm}{Numeric vector of subscapular skinfold measurement(s) in
mm.}

\item{triceps_sf_mm}{Numeric vector of triceps skinfold measurement(s) in mm.}
}
\value{
Numeric vector of expected measurements with length equal to the
outputs from \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}.
}
\description{
Convert values to z-scores/centiles in the WHO Child Growth Standards
}
\note{
Input vectors are recycled by \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}.
}
\examples{
# Convert values to centiles
who_gs_value2centile(y = 10.1, x = 505, sex = "M", acronym = "wfa") |>
  round(digits = 2)

# Or values to z-scores
who_gs_value2zscore(y = 10.1, x = 505, sex = "M", acronym = "wfa") |>
  round(digits = 2)

# Specify which standard to use with the acronym parameter...
who_gs_value2centile(y = 75.2, x = 300, sex = "M", acronym = "lhfa") |>
  round(digits = 2)

# ... or by using a standard-specific function
who_gs_lhfa_value2centile(lenht_cm = 75.2, age_days = 300, sex = "M") |>
  round(digits = 2)

# Inputs are recycled to the input of the longest length
who_gs_ssfa_value2centile(subscap_sf_mm = 6.1,
                          age_days = seq(450, 650, by = 50),
                          sex = "M") |>
  round(digits = 2)

# Bad inputs will not stop the function but will instead return NA - here
# 121 cm for height_cm is incompatible with the WHO Growth Standard for
# weight-for-height
who_gs_wfh_value2zscore(weight_kg = c(7.4, 14.2, 21.8, 22.4),
                        height_cm = c(65, 95, 120, 121),
                        sex = "M") |>
  round(digits = 2)
}
\references{
de Onis M, Garza C, Victora CG, Onyango AW, Frongillo EA, Martines J. \strong{The
WHO Multicentre Growth Reference Study: planning, study design, and
methodology} \emph{Food Nutr Bull.} 2004, \strong{25(1 Suppl):S15-26.}
\doi{10.1177/15648265040251s104}

World Health Organisation. \strong{WHO child growth standards:
length/height-for-age, weight-for-age, weight-for-length, weight-for-height
and body mass index-for-age: methods and development.} \emph{Technical report,
WHO, Geneva}, 2006.

World Health Organisation. \strong{WHO child growth standards: head
circumference-for-age, arm circumference-for-age, triceps skinfold-for-age
and subscapular skinfold-for-age: methods and development.} \emph{Technical
report, WHO, Geneva}, 2007.

Cole TJ. \strong{The LMS method for constructing normalized growth standards} \emph{Eur
J Clin Nutr.} 1990, \strong{44(1):45-60.} PMID:
\href{https://pubmed.ncbi.nlm.nih.gov/2354692/}{2354692}
}
