% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_categorise.R
\name{categorise_wfa}
\alias{categorise_wfa}
\title{Classify weight-for-age z-scores into weight-for-age strata}
\usage{
categorise_wfa(waz, outliers = FALSE)
}
\arguments{
\item{waz}{A numeric vector of length one or more containing
weight-for-age z-scores (WAZs).}

\item{outliers}{A single \code{TRUE} or \code{FALSE} value specifying whether
implausible z-score thresholds should be applied. Default = \code{FALSE}.}
}
\value{
An object of class factor with the same length as \code{waz}, containing
weight-for-age classifications. Its levels are \code{c("underweight_severe", "underweight", "normal", "overweight")} if \code{outliers = FALSE} (the
default), else \code{c("underweight_severe", "underweight", "normal", "overweight", "outlier")}.

\tabular{lll}{
\strong{Category}    \tab \strong{Factor level}  \tab
\strong{Z-score bounds} \cr
Severely underweight \tab \code{"underweight_severe"} \tab \code{waz} ≤ -3      \cr
Underweight          \tab \code{"underweight"}        \tab -3 < \code{waz} ≤ -2 \cr
Normal weight        \tab \code{"normal_weight"}      \tab \code{abs(waz)} < 2  \cr
Overweight           \tab \code{"overweight"}         \tab \code{waz} ≥ 2       \cr
Outlier              \tab \code{"outlier"}        \tab \code{waz} < -6 or \code{waz} > 5
}
}
\description{
Classify weight-for-age z-scores into weight-for-age strata
}
\examples{
waz <- c(-6.5, -3.5, -2.5, 0, 2.5, 3.5)
categorise_wfa(waz, outliers = FALSE)
categorise_wfa(waz, outliers = TRUE)
}
\references{
\strong{'Implausible z-score values'} \emph{in} World Health Organization (ed.)
\emph{Recommendations for data collection, analysis and reporting on
anthropometric indicators in children under 5 years old}. Geneva: World
Health Organization and the United Nations Children's Fund UNICEF, (2019).
pp. 64-65.

\strong{'Percentage of children stunted, wasted, and underweight, and mean z-scores
for stunting, wasting and underweight'} \emph{in} \emph{Guide to DHS Statistics DHS-7}
Rockville, Maryland, USA: ICF (2020). pp. 431-435.
\url{https://dhsprogram.com/data/Guide-to-DHS-Statistics/Nutritional_Status.htm}
}
