% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_classify.R
\name{classify_growth}
\alias{classify_growth}
\title{Classify multiple growth indicators at the same time using GIGS-recommended
growth standards}
\usage{
classify_growth(
  .data,
  gest_days,
  age_days,
  sex,
  weight_kg = NULL,
  lenht_cm = NULL,
  headcirc_cm = NULL,
  id = NULL,
  .outcomes = c("sfga", "svn", "stunting", "wasting", "wfa", "headsize"),
  .new = list(sfga = c("birthweight_centile", "sfga", "sfga_severe"), svn =
    c("birthweight_centile", "svn"), stunting = c("lhaz", "stunting",
    "stunting_outliers"), wasting = c("wlz", "wasting", "wasting_outliers"), wfa =
    c("waz", "wfa", "wfa_outliers"), headsize = c("hcaz", "headsize")),
  .verbose = TRUE
)
}
\arguments{
\item{.data}{A \code{data.frame}-like tabular object with one or more rows. Must
be permissible by \code{\link[checkmate:checkDataFrame]{checkmate::assert_data_frame()}}, so you can also supply
a \code{tibble}, \code{data.table}, or similar.}

\item{gest_days}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of gestational age(s) at birth
in days. This column, in conjunction with the column referred to by
\code{age_days}, is used to select which growth standard to use for each
observation.}

\item{age_days}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of age values in days.}

\item{sex}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a case-sensitive character vector of sexes,
either \code{"M"} (male) or \code{"F"} (female). By default, gigs will warn you if
any elements of \code{sex} are not \code{"M"} or \code{"F"}, or are missing (\code{NA}). You
can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}

\item{weight_kg}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of weight values in kg. When
performing size-for-GA and small vulnerable newborn classifications,
centiles and classifications will only be provided where measurements were
taken <12hrs after birth, i.e. \code{age_days < 0.5}.}

\item{lenht_cm}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of length/height values in cm.}

\item{headcirc_cm}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of head circumference values in
cm.}

\item{id}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a factor variable with IDs for each observation.
When not \code{NULL}, this variable is used to ensure that only the first
measurement taken from each infant is used as a birth measure. If all your
data is from one individual, leave this parameter as \code{NULL}. Default =
\code{NULL}.}

\item{.outcomes}{A character vector of up to six elements in length
describing which growth outcomes you want to assess. Use this when you
supply data that can be used to generate multiple growth indicators, but
only want to run a specific set. This argument is case-sensitive, and an
error will be thrown if you supply a string in \code{.analyses} which is either
\code{NA} or is not a member of this argument's default. Default = \code{c("sfga", "svn", "stunting", "wasting", "wfa", "headsize")}.}

\item{.new}{A list with names corresponding to \code{.analyses}, which describes
the names of new columns to be added to \code{.data}. If any elements of the
vectors in \code{.new} are equal to existing elements of \code{colnames(.data)}, the
function will throw an error. Excluding the birthweight centiles produced
when getting size-for-gestational age (\code{"sfga"}) and small vulnerable
newborn (\code{"svn"}) classifications, all strings in \code{.new} must be unique.
Each character vector in \code{.new} is repaired with \code{vctrs::as_names()}, which
will issue messages if any elements in \code{.new} are changed.

\describe{
\item{\code{"sfga"}:}{  \code{c("birthweight_centile", "sfga", "sfga_severe")}}
\item{\code{"svn"}:}{  \code{c("birthweight_centile", "svn")}}
\item{\code{"stunting"}:}{  \code{c("lhaz", "stunting", "stunting_outliers")}}
\item{\code{"wasting"}:}{  \code{c("wlz", "wasting", "wasting_outliers")}}
\item{\code{"wfa"}:}{  \code{c("waz", "wfa", "wfa_outliers")}}
\item{\code{"headsize"}:}{  \code{c("hcaz", "headsize")}}
}}

\item{.verbose}{A single logical value. When \code{TRUE} (the default),
messages about which analyses were requested versus which were performed
will be printed to the console. Warnings from \code{classify_growth()} will
still be printed even if \code{.verbose} is \code{FALSE}.}
}
\value{
A tabular object of the same class that was provided as \code{.data},
with new columns named according to \code{.new}. The new columns will depend
on the values supplied for \code{.outcomes}.
}
\description{
This function permits classification of multiple growth indicators (stunting,
wasting, weight-for-age, and more) at once in \code{data.frame}-like objects.
}
\note{
For size-for-GA and small vulnerable newborn analyses, centiles and
categorisations will only be applied for birth measurements. These are
considered to be the observation per level of \code{id} where \code{age_days} is
smallest, provided that \code{age_days} is \verb{<3}.

Categorical (factor) columns produced here may contain unused factor
levels. By default, gigs will inform you if these columns have unused
factor levels. You can change this behaviour using the
\link[=gigs_options]{GIGS package-level option}
\code{.gigs_options$handle_unused_levels}.
}
\examples{
# This dummy dataset contains data from two people, from birth (<3 days) to
# 500 days of age.
data <- data.frame(
  child = factor(rep.int(c("A", "B"), c(3, 3))),
  agedays = c(0, 100, 500, 2, 100, 500),
  gestage  = c(rep(35 * 7, 3), rep(35 * 7, 3)),
  sex = rep.int(c("M", "F"), c(3, 3)),
  wt_kg = c(3, 6, 9, 3, 6, 9),
  len_cm = rep.int(c(52.2, 60.4, 75), 2),
  head_cm = rep.int(c(30, 40, 49), 2)
)

data_classified <- classify_growth(data,
                                   age_days = agedays,
                                   gest_days = gestage,
                                   weight_kg = wt_kg,
                                   lenht_cm = len_cm,
                                   headcirc_cm = head_cm,
                                   sex = sex,
                                   id = child)

data_classified

# Use `.outcomes` to set which growth indicators will be computed
data_svn <- classify_growth(data,
                            age_days = agedays,
                            gest_days = gestage,
                            sex = sex,
                            weight_kg = wt_kg,
                            id = child,
                            .outcomes = "svn")

data_svn

# Use `.new` to specify new column names
data_svn <- classify_growth(data,
                            age_days = agedays,
                            gest_days = gestage,
                            sex = sex,
                            weight_kg = wt_kg,
                            id = child,
                            .outcomes = "svn",
                            .new = list("svn" = c("ig_nbs_centile",
                                                  "SVN_Category")))
data_svn

# The function will warn you if you don't provide data for your outcomes
data_svn_stunting <- classify_growth(data,
                                     age_days = agedays,
                                     gest_days = gestage,
                                     sex = sex,
                                     weight_kg = wt_kg,
                                     id = child,
                                     .outcomes = c("svn", "stunting"))
data_svn_stunting
}
\references{
WHO. \strong{Physical status: the use and interpretation of anthropometry. Report
of a WHO Expert Committee.} \emph{World Health Organisation Technical Report
Series 1995,} \strong{854: 1–452}

Royal College of Obstetricians and Gynaecologists. \strong{The Investigation and
Management of the Small-for-Gestational-Age Fetus: Green-top Guideline No.
31.} \emph{Technical report, Royal College of Obstetricians and Gynaecologists,
London, 2013.}

Lawn JE, Ohuma EO, Bradley E, Idueta LS, Hazel E, Okwaraji YB et al.
\strong{Small babies, big risks: global estimates of prevalence and mortality for
vulnerable newborns to accelerate change and improve counting.} \emph{The Lancet}
2023, \emph{401(10389):1707-1719.} \doi{10.1016/S0140-6736(23)00522-6}

\strong{'Implausible z-score values'} \emph{in} World Health Organization (ed.)
\emph{Recommendations for data collection, analysis and reporting on
anthropometric indicators in children under 5 years old}. Geneva: World
Health Organization and the United Nations Children's Fund UNICEF, (2019).
pp. 64-65.

\strong{'Percentage of children stunted, wasted, and underweight, and mean z-scores
for stunting, wasting and underweight'} \emph{in} \emph{Guide to DHS Statistics DHS-7}
Rockville, Maryland, USA: ICF (2020). pp. 431-435.
\url{https://dhsprogram.com/data/Guide-to-DHS-Statistics/Nutritional_Status.htm}

Victora CG, Schuler-Faccini L, Matijasevich A, Ribeiro E, Pessoa A,
Barros FC. \strong{Microcephaly in Brazil: how to interpret reported numbers?}
\emph{The Lancet} 2016, \emph{387(10019):621-624} \doi{10.1016/S0140-6736(16)00273-7}

Accogli A, Geraldo AF, Piccolo G, Riva A, Scala M, Balagura G, et al.
\strong{Diagnostic Approach to Macrocephaly in Children}. \emph{Frontiers in
Paediatrics} 2022, \emph{9:794069} \doi{10.3389/fped.2021.794069}
}
\seealso{
See \code{\link[=gigs_waz]{gigs_waz()}} to learn which growth standard will be used for
different combinations of gestational/chronological age.
}
