% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_classify.R
\name{classify_headsize}
\alias{classify_headsize}
\title{Classify head size in \code{data.frame}-like objects with GIGS-recommended
growth standards}
\usage{
classify_headsize(
  .data,
  headcirc_cm,
  age_days,
  gest_days,
  sex,
  id = NULL,
  .new = c("hcaz", "headsize")
)
}
\arguments{
\item{.data}{A \code{data.frame}-like tabular object with one or more rows. Must
be permissible by \code{\link[checkmate:checkDataFrame]{checkmate::assert_data_frame()}}, so you can also supply
a \code{tibble}, \code{data.table}, or similar.}

\item{headcirc_cm}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of head circumference values in
cm.}

\item{age_days}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of age values in days.}

\item{gest_days}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of gestational age(s) at birth
in days. This column, in conjunction with the column referred to by
\code{age_days}, is used to select which growth standard to use for each
observation.}

\item{sex}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a case-sensitive character vector of sexes,
either \code{"M"} (male) or \code{"F"} (female). By default, gigs will warn you if
any elements of \code{sex} are not \code{"M"} or \code{"F"}, or are missing (\code{NA}). You
can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}

\item{id}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a factor variable with IDs for each observation.
When not \code{NULL}, this variable is used to ensure that only the first
measurement taken from each infant is used as a birth measure. If all your
data is from one individual, leave this parameter as \code{NULL}. Default =
\code{NULL}.}

\item{.new}{A three-length character vector with names for the output
columns. These inputs will be repaired if necessary using
\code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}, which will print any changes to the console. If
any elements in \code{.new} are the same as elements in \code{colnames(.data)}, the
function will throw an error. Default = \code{c("hcaz", "headsize")}.}
}
\value{
A tabular object of the same class that was provided as \code{.data},
with three new columns named according to \code{.new}. These columns will be
(from left to right):
\itemize{
\item \code{hcaz} - Numeric vector of weight-for-length/height zscores
\item \code{headsize} - Factor of head size categories
}
}
\description{
Classify head size in \code{data.frame}-like objects with GIGS-recommended
growth standards
}
\details{
Cut-offs for head size categories are:

\tabular{lll}{
\strong{Category}    \tab \strong{Factor level}  \tab
\strong{Z-score bounds} \cr
Severe microcephaly       \tab \code{"microcephaly_severe"} \tab \code{hcaz} =< -3      \cr
Microcephaly              \tab \code{"microcephaly"}        \tab -3 < \code{hcaz} =< -2 \cr
Normal head circumference \tab \code{"normal_headcirc"}     \tab \code{abs(hcaz)} < 2   \cr
Macrocephaly              \tab \code{"macrocephaly"}        \tab \code{hcaz} >= 2       \cr
Severe macrocephaly       \tab \code{"macrocephaly_severe"} \tab \code{hcaz} >= 3
}
}
\note{
Categorical (factor) columns produced here may contain unused factor
levels. By default, gigs will inform you if these columns have unused
factor levels. You can change this behaviour using the
\link[=gigs_options]{GIGS package-level option}
\code{.gigs_options$handle_unused_levels}.
}
\examples{
# This dummy dataset contains data from two people, from birth (<3 days) to
# 500 days of age.
data <- data.frame(
  child = factor(rep.int(c("A", "B"), c(3, 3))),
  agedays = c(0, 100, 500, 2, 100, 500),
  gestage  = c(rep(35 * 7, 3), rep(35 * 7, 3)),
  sex = rep.int(c("M", "F"), c(3, 3)),
  headcirc_cm = rep.int(c(40, 50, 60), 2)
)

# Use the `id` argument to ensure that `classify_headsize()` uses the correct
# standard for each observation
data |>
  classify_headsize(headcirc_cm = headcirc_cm,
                    age_days = agedays,
                    gest_days = gestage,
                    sex = sex,
                    id = child)

# If you don't specify `id`, `classify_headsize()` will assume data is from
# one child only
data |>
  classify_headsize(headcirc_cm = headcirc_cm,
                    age_days = agedays,
                    gest_days = gestage,
                    sex = sex)
}
\references{
Victora CG, Schuler-Faccini L, Matijasevich A, Ribeiro E, Pessoa A,
Barros FC. \strong{Microcephaly in Brazil: how to interpret reported numbers?}
\emph{The Lancet} 2016, \emph{387(10019):621-624} \doi{10.1016/S0140-6736(16)00273-7}

Accogli A, Geraldo AF, Piccolo G, Riva A, Scala M, Balagura G, et al.
\strong{Diagnostic Approach to Macrocephaly in Children}. \emph{Frontiers in
Paediatrics} 2022, \emph{9:794069} \doi{10.3389/fped.2021.794069}
}
\seealso{
See \code{\link[=classify_growth]{classify_growth()}} to run this analysis and others at the same
time. See \code{\link[=gigs_waz]{gigs_waz()}} to learn which growth standard will be used for
different combinations of gestational/chronological age.
}
