% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_classify.R
\name{classify_stunting}
\alias{classify_stunting}
\title{Classify stunting using INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} or WHO
length/height-for-age standards}
\usage{
classify_stunting(
  .data,
  lenht_cm,
  age_days,
  gest_days,
  sex,
  .new = c("lhaz", "stunting", "stunting_outliers")
)
}
\arguments{
\item{.data}{A \code{data.frame}-like tabular object with one or more rows. Must
be permissible by \code{\link[checkmate:checkDataFrame]{checkmate::assert_data_frame()}}, so you can also supply
a \code{tibble}, \code{data.table}, or similar.}

\item{lenht_cm}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of length/height values in cm.}

\item{age_days}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of age values in days.}

\item{gest_days}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of gestational age(s) at birth
in days. This column, in conjunction with the column referred to by
\code{age_days}, is used to select which growth standard to use for each
observation.}

\item{sex}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a case-sensitive character vector of sexes,
either \code{"M"} (male) or \code{"F"} (female). By default, gigs will warn you if
any elements of \code{sex} are not \code{"M"} or \code{"F"}, or are missing (\code{NA}). You
can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}

\item{.new}{A three-length character vector with names for the output
columns. These inputs will be repaired if necessary using
\code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}, which will print any changes to the console. If
any elements in \code{.new} are the same as elements in \code{colnames(.data)}, the
function will throw an error. Default = \code{c("lhaz", "stunting", "stunting_outliers")}.}
}
\value{
A tabular object of the same class that was provided as \code{.data},
with three new columns named according to \code{.new}. These columns will be
(from left to right):
\itemize{
\item \code{lhaz} - Numeric vector of length/height-for-age zscores
\item \code{stunting} - Factor of stunting categories without outlier flagging
\item \code{stunting_outliers} - Factor of stunting categories with outlier flagging
}
}
\description{
Classify stunting using INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} or WHO
length/height-for-age standards
}
\examples{
# The first observation uses the INTERGROWTH-21st Postnatal Growth standards;
# the next two use the WHO Child Growth Standards.
data <- data.frame(
  length_cm = c(52.2, 75.4, 63.1),
  agedays = c(357, 375, 250),
  gestage  = c(196, 287, 266),
  sex = c("M", "M", "F")
)

data |>
  classify_stunting(lenht_cm = length_cm,
                    age_days = agedays,
                    gest_days = gestage,
                    sex = sex)
}
\references{
\strong{'Implausible z-score values'} \emph{in} World Health Organization (ed.)
\emph{Recommendations for data collection, analysis and reporting on
anthropometric indicators in children under 5 years old}. Geneva: World
Health Organization and the United Nations Children's Fund UNICEF, (2019).
pp. 64-65.

\strong{'Percentage of children stunted, wasted, and underweight, and mean z-scores
for stunting, wasting and underweight'} \emph{in} \emph{Guide to DHS Statistics DHS-7}
Rockville, Maryland, USA: ICF (2020). pp. 431-435.
\url{https://dhsprogram.com/data/Guide-to-DHS-Statistics/Nutritional_Status.htm}
}
\seealso{
Check out the \code{\link[=categorise_stunting]{categorise_stunting()}} documentation to see the
stunting categorisation thresholds.
}
