% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify_stunting}
\alias{classify_stunting}
\title{Classify stunting using WHO or INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
length/height-for-age standards}
\usage{
classify_stunting(lenht_cm, age_days, gest_days, sex, outliers = FALSE)
}
\arguments{
\item{lenht_cm}{Numeric vector with length/height measurement(s) in cm.}

\item{age_days}{Numeric vector with age(s) in days for each child. Should be
between \code{0} to \code{1856} days.}

\item{gest_days}{Numeric vector with gestational age(s) at birth in days.}

\item{sex}{Character vector of sex(es), either \code{"M"} (male) or \code{"F"}
(female).}

\item{outliers}{A single \code{TRUE} or \code{FALSE} value specifying whether
implausible z-score value thresholds should be applied. Default = \code{FALSE}.}
}
\value{
Factor of stunting classification(s) with levels
\code{c("stunting_severe", "stunting", "normal")} or \code{c("stunting_severe", "stunting", "normal", "outlier")}, depending on whether the \code{outliers}
argument is set to \code{TRUE}.
}
\description{
Classify stunting (low length/height-for-age) using WHO or
INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} length/height-for-age standards depending on the
gestational age at birth of the child. Severe stunting is below <-3 SD
relative to median length/height at a given age, whereas moderate stunting is
-2SD from the median.
}
\note{
Input vectors will be recycled by \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}. The
function assumes that your measurements were taken according to the WHO
guidelines, which stipulate that recumbent length should not be measured
after 730 days. Implausible z-score bounds are sourced from the referenced
WHO report, and classification cut-offs from the DHS manual.
}
\examples{
# The first observation uses the INTERGROWTH-21st Postnatal Growth standards;
# the next two use the WHO Child Growth Standards.
classify_stunting(
  lenht_cm = c(52.2, 75.4, 63.1),
  age_days = c(357, 375, 250),
  gest_days = c(196, 287, 266),
  sex = c("M", "M", "F")
)

# And with outlier flagging:
classify_stunting(
  lenht_cm = c(52.2, 75.4, 63.1),
  age_days = c(357, 375, 250),
  gest_days = c(196, 287, 266),
  sex = c("M", "M", "F"),
  outliers = TRUE
)
}
\references{
\strong{'Implausible z-score values'} \emph{in} World Health Organization (ed.)
\emph{Recommendations for data collection, analysis and reporting on
anthropometric indicators in children under 5 years old}. Geneva: World
Health Organization and the United Nations Children's Fund UNICEF, (2019).
pp. 64-65.

\strong{'Percentage of children stunted, wasted, and underweight, and mean z-scores
for stunting, wasting and underweight'} \emph{in} \emph{Guide to DHS Statistics DHS-7}
Rockville, Maryland, USA: ICF (2020). pp. 431-435.
\url{https://dhsprogram.com/data/Guide-to-DHS-Statistics/Nutritional_Status.htm}
}
