% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify_stunting}
\alias{classify_stunting}
\title{Classify stunting using WHO or INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
length/height-for-age standards}
\usage{
classify_stunting(lenht_cm, age_days, gest_days, sex, outliers = FALSE)
}
\arguments{
\item{lenht_cm}{Numeric vector of length one or more with length/height
measurement(s) in cm.}

\item{age_days}{Numeric vector of length one or more with age(s) in days
for each child. Should be between \code{0} to \code{1856} days. By default, gigs will
replace out-of-bounds elements in \code{age_days} with \code{NA} and warn you. This
behaviour can be customised using the functions in \link{gigs_options}.}

\item{gest_days}{Numeric vector with gestational age(s) at birth in days.}

\item{sex}{Character vector of length one or more with sex(es), either \code{"M"}
(male) or \code{"F"} (female). This argument is case-sensitive. By default, gigs
will replace elements of \code{sex} which are not \code{"M"} or \code{"F"} with \code{NA}
and warn you. This behaviour can be customised using the functions in
\link{gigs_options}.}

\item{outliers}{A single \code{TRUE} or \code{FALSE} value specifying whether
implausible z-score value thresholds should be applied. Default = \code{FALSE}.}
}
\value{
An object of class factor with the same length as the longest input
vector, containing stunting classifications. Its levels are
\code{c("stunting_severe", "stunting", "not_stunting")} if \code{outliers = FALSE} (the default), else \code{c("stunting_severe", "stunting", "not_stunting", "outlier")}.

\tabular{lll}{
\strong{Category} \tab \strong{Factor level} \tab
\strong{Z-score bounds} \cr
Severe stunting \tab \code{"stunting_severe"}     \tab \code{z} ≤ -3         \cr
Stunting        \tab \code{"stunting"}            \tab -3 < \code{z} ≤ -2    \cr
No stunting     \tab \code{"not_stunting"}        \tab \code{z} > -2         \cr
Outlier         \tab \code{"outlier"}             \tab \code{abs(z)} > 5
}
}
\description{
Classify stunting (low length/height-for-age) using WHO or
INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} length/height-for-age standards depending on
the gestational age at birth of the child. Severe stunting is below <-3 SD
relative to mean length/height at a given age, whereas moderate stunting is
-2 SD from the mean.
}
\note{
Input vectors are recycled by \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}, and must
adhere to the \link{vctrs} recycling rules. This function assumes that your
measurements were taken according to the WHO guidelines, which stipulate
that recumbent length should not be measured after 730 days. Instead,
standing height should be used. Implausible z-score bounds are sourced from
the referenced WHO report, and classification cut-offs from the DHS manual.
}
\examples{
# The first observation uses the INTERGROWTH-21st Postnatal Growth standards;
# the next two use the WHO Child Growth Standards.
classify_stunting(
  lenht_cm = c(52.2, 75.4, 63.1),
  age_days = c(357, 375, 250),
  gest_days = c(196, 287, 266),
  sex = c("M", "M", "F")
)

# And with outlier flagging:
classify_stunting(
  lenht_cm = c(52.2, 75.4, 63.1),
  age_days = c(357, 375, 250),
  gest_days = c(196, 287, 266),
  sex = c("M", "M", "F"),
  outliers = TRUE
)
}
\references{
\strong{'Implausible z-score values'} \emph{in} World Health Organization (ed.)
\emph{Recommendations for data collection, analysis and reporting on
anthropometric indicators in children under 5 years old}. Geneva: World
Health Organization and the United Nations Children's Fund UNICEF, (2019).
pp. 64-65.

\strong{'Percentage of children stunted, wasted, and underweight, and mean z-scores
for stunting, wasting and underweight'} \emph{in} \emph{Guide to DHS Statistics DHS-7}
Rockville, Maryland, USA: ICF (2020). pp. 431-435.
\url{https://dhsprogram.com/data/Guide-to-DHS-Statistics/Nutritional_Status.htm}
}
