% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_classify.R
\name{classify_svn}
\alias{classify_svn}
\title{Classify small vulnerable newborns in \code{data.frame}-like objects with the
INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} weight-for-gestational age standard}
\usage{
classify_svn(
  .data,
  weight_kg,
  gest_days,
  sex,
  .new = c("birthweight_centile", "svn")
)
}
\arguments{
\item{.data}{A \code{data.frame}-like tabular object with one or more rows. Must
be permissible by \code{\link[checkmate:checkDataFrame]{checkmate::assert_data_frame()}}, so you can also supply
a \code{tibble}, \code{data.table}, or similar.}

\item{weight_kg}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of birth weight values in kg.
It is assumed that weight measurements provided to this function are
birth weights recorded <12 hours after an infant's birth.}

\item{gest_days}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of gestational age(s) at birth
in days between \code{168} and \code{300}. By default, gigs will warn you about
elements of this vector which are outside these bounds, are \code{NA}, or \code{Inf}.
You can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}

\item{sex}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a case-sensitive character vector of sexes,
either \code{"M"} (male) or \code{"F"} (female). By default, gigs will warn you if
any elements of \code{sex} are not \code{"M"} or \code{"F"}, or are missing (\code{NA}). You
can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}

\item{.new}{A two-length character vector with names for the output
columns. These inputs will be repaired if necessary using
\code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}, which will print any changes to the console. If
any elements in \code{.new} are the same as elements in \code{colnames(.data)}, the
function will throw an error. Default = \code{c("birthweight_centile", "svn")}.}
}
\value{
A tabular object of the same class that was provided as \code{.data},
with two new columns named according to \code{.new}. These columns will be
(from left to right):
\itemize{
\item \code{birthweight_centile} - Numeric vector of birthweight centiles from the
INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Newborn Size standard for
weight-for-GA
\item \code{svn} - Factor of small vulnerable newborn (SVN) categories
}
}
\description{
Classify small vulnerable newborns in \code{data.frame}-like objects with the
INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} weight-for-gestational age standard
}
\details{
Cut-offs for small, vulnerable newborn categorisations are:

\tabular{llll}{
\strong{SVN category} \tab \strong{Factor level} \tab
\strong{Gestational age range} \tab \strong{Centile bounds} \cr
Preterm-SGA \tab \code{"Preterm SGA"} \tab \code{p} < 0.1       \tab
\code{gest_days} < 259  \cr
Preterm-AGA \tab \code{"Preterm AGA"} \tab 0.1 ≤ \code{p} ≤ 0.9 \tab
\code{gest_days} < 259  \cr
Preterm-LGA \tab \code{"Preterm LGA"} \tab \code{p} > 0.9       \tab
\code{gest_days} < 259  \cr
Term-SGA \tab \code{"Term SGA"}    \tab \code{p} < 0.1       \tab
\code{gest_days} ≥ 259  \cr
Term-AGA \tab \code{"Term AGA"}    \tab 0.1 ≤ \code{p} ≤ 0.9 \tab
\code{gest_days} ≥ 259  \cr
Term-LGA \tab \code{"Term LGA"}    \tab \code{p} > 0.9       \tab
\code{gest_days} ≥ 259
}

\emph{Abbreviations:} SGA, small-for-gestational age; AGA,
appropriate-for-gestational age; LGA, large-for-gestational age.
}
\examples{
data <- data.frame(
  wtkg = c(1.5, 2.6, 2.6, 3.5),
  gestage = c(235, 257, 275, 295),
  sex = c("F", "M", "F", "M")
)

data |>
  classify_svn(weight_kg = wtkg,
               gest_days = gestage,
               sex = sex)
}
\references{
WHO. \strong{Physical status: the use and interpretation of anthropometry. Report
of a WHO Expert Committee.} \emph{World Health Organisation Technical Report
Series 1995,} \strong{854: 1–452}

Royal College of Obstetricians and Gynaecologists. \strong{The Investigation and
Management of the Small-for-Gestational-Age Fetus: Green-top Guideline No.
31.} \emph{Technical report, Royal College of Obstetricians and Gynaecologists,
London, 2013.}
}
\seealso{
\code{\link[=ig_nbs_wfga_value2centile]{ig_nbs_wfga_value2centile()}}, which this function uses to get
a centile for each observation.
}
