% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_classify.R
\name{classify_wasting}
\alias{classify_wasting}
\title{Classify wasting in \code{data.frame}-like objects with GIGS-recommended growth
standards}
\usage{
classify_wasting(
  .data,
  weight_kg,
  lenht_cm,
  age_days,
  gest_days,
  sex,
  .new = c("wlz", "wasting", "wasting_outliers")
)
}
\arguments{
\item{.data}{A \code{data.frame}-like tabular object with one or more rows. Must
be permissible by \code{\link[checkmate:checkDataFrame]{checkmate::assert_data_frame()}}, so you can also supply
a \code{tibble}, \code{data.table}, or similar.}

\item{weight_kg}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of weight values in kg.}

\item{lenht_cm}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of length/height values in cm.}

\item{age_days}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of age values in days.}

\item{gest_days}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of gestational age(s) at birth
in days. This column, in conjunction with the column referred to by
\code{age_days}, is used to select which growth standard to use for each
observation.}

\item{sex}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a case-sensitive character vector of sexes,
either \code{"M"} (male) or \code{"F"} (female). By default, gigs will warn you if
any elements of \code{sex} are not \code{"M"} or \code{"F"}, or are missing (\code{NA}). You
can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}

\item{.new}{A three-length character vector with names for the output
columns. These inputs will be repaired if necessary using
\code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}, which will print any changes to the console. If
any elements in \code{.new} are the same as elements in \code{colnames(.data)}, the
function will throw an error. Default = \code{c("wlz", "wasting", "wasting_outliers")}.}
}
\value{
A tabular object of the same class that was provided as \code{.data},
with three new columns named according to \code{.new}. These columns will be
(from left to right):
\itemize{
\item \code{wlz} - Numeric vector of weight-for-length/height zscores
\item \code{wasting} - Factor of wasting categories without outlier flagging
\item \code{wasting_outliers} - Factor of wasting categories with outlier flagging
}
}
\description{
Classify wasting in \code{data.frame}-like objects with GIGS-recommended growth
standards
}
\details{
Cut-offs for wasting categories are:

\tabular{lll}{
\strong{Category} \tab \strong{Factor level} \tab
\strong{Z-score bounds} \cr
Severe wasting \tab \code{"wasting_severe"}       \tab \code{wlz} ≤ -3      \cr
Wasting        \tab \code{"wasting"}              \tab -3 < \code{wlz} ≤ -2 \cr
No wasting     \tab \code{"not_wasting"}          \tab \code{abs(wlz)} < 2  \cr
Overweight     \tab \code{"overweight"}           \tab \code{wlz} ≥ 2       \cr
Outlier        \tab \code{"outlier"}              \tab \code{abs(wlz)} > 5
}
}
\examples{
# Returns factor with stunting classifications
data <- data.frame(
  wt_kg = c(5.75, 2.18, 3.00, 6.75),
  length_height = c(67.7, 46.6, 50.0, 80.1),
  gestage = c(251, 197, 225, 243),
  age = c(251, 197, 225, 243),
  sex =  c("F", "M", "F", "M")
)
data |>
  classify_wasting(weight_kg = wt_kg,
                   lenht_cm = length_height,
                   gest_days = gestage,
                   age_days = age,
                   sex = sex)
}
\references{
\strong{'Implausible z-score values'} \emph{in} World Health Organization (ed.)
\emph{Recommendations for data collection, analysis and reporting on
anthropometric indicators in children under 5 years old}. Geneva: World
Health Organization and the United Nations Children's Fund UNICEF, (2019).
pp. 64-65.

\strong{'Percentage of children stunted, wasted, and underweight, and mean z-scores
for stunting, wasting and underweight'} \emph{in} \emph{Guide to DHS Statistics DHS-7}
Rockville, Maryland, USA: ICF (2020). pp. 431-435.
\url{https://dhsprogram.com/data/Guide-to-DHS-Statistics/Nutritional_Status.htm}
}
