% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify_wasting}
\alias{classify_wasting}
\title{Classify wasting using INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} weight-for-length or WHO
weight-for-length/height standards}
\usage{
classify_wasting(
  weight_kg,
  lenht_cm,
  gest_days,
  age_days,
  sex,
  outliers = FALSE
)
}
\arguments{
\item{weight_kg}{Numeric vector of length one or greater with weight
value(s) in kg.}

\item{lenht_cm}{Numeric vector of length one or greater  with length/height
measurement(s) in cm.}

\item{gest_days}{Numeric vector of length one or greater with gestational
age(s) at birth in days.}

\item{age_days}{Numeric vector of length one or greater  with age(s) in days
for each child. Should be between \code{0} to \code{1856} days.}

\item{sex}{Character vector of length one or more with sex(es), either \code{"M"}
(male) or \code{"F"} (female). This argument is case-sensitive. By default, gigs
will replace elements of \code{sex} which are not \code{"M"} or \code{"F"} with \code{NA}
and warn you. This behaviour can be customised using the functions in
\link{gigs_options}.}

\item{outliers}{A single \code{TRUE} or \code{FALSE} value specifying whether
implausible z-score value thresholds should be applied. Default = \code{FALSE}.}
}
\value{
Factor of wasting classifications with same length as longest input.
If \code{outliers = FALSE}, levels are \code{c("wasting_severe", "wasting", "not_wasting", "overweight")}. If \code{outliers = TRUE}, levels are
\code{c("wasting_severe", "wasting", "not_wasting", "overweight", "outlier")}.
}
\description{
Classify wasting (low weight-for-length/height) using the
INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} weight-for-length or WHO Child Growth standards,
specifically either the weight-for-length or weight-for-height standard
depending on the age of the child. Severe wasting is <-3SD relative to the
mean expected weight, whereas moderate wasting is -2SD from the mean.
}
\note{
Input vectors will be recycled by \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}.
Implausible z-score bounds are sourced from the referenced WHO report, and
classification cut-offs from the DHS manual.
}
\examples{
# Returns factor with stunting classifications
classify_wasting(
  weight_kg = c(5.75, 2.18, 3.00, 6.75),
  lenht_cm = c(67.7, 46.6, 50.0, 80.1),
  gest_days = c(251, 197, 225, 243),
  age_days = c(251, 197, 225, 243),
  sex =  c("F", "M", "F", "M")
)

# Outliers  can be flagged if `outliers` set to TRUE
classify_wasting(
  weight_kg = c(5.75, 2.18, 3.00, 6.75),
  lenht_cm = c(67.7, 46.6, 50.0, 80.1),
  gest_days = c(251, 197, 225, 243),
  age_days = c(251, 197, 225, 243),
  sex =  c("F", "M", "F", "M"),
  outliers = TRUE
)
}
\references{
\strong{'Implausible z-score values'} \emph{in} World Health Organization (ed.)
\emph{Recommendations for data collection, analysis and reporting on
anthropometric indicators in children under 5 years old}. Geneva: World
Health Organization and the United Nations Children's Fund UNICEF, (2019).
pp. 64-65.

\strong{'Percentage of children stunted, wasted, and underweight, and mean z-scores
for stunting, wasting and underweight'} \emph{in} \emph{Guide to DHS Statistics DHS-7}
Rockville, Maryland, USA: ICF (2020). pp. 431-435.
\url{https://dhsprogram.com/data/Guide-to-DHS-Statistics/Nutritional_Status.htm}
}
