% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify_wasting}
\alias{classify_wasting}
\title{Classify wasting using WHO weight-for-length/height standards}
\usage{
classify_wasting(weight_kg, lenht_cm, sex, lenht_method)
}
\arguments{
\item{weight_kg}{Numeric vector with weight value(s) in kg.}

\item{lenht_cm}{Numeric vector with length/height measurement(s) in cm.}

\item{sex}{Character vector of sex(es), either \code{"M"} (male) or \code{"F"}
(female).}

\item{lenht_method}{Character vector containig \code{"H"} or \code{"L"} value(s)
describing whether \code{lenht_cm} was recorded as recumbent length or standing
height. Missing (\code{NA}) values will be set to \code{"L"} for children <731 days
old and to \code{"H"} for children 731 days old or more. Default = \code{NA}.}
}
\value{
Factor of wasting classification(s) with levels \code{c("implausible", "wasting_severe", "wasting", "normal", "overweight")}.
}
\description{
Classify wasting (low weight-for-length/height) using the WHO Child Growth
Standards, specifically either the weight-for-length or weight-for-height
standard depending on the age of the child. Severe wasting is <-3SD relative
to the median expected weight, whereas moderate wasting is -2SD from the
median.
}
\note{
Input vectors will be recycled by \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}.
Implausible z-score bounds are sourced from the referenced WHO report, and
classification cut-offs from the DHS manual. Observations with invalid or
missing \code{lenht_method} values will be returned as \code{NA}.
}
\examples{
# Where no lenht_method is given, classify_wasting() returns NA.
classify_wasting(
  weight_kg = c(5.75, 2.18, 3.00, 6.75),
  lenht_cm = c(67.7, 46.6, 50.0, 80.1),
  sex =  c("F", "M", "F", "M"),
  lenht_method = c("H", "L", NA_character_, "H")
)
}
\references{
\strong{'Implausible z-score values'} \emph{in} World Health Organization (ed.)
\emph{Recommendations for data collection, analysis and reporting on
anthropometric indicators in children under 5 years old}. Geneva: World
Health Organization and the United Nations Children's Fund UNICEF, (2019).
pp. 64-65.

\strong{'Percentage of children stunted, wasted, and underweight, and mean z-scores
for stunting, wasting and underweight'} \emph{in} \emph{Guide to DHS Statistics DHS-7}
Rockville, Maryland, USA: ICF (2020). pp. 431-435.
\url{https://dhsprogram.com/data/Guide-to-DHS-Statistics/Nutritional_Status.htm}
}
