% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify_wfa}
\alias{classify_wfa}
\title{Classify weight-for-age according to INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} postnatal
growth and WHO growth standards}
\usage{
classify_wfa(weight_kg, age_days, ga_at_birth, sex)
}
\arguments{
\item{weight_kg}{Weight measurement(s) in kg.}

\item{age_days}{Age(s) at recording of each infant in days. Should be between
\code{0} and \code{1856}.}

\item{ga_at_birth}{Gestational age(s) at birth in days.}

\item{sex}{Sex(es), either \code{"M"} (male) or \code{"F"} (female).}
}
\value{
Factor of weight classification(s) with levels \code{c("implausible", "underweight_severe", "underweight", "normal", "overweight")}.
}
\description{
Classify weight-for-age z-scores using the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
Postnatal Growth standards or WHO Child Growth Standards depending on the
gestational age at birth for the infant. Severely underweight is less than 3
SD below the median, underweight is less than 2 SD below the median, and
overweight is > 2 SDs above the median.
}
\note{
Implausible z-score bounds are sourced from the referenced WHO report,
and classification criteria from the DHS manual.
}
\examples{
classify_wfa(
  weight_kg = c(7.2, 4.5, 9.1, 24),
  age_days = c(401, 185, 101, 607),
  ga_at_birth = 7 * c(27, 36, 40, 41),
  sex = c("F", "M", "F", "M")
)
}
\references{
\strong{'Implausible z-score values'} \emph{in} World Health Organization (ed.)
\emph{Recommendations for data collection, analysis and reporting on
anthropometric indicators in children under 5 years old}. Geneva: World
Health Organization and the United Nations Children’s Fund UNICEF, (2019).
pp. 64-65.

\strong{'Percentage of children stunted, wasted, and underweight, and mean z-scores
for stunting, wasting and underweight'} \emph{in} \emph{Guide to DHS Statistics DHS-7}
Rockville, Maryland, USA: ICF (2020). pp. 431-435.
\url{https://dhsprogram.com/data/Guide-to-DHS-Statistics/Nutritional_Status.htm}
}
