% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_classify.R
\name{classify_wfa}
\alias{classify_wfa}
\title{Classify weight-for-age in \code{data.frame}-like objects with GIGS-recommended
growth standards}
\usage{
classify_wfa(
  .data,
  weight_kg,
  age_days,
  gest_days,
  sex,
  id = NULL,
  .new = c("waz", "wfa", "wfa_outliers")
)
}
\arguments{
\item{.data}{A \code{data.frame}-like tabular object with one or more rows. Must
be permissible by \code{\link[checkmate:checkDataFrame]{checkmate::assert_data_frame()}}, so you can also supply
a \code{tibble}, \code{data.table}, or similar.}

\item{weight_kg}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of birth weight values in kg.
It is assumed that weight measurements provided to this function are
birth weights recorded <12 hours after an infant's birth.}

\item{age_days}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of age values in days.}

\item{gest_days}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a numeric vector of gestational age(s) at birth
in days. This column, in conjunction with the column referred to by
\code{age_days}, is used to select which growth standard to use for each
observation.}

\item{sex}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a case-sensitive character vector of sexes,
either \code{"M"} (male) or \code{"F"} (female). By default, gigs will warn you if
any elements of \code{sex} are not \code{"M"} or \code{"F"}, or are missing (\code{NA}). You
can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}

\item{id}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
column in \code{.data} which is a factor variable with IDs for each observation.
When not \code{NULL}, this variable is used to ensure that only the first
measurement taken from each infant is used as a birth measure. If all your
data is from one individual, leave this parameter as \code{NULL}. Default =
\code{NULL}.}

\item{.new}{A three-length character vector with names for the output
columns. These inputs will be repaired if necessary using
\code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}, which will print any changes to the console. If
any elements in \code{.new} are the same as elements in \code{colnames(.data)}, the
function will throw an error. Default = \code{c("waz", "wfa", "wfa_outliers")}.}
}
\value{
A tabular object of the same class that was provided as \code{.data},
with three new columns named according to \code{.new}. These columns will be
(from left to right):
\itemize{
\item \code{wlz} - Numeric vector of weight-for-length/height zscores
\item \code{wfa} - Factor of weight-for-age categories without outlier flagging
\item \code{wfa_outliers} - Factor of weight-for-age categories with outlier
flagging
}
}
\description{
Classify weight-for-age in \code{data.frame}-like objects with GIGS-recommended
growth standards
}
\details{
Cut-offs for weight-for-age categories are:

\tabular{lll}{
\strong{Category}    \tab \strong{Factor level}  \tab
\strong{Z-score bounds} \cr
Severely underweight \tab \code{"underweight_severe"} \tab \code{waz} =< -3      \cr
Underweight          \tab \code{"underweight"}        \tab -3 < \code{waz} =< -2 \cr
Normal weight        \tab \code{"normal_weight"}      \tab \code{abs(waz)} < 2  \cr
Overweight           \tab \code{"overweight"}         \tab \code{waz} >= 2      \cr
Outlier              \tab \code{"outlier"}        \tab \code{waz} < -6 or \code{waz} > 5
}
}
\note{
Categorical (factor) columns produced here may contain unused factor
levels. By default, gigs will inform you if these columns have unused
factor levels. You can change this behaviour using the
\link[=gigs_options]{GIGS package-level option}
\code{.gigs_options$handle_unused_levels}.
}
\examples{
# This dummy dataset contains data from two people, from birth (<3 days) to
# 500 days of age.
data <- data.frame(
  child = factor(rep.int(c("A", "B"), c(3, 3))),
  agedays = c(0, 100, 500, 2, 100, 500),
  gestage  = c(rep(35 * 7, 3), rep(35 * 7, 3)),
  sex = rep.int(c("M", "F"), c(3, 3)),
  weight_kg = c(3, 6, 9, 3, 6, 9)
)

# Use the `id` argument to ensure that `classify_wfa()` uses the correct
# standard for each observation
data |>
  classify_wfa(weight_kg = weight_kg,
               age_days = agedays,
               gest_days = gestage,
               sex = sex,
               id = child)

# If you don't specify `id`, `classify_wfa()` will assume data is from one
# child only
data |>
  classify_wfa(weight_kg = weight_kg,
               age_days = agedays,
               gest_days = gestage,
               sex = sex)
}
\references{
\strong{'Implausible z-score values'} \emph{in} World Health Organization (ed.)
\emph{Recommendations for data collection, analysis and reporting on
anthropometric indicators in children under 5 years old}. Geneva: World
Health Organization and the United Nations Children's Fund UNICEF, (2019).
pp. 64-65.

\strong{'Percentage of children stunted, wasted, and underweight, and mean z-scores
for stunting, wasting and underweight'} \emph{in} \emph{Guide to DHS Statistics DHS-7}
Rockville, Maryland, USA: ICF (2020). pp. 431-435.
\url{https://dhsprogram.com/data/Guide-to-DHS-Statistics/Nutritional_Status.htm}
}
\seealso{
See \code{\link[=classify_growth]{classify_growth()}} to run this analysis and others at the same
time. See \code{\link[=gigs_waz]{gigs_waz()}} to learn which growth standard will be used for
different combinations of gestational/chronological age.
}
