% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_compute.R
\name{compute_svn}
\alias{compute_svn}
\title{Get small vulnerable newborn categories using multiple vectors and the
INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} weight-for-gestational age standard}
\usage{
compute_svn(weight_kg, gest_days, sex)
}
\arguments{
\item{weight_kg}{Numeric vector of length one or more with weight value(s) in
kg. It is assumed that weight measurements provided to this function are
birth weights recorded <12 hours after an infant's birth.}

\item{gest_days}{Numeric vector of length one or more with gestational ages
in days. Classifications can only be computed when \code{gest_days} is between
\code{168} and \code{300}, as these are the limits of the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
Newborn Size standards (including the Very Preterm standards). By default,
gigs will warn you about elements of \code{gest_days} which are either outside
these bounds, are \code{NA}, or are \code{Inf}. You can customise this behaviour
using the \link[=gigs_options]{GIGS package-level options}.}

\item{sex}{Character vector of length one or more with sex(es), either \code{"M"}
(male) or \code{"F"} (female). This argument is case-sensitive. By default, gigs
will replace elements of \code{sex} which are not \code{"M"} or \code{"F"} with \code{NA}
and warn you. You can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}
}
\value{
An object of class factor with the same length as the longest input
vector, containing small vulnerable newborn classifications. Its levels
are \code{c("Preterm SGA", "Preterm AGA", "Preterm LGA", "Term SGA", "Term AGA", "Term LGA")}.
}
\description{
Get small vulnerable newborn categories using multiple vectors and the
INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} weight-for-gestational age standard
}
\details{
Cut-offs for small, vulnerable newborn categorisations are:

\tabular{llll}{
\strong{SVN category} \tab \strong{Factor level} \tab
\strong{Gestational age range} \tab \strong{Centile bounds} \cr
Preterm-SGA \tab \code{"Preterm SGA"} \tab \code{p} < 0.1       \tab
\code{gest_days} < 259  \cr
Preterm-AGA \tab \code{"Preterm AGA"} \tab 0.1 ≤ \code{p} ≤ 0.9 \tab
\code{gest_days} < 259  \cr
Preterm-LGA \tab \code{"Preterm LGA"} \tab \code{p} > 0.9       \tab
\code{gest_days} < 259  \cr
Term-SGA \tab \code{"Term SGA"}    \tab \code{p} < 0.1       \tab
\code{gest_days} ≥ 259  \cr
Term-AGA \tab \code{"Term AGA"}    \tab 0.1 ≤ \code{p} ≤ 0.9 \tab
\code{gest_days} ≥ 259  \cr
Term-LGA \tab \code{"Term LGA"}    \tab \code{p} > 0.9       \tab
\code{gest_days} ≥ 259
}

\emph{Abbreviations:} SGA, small-for-gestational age; AGA,
appropriate-for-gestational age; LGA, large-for-gestational age.
}
\note{
Input vectors are recycled by \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}, and must
adhere to the \link{vctrs} recycling rules.
}
\examples{
compute_svn(
  weight_kg = c(1.5, 2.6, 2.6, 3.5),
  gest_days = c(235, 257, 275, 295),
  sex = c("F", "M", "F", "M")
)
}
\references{
Lawn JE, Ohuma EO, Bradley E, Idueta LS, Hazel E, Okwaraji YB et al.
\strong{Small babies, big risks: global estimates of prevalence and mortality for
vulnerable newborns to accelerate change and improve counting.} \emph{The Lancet}
2023, \emph{401(10389):1707-1719.} \doi{10.1016/S0140-6736(23)00522-6}
}
