% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{gigs-package}
\alias{gigs-package}
\alias{gigs}
\title{gigs: Assess Growth in Infants and Newborns}
\description{
Convert between anthropometric measures and z-scores/centiles using WHO
and INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} growth standards. This includes functions for
classification of newborn and infant growth, including size for gestational
age, stunting, wasting, and weight-for-age.
}
\examples{
\dontrun{
help(package = "gigs")
browseURL("https://www.github.com/lshtm-gigs/gigs")
}
}
\keyword{internal}
