% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{gigs-package}
\alias{gigs-package}
\alias{gigs}
\title{gigs: Assess Growth in Infants and Newborns}
\description{
Convert between anthropometric measures and z-scores/percentiles using WHO and INTERGROWTH-21st growth standards.
This package also enables conversion of z-scores/percentiles to newborn and infant size classifications, including
size for gestational age, stunting, and wasting.
}
\examples{
\dontrun{
help(package = "gigs")
browseURL("https://www.github.com/simpar1471/gigs")
}
}
\keyword{internal}
