% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gigs-package.R
\docType{package}
\name{gigs-package}
\alias{gigs-package}
\alias{gigs}
\title{gigs: Assess Fetal, Newborn, and Child Growth with International Standards}
\description{
Convert between anthropometric measures and z-scores/centiles using WHO
and INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} growth standards. This includes functions for
classification of newborn and infant growth, including size for gestational
age, stunting, wasting, and weight-for-age.
}
\examples{
\dontrun{
help(package = "gigs")
browseURL("https://www.github.com/lshtm-gigs/gigs")
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lshtm-gigs/gigs/}
  \item \url{https://lshtm-gigs.github.io/gigs/}
  \item Report bugs at \url{https://github.com/lshtm-gigs/gigs/issues}
}

}
\author{
\strong{Maintainer}: Simon R Parker \email{simon.parker1471@outlook.com} (\href{https://orcid.org/0009-0003-8214-4496}{ORCID})

Other contributors:
\itemize{
  \item Linda Vesel (\href{https://orcid.org/0000-0003-3753-4172}{ORCID}) [data contributor, contributor]
  \item Eric O Ohuma (\href{https://orcid.org/0000-0002-3116-2593}{ORCID}) [data contributor, contributor]
  \item Chitra Saraswati (\href{https://orcid.org/0000-0002-8159-0414}{ORCID}) [reviewer]
  \item Victor Ordu (\href{https://orcid.org/0000-0003-3716-0668}{ORCID}) [reviewer]
  \item Bill & Melinda Gates Foundation [funder]
  \item London School of Hygiene & Tropical Medicine [copyright holder]
}

}
\keyword{internal}
