% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_fet_estimation.R
\name{ig_fet_estimate_fetal_weight}
\alias{ig_fet_estimate_fetal_weight}
\title{Estimate fetal weight in grams using the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
predictive equation}
\usage{
ig_fet_estimate_fetal_weight(abdocirc_mm, headcirc_mm)
}
\arguments{
\item{abdocirc_mm}{Numeric vector with abdominal circumference value(s) in
cm. Should have length one or same length as \code{headcirc_mm}.}

\item{headcirc_mm}{Numeric vector with head circumference value(s) in cm.
Should have length one or same length as \code{abdocirc_mm}.}
}
\value{
Numeric vector with estimated fetal weight(s) in g, with the same
length as the longest input vector.
}
\description{
Estimate fetal weight in grams using the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
predictive equation
}
\note{
Inputs are recycled using \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}.
}
\examples{
# Estimate fetal weight in grams
ig_fet_estimate_fetal_weight(abdocirc_mm = 31:33,
                             headcirc_mm = 25:27)

# Input vectors are recycled using vctrs::vec_recycle_common
ig_fet_estimate_fetal_weight(abdocirc_mm = 25.0,
                             headcirc_mm = 24:26)
}
\references{
Stirnemann J, Villar J, Salomon LJ, Ohuma EO, Lamber A, Victoria CG et al.
\strong{International Estimated Fetal Weight Standards of the INTERGROWTH-21st
Project.} \emph{Ultrasound Obstet Gynecol} 2016, \strong{49:478-486}
\doi{10.1002/uog.17347}
}
\seealso{
Get z-scores/centiles for estimated fetal weights at known
gestational ages using the \code{\link[=ig_fet_value2zscore]{ig_fet_value2zscore()}} or
\code{\link[=ig_fet_value2centile]{ig_fet_value2centile()}} functions, respectively.
}
