% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_fet.R
\name{ig_fet_zscore2value}
\alias{ig_fet_zscore2value}
\alias{ig_fet_hcfga_zscore2value}
\alias{ig_fet_bpdfga_zscore2value}
\alias{ig_fet_acfga_zscore2value}
\alias{ig_fet_flfga_zscore2value}
\alias{ig_fet_ofdfga_zscore2value}
\alias{ig_fet_efwfga_zscore2value}
\alias{ig_fet_centile2value}
\alias{ig_fet_hcfga_centile2value}
\alias{ig_fet_bpdfga_centile2value}
\alias{ig_fet_acfga_centile2value}
\alias{ig_fet_flfga_centile2value}
\alias{ig_fet_ofdfga_centile2value}
\alias{ig_fet_efwfga_centile2value}
\title{Convert z-scores/centiles to values in the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Fetal
Growth standards}
\usage{
ig_fet_zscore2value(z, gest_days, acronym)

ig_fet_hcfga_zscore2value(z, gest_days)

ig_fet_bpdfga_zscore2value(z, gest_days)

ig_fet_acfga_zscore2value(z, gest_days)

ig_fet_flfga_zscore2value(z, gest_days)

ig_fet_ofdfga_zscore2value(z, gest_days)

ig_fet_efwfga_zscore2value(z, gest_days)

ig_fet_centile2value(p, gest_days, acronym)

ig_fet_hcfga_centile2value(p, gest_days)

ig_fet_bpdfga_centile2value(p, gest_days)

ig_fet_acfga_centile2value(p, gest_days)

ig_fet_flfga_centile2value(p, gest_days)

ig_fet_ofdfga_centile2value(p, gest_days)

ig_fet_efwfga_centile2value(p, gest_days)
}
\arguments{
\item{z, p}{Numeric vector of centiles/z-scores to convert to values.}

\item{gest_days}{Numeric vector of gestational age(s) in days. Elements
should be between \code{98} to \code{280} for all standards other than estimated
fetal weight-for-GA (\code{"efwfga"}), which takes values betweem \code{154} and
\code{280}. Will return NA for each observation where GA is not within bounds.}

\item{acronym}{Acronym(s) denoting an INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Fetal
Growth standard. Each value should be one of \code{"hcfga"} (head
circumference-for-GA), \code{"bpdfga"} (biparietal diameter-for-age), \code{"acfga"}
(abdominal circumference-for-GA), \code{"flfga"} (femur length-for-GA),
\code{"ofdfga"} (occipito-frontal diameter-for-GA), or \code{"efwfga"} (estimated
fetal length-for-GA). If none of these, the function will return NA.}
}
\description{
Convert z-scores/centiles to values in the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Fetal
Growth standards
}
\examples{
# Convert centiles to values
p <- 0.25 # 25th centile
ig_fet_centile2value(p = p, gest_days = 280, acronym = "hcfga") |>
  round(digits = 2)

# Or z-scores to values
z <- qnorm(p)
ig_fet_zscore2value(z = z, gest_days = 280, acronym = "hcfga") |>
  round(digits = 2)

# Specify which standard to use with the acronym parameter...
ig_fet_zscore2value(z = z, gest_days = 280, acronym = "acfga") |>
  round(digits = 2)

# ... or by using a standard-specific function
ig_fet_acfga_zscore2value(z = z, gest_days = 280) |>
  round(digits = 2)

# Inputs are recycled to the input of the longest length
ig_fet_efwfga_zscore2value(z = seq(0.1, 0.9, by = 0.2),
                           gest_days = 280) |>
  round(digits = 2)

# Bad inputs will not stop the function but will instead return NA - here 91
# days for gest_days is outside the bounds of the INTERGROWTH-21st Fetal
# Growth standards
ig_fet_hcfga_zscore2value(z = z, gest_days = c(91, 98, 224, 266)) |>
  round(digits = 2)
}
\references{
Papageorghiou AT, Ohuma EO, Altman DG, Todros T, Cheikh Ismail L, Lambert A
et al. \strong{International standards for fetal growth based on serial ultrasound
measurements: the Fetal Growth Longitudinal Study of the INTERGROWTH-21st
Project.} \emph{Lancet} 2014, \strong{384(9946):869-79.}
\doi{0.1016/S0140-6736(14)61490-2}

Stirnemann J, Villar J, Salomon LJ, Ohuma EO, Lamber A, Victoria CG et al.
\strong{International Estimated Fetal Weight Standards of the INTERGROWTH-21st
Project.} \emph{Ultrasound Obstet Gynecol} 2016, \strong{49:478-486}
\doi{10.1002/uog.17347}
}
