% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_nbs.R
\name{ig_nbs_value2percentile}
\alias{ig_nbs_value2percentile}
\alias{ig_nbs_wfga_value2percentile}
\alias{ig_nbs_lfga_value2percentile}
\alias{ig_nbs_hcfga_value2percentile}
\alias{ig_nbs_wlrfga_value2percentile}
\alias{ig_nbs_fmfga_value2percentile}
\alias{ig_nbs_bfpfga_value2percentile}
\alias{ig_nbs_ffmfga_value2percentile}
\alias{ig_nbs_value2zscore}
\alias{ig_nbs_wfga_value2zscore}
\alias{ig_nbs_lfga_value2zscore}
\alias{ig_nbs_hcfga_value2zscore}
\alias{ig_nbs_wlrfga_value2zscore}
\alias{ig_nbs_fmfga_value2zscore}
\alias{ig_nbs_bfpfga_value2zscore}
\alias{ig_nbs_ffmfga_value2zscore}
\title{Convert values to z-scores/percentiles in the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
Newborn Size Standards}
\usage{
ig_nbs_value2percentile(y, gest_days, sex, acronym)

ig_nbs_wfga_value2percentile(weight_kg, gest_days, sex)

ig_nbs_lfga_value2percentile(length_cm, gest_days, sex)

ig_nbs_hcfga_value2percentile(headcirc_cm, gest_days, sex)

ig_nbs_wlrfga_value2percentile(wei_len_ratio, gest_days, sex)

ig_nbs_fmfga_value2percentile(fat_mass_g, gest_days, sex)

ig_nbs_bfpfga_value2percentile(body_fat_perc, gest_days, sex)

ig_nbs_ffmfga_value2percentile(fatfree_mass_g, gest_days, sex)

ig_nbs_value2zscore(y, gest_days, sex, acronym)

ig_nbs_wfga_value2zscore(weight_kg, gest_days, sex)

ig_nbs_lfga_value2zscore(length_cm, gest_days, sex)

ig_nbs_hcfga_value2zscore(headcirc_cm, gest_days, sex)

ig_nbs_wlrfga_value2zscore(wei_len_ratio, gest_days, sex)

ig_nbs_fmfga_value2zscore(fat_mass_g, gest_days, sex)

ig_nbs_bfpfga_value2zscore(body_fat_perc, gest_days, sex)

ig_nbs_ffmfga_value2zscore(fatfree_mass_g, gest_days, sex)
}
\arguments{
\item{y}{Numeric vector of measurement(s) to convert to percentiles/z-scores.
Units depend on which \code{acronym}(s) are in use.}

\item{gest_days}{Numeric vector of gestational age(s) in days. Elements should
be between \code{266} to \code{294} for body composition equations (\code{"fmfga"},
\code{"bfpfga"}, or \code{"ffmfga"}), or between \code{168} and \code{300} for the other
standards. If not inside these bounds, will return NA.}

\item{sex}{Character vector of sex(es), either \code{"M"} (male) or \code{"F"}
(female).}

\item{acronym}{Acronym(s) denoting the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} NBS
standard to use. Must be one of \code{"wfga"} (weight-for-GA), \code{"lfga"}
(length-for-GA), \code{"hcfga"} (head circumference-for-GA), \code{"wlrfga"}
(weight/length ratio-for-GA), \code{"fmfga"} (fat mass-for-GA), \code{"bfpfga"}
(body fat \%-for-GA), or \code{"ffmfga"} (fat-free mass-for-GA).}

\item{weight_kg}{Birth weight(s) in kg.}

\item{length_cm}{Birth length(s) in cm.}

\item{headcirc_cm}{Birth head circumference(s) in cm.}

\item{wei_len_ratio}{Weight-length ratio(s) in kg per m.}

\item{fat_mass_g}{Fat mass(es) in g.}

\item{body_fat_perc}{Body fat percentage(s).}

\item{fatfree_mass_g}{Fat-free mass(es) in g.}
}
\value{
Numeric vector of z-scores/percentiles with length equal to the
outputs from \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}.
}
\description{
Convert values to z-scores/percentiles in the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
Newborn Size Standards
}
\note{
Input vectors are recycled by \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}.
}
\examples{
# Convert values to percentiles
ig_nbs_value2percentile(y = 3.12, gest_days = 280, sex = "M", acronym = "wfga") |>
  round(digits = 2)

# Or values to z-scores
ig_nbs_value2zscore(y = 3.12, gest_days = 280, sex = "M", acronym = "wfga") |>
  round(digits = 2)

# Specify which standard to use with the acronym parameter...
ig_nbs_value2zscore(y = 48.84, gest_days = 280, sex = "F", acronym = "lfga") |>
  round(digits = 2)

# ... or by using a standard-specific function
ig_nbs_lfga_value2zscore(length_cm = 48.84, gest_days = 280, sex = "F") |>
  round(digits = 2)

# Inputs are recycled to the input of the longest length
ig_nbs_wlrfga_value2zscore(wei_len_ratio = c(7.37, 6.47, 6.12, 6.86),
                           gest_days = 280,
                           sex = "M") |>
  round(digits = 2)

# Bad inputs will not stop the function but will instead return NA - here 301
# days gest_days is outside the bounds of the INTERGROWTH-21st newborn size
# standards
ig_nbs_hcfga_value2percentile(headcirc_cm = c(23.0, 28.0, 33.0, 35.0),
                              gest_days = c(168, 217, 266, 301),
                              sex = "F") |>
  round(digits = 2)
}
\references{
Villar J, Cheikh Ismail L, Victora CG, Ohuma EO, Bertino E, Altman DG, et al.
\strong{International standards for newborn weight, length, and head circumference
by gestational age and sex: the Newborn Cross-Sectional Study of the
INTERGROWTH-21st Project.} \emph{Lancet} 2014, \strong{384(9946):857-68.}
\doi{10.1016/S0140-6736(14)60932-6}

Villar J, Giuliani F, Fenton TR, Ohuma EO, Ismail LC, Kennedy SH et al.
\strong{INTERGROWTH-21st very preterm size at birth reference charts.} \emph{Lancet}
2016, \strong{387(10021):844-45.} \doi{10.1016/S0140-6736(16)00384-6}

Villar J, Puglia FA, Fenton TR, Ismal LC, Staines-Urias E, Giuliani F, et al.
\strong{Body composition at birth and its relationship with neonatal anthropometric
ratios: the newborn body composition study of the INTERGROWTH-21st
project.} \emph{Pediatric Research} 2017, \strong{82:305-316.}
\doi{10.1038/pr.2017.52}
}
