% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ig_png}
\alias{ig_png}
\title{INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Postnatal Growth Standards growth curve data}
\source{
\href{https://intergrowth21.tghn.org/postnatal-growth-preterm-infants/}{INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Postnatal Growth of Preterm Infants}
}
\description{
A set of nested lists containing tables with reference values at different
z-scores/centiles for valid post-menstrual ages. The list is ordered by
acronym first, then by sex, and finally by z-score/centile.
}
\examples{
names(gigs::ig_png)
head(gigs::ig_png$wfa$male$zscores)
}
\references{
Villar J, Giuliani F, Bhutta ZA, Bertino E, Ohuma EO, Ismail LC et al.
\strong{Postnatal growth standards for preterm infants: the Preterm Postnatal
Follow-up Study of the INTERGROWTH-21st Project.} \emph{Lancet Glob Health} 2015,
\emph{3(11):e681-e691.} \doi{10.1016/S2214-109X(15)00163-1}
}
\keyword{data}
