% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_png.R
\name{ig_png_value2zscore}
\alias{ig_png_value2zscore}
\alias{ig_png_wfa_value2zscore}
\alias{ig_png_lfa_value2zscore}
\alias{ig_png_hcfa_value2zscore}
\alias{ig_png_wfl_value2zscore}
\alias{ig_png_value2percentile}
\alias{ig_png_wfa_value2percentile}
\alias{ig_png_lfa_value2percentile}
\alias{ig_png_hcfa_value2percentile}
\alias{ig_png_wfl_value2percentile}
\title{Convert postnatal growth measures to INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} postnatal
growth standards}
\usage{
ig_png_value2zscore(y, x, sex, acronym)

ig_png_wfa_value2zscore(weight_kg, pma_weeks, sex)

ig_png_lfa_value2zscore(length_cm, pma_weeks, sex)

ig_png_hcfa_value2zscore(headcirc_cm, pma_weeks, sex)

ig_png_wfl_value2zscore(weight_kg, length_cm, sex)

ig_png_value2percentile(y, x, sex, acronym)

ig_png_wfa_value2percentile(weight_kg, pma_weeks, sex)

ig_png_lfa_value2percentile(length_cm, pma_weeks, sex)

ig_png_hcfa_value2percentile(headcirc_cm, pma_weeks, sex)

ig_png_wfl_value2percentile(weight_kg, length_cm, sex)
}
\arguments{
\item{y}{Value of the anthropometric measure.}

\item{x, pma_weeks}{The \code{x} value at which to convert a
value to a z-score/percentile. Must be within bounds of available \code{x} values
for given acronym. The standard-specific versions of each function specify a
specific type of \code{x}, either \code{pma_weeks} (post-menstrual age in exact weeks;
between \code{27} and \code{64} weeks), or \code{length_cm} (recumbent length measurement(s)
in cm; between \code{35} and \code{65} cm).}

\item{sex}{Sex(es), either \code{"M"} (male) or \code{"F"} (female).}

\item{acronym}{Acronym(s) denoting an INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} standard
for postnatal growth in preterm infants. Should be one of \code{"wfa"}
(weight-for-age), \code{"lfa"} (length-for-age), \code{"hcfa"}
(head circumference-for-age), or \code{"wfl"} (weight-for-length).}

\item{weight_kg}{Weight measurement(s) in kg.}

\item{length_cm}{Recumbent length measurement(s) in cm. Can be either an \code{x}
or \code{y} variable depending on the standard-specific function used.}

\item{headcirc_cm}{Head circumference measurement(s) in cm.}
}
\value{
Z-scores/percentiles for each combination of measurement,
post-menstrual age, sex, and acronym provided to the function.
}
\description{
Convert postnatal growth measures to INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} postnatal
growth standards
}
\examples{
# Convert values to percentiles
ig_png_value2percentile(y = 5.94, x = 55, sex = "M", acronym = "wfa") |>
  round(digits = 2)

# Or values to z-scores
ig_png_value2zscore(y = 5.94, x = 55, sex = "M", acronym = "wfa") |>
  round(digits = 2)

# Specify which standard to use with the acronym parameter...
ig_png_value2zscore(y = 65.1, x = 55, sex = "F", acronym = "lfa") |>
  round(digits = 2)

# ... or by using a standard-specific function
ig_png_lfa_value2zscore(length_cm = 65.1, pma_weeks = 55, sex = "F") |>
  round(digits = 2)

# Inputs are recycled to the input of the longest length
ig_png_lfa_value2zscore(length_cm = c(51.10, 51.48, 51.86, 52.24, 52.63),
                        pma_weeks = 40,
                        sex = "M") |>
  round(digits = 2)

# Bad inputs will not stop the function but will instead return NA - here 25
# weeks post-menstrual age is outside the bounds of the INTERGROWTH-21st
# postnatal growth standards for preterm infants
ig_png_hcfa_value2percentile(headcirc_cm = c(20.6, 22.5, 38.2, 42.8),
                             pma_weeks = c(25, 27, 46, 64),
                             sex = "M") |>
  round(digits = 2)
}
\references{
Villar J, Giuliani F, Bhutta ZA, Bertino E, Ohuma EO, Ismail LC et al.
\strong{Postnatal growth standards for preterm infants: the Preterm Postnatal
Follow-up Study of the INTERGROWTH-21st Project.} \emph{Lancet Glob Health} 2015,
\emph{3(11):e681-e691.} \doi{10.1016/S2214-109X(15)00163-1}
}
