% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_png.R
\name{ig_png_zscore2value}
\alias{ig_png_zscore2value}
\alias{ig_png_wfa_zscore2value}
\alias{ig_png_lfa_zscore2value}
\alias{ig_png_hcfa_zscore2value}
\alias{ig_png_wfl_zscore2value}
\alias{ig_png_percentile2value}
\alias{ig_png_wfa_percentile2value}
\alias{ig_png_lfa_percentile2value}
\alias{ig_png_hcfa_percentile2value}
\alias{ig_png_wfl_percentile2value}
\title{Convert z-scores/percentiles to values in the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
Postnatal Growth Standards for preterm infants}
\usage{
ig_png_zscore2value(z, x, sex, acronym)

ig_png_wfa_zscore2value(z, pma_weeks, sex)

ig_png_lfa_zscore2value(z, pma_weeks, sex)

ig_png_hcfa_zscore2value(z, pma_weeks, sex)

ig_png_wfl_zscore2value(z, length_cm, sex)

ig_png_percentile2value(p, x, sex, acronym)

ig_png_wfa_percentile2value(p, pma_weeks, sex)

ig_png_lfa_percentile2value(p, pma_weeks, sex)

ig_png_hcfa_percentile2value(p, pma_weeks, sex)

ig_png_wfl_percentile2value(p, length_cm, sex)
}
\arguments{
\item{z, p}{Z-score(s)/percentiles to convert to a value/values.}

\item{x, pma_weeks, length_cm}{The \code{x} value at which to convert a
value to a z-score/percentile. Must be within bounds of available \code{x} values
for given acronym. The standard-specific versions of each function specify
\code{x}, either as \code{pma_weeks} (post-menstrual age in exact weeks; between \code{27}
and \code{64} weeks), or \code{length_cm} (recumbent length measurement(s) in cm;
between \code{35} and \code{65} cm).}

\item{sex}{Sex(es), either \code{"M"} (male) or \code{"F"} (female).}

\item{acronym}{Acronym(s) denoting an INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} standard
for postnatal growth in preterm infants. Should be one of \code{"wfa"}
(weight-for-age), \code{"lfa"} (length-for-age), \code{"hcfa"}
(head circumference-for-age), or \code{"wfl"} (weight-for-length).}
}
\value{
Expected measurements for each combination of z-score/percentile,
x variable, sex, and acronym provided to the function.
}
\description{
Convert z-scores/percentiles to values in the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
Postnatal Growth Standards for preterm infants
}
\note{
Input vectors will be recycled to the length of the longest vector
according to the rules of \code{vctrs::vec_recycle_common()}.
}
\examples{
# Convert percentiles to values
p <- 0.25 # 25th percentile
ig_png_percentile2value(p = p, x = 55, sex = "M", acronym = "wfa") |>
  round(digits = 2)

# Or z-scores to values
z <- qnorm(p)
ig_png_zscore2value(z = z, x = 55, sex = "M", acronym = "wfa") |>
  round(digits = 2)

# Specify which standard to use with the acronym parameter...
ig_png_zscore2value(z = z, x = 55, sex = "M", acronym = "lfa") |>
  round(digits = 2)

# ... or by using a standard-specific function
ig_png_lfa_zscore2value(z = z, pma_weeks = 55, sex = "M") |>
  round(digits = 2)

# Inputs are recycled to the input of the longest length
ig_png_lfa_zscore2value(z = seq(0.1, 0.9, by = 0.2),
                        pma_weeks = 40,
                        sex = "M") |>
  round(digits = 2)

# Bad inputs will not stop the function but will instead return NA - here 25
# weeks post-menstrual age is outside the bounds of the INTERGROWTH-21st
# postnatal growth standards for preterm infants
ig_png_hcfa_zscore2value(z = z,
                         pma_weeks = c(25, 27, 46, 64),
                         sex = "M") |>
  round(digits = 2)
}
\references{
Villar J, Giuliani F, Bhutta ZA, Bertino E, Ohuma EO, Ismail LC et al.
\strong{Postnatal growth standards for preterm infants: the Preterm Postnatal
Follow-up Study of the INTERGROWTH-21st Project.} \emph{Lancet Glob Health} 2015,
\emph{3(11):e681-e691.} \doi{10.1016/S2214-109X(15)00163-1}
}
