% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_png.R
\name{ig_png_zscore2value}
\alias{ig_png_zscore2value}
\alias{ig_png_wfa_zscore2value}
\alias{ig_png_lfa_zscore2value}
\alias{ig_png_hcfa_zscore2value}
\alias{ig_png_wfl_zscore2value}
\alias{ig_png_centile2value}
\alias{ig_png_wfa_centile2value}
\alias{ig_png_lfa_centile2value}
\alias{ig_png_hcfa_centile2value}
\alias{ig_png_wfl_centile2value}
\title{Convert z-scores/centiles to values in the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
Postnatal Growth Standards for preterm infants}
\usage{
ig_png_zscore2value(z, x, sex, acronym)

ig_png_wfa_zscore2value(z, pma_weeks, sex)

ig_png_lfa_zscore2value(z, pma_weeks, sex)

ig_png_hcfa_zscore2value(z, pma_weeks, sex)

ig_png_wfl_zscore2value(z, length_cm, sex)

ig_png_centile2value(p, x, sex, acronym)

ig_png_wfa_centile2value(p, pma_weeks, sex)

ig_png_lfa_centile2value(p, pma_weeks, sex)

ig_png_hcfa_centile2value(p, pma_weeks, sex)

ig_png_wfl_centile2value(p, length_cm, sex)
}
\arguments{
\item{z, p}{Numeric vector of length one or more with centiles/z-scores to
convert to values. For \code{p}, if an element of \code{p} is not between \code{0} and
\code{1}, gigs will replace it with \code{NA} and warn you. This behaviour can be
customised using the functions in \link{gigs_options}.}

\item{x}{Numeric vector of length one or more with x values. Elements of \code{x}
or its standard-specific equivalents (\code{pma_weeks}, \code{length_cm}) should have
specific units and be between certain values depending on the standard in
use (defined by \code{acronym}). These are:
\itemize{
\item Between 27 and 64 weeks for \code{"wfa"}, \code{"lfa"}, and \code{"hcfa"}.
\item Between 35 and 65 days for \code{"wfl"}.
}

By default, gigs will replace out-of-bounds elements in \code{x} with \code{NA} and
warn you. This behaviour can be customised using the functions in
\link{gigs_options}.}

\item{sex}{Character vector of length one or more with sex(es), either \code{"M"}
(male) or \code{"F"} (female). This argument is case-sensitive. By default, gigs
will replace elements of \code{sex} which are not \code{"M"} or \code{"F"} with \code{NA}
and warn you. This behaviour can be customised using the functions in
\link{gigs_options}.}

\item{acronym}{Character vector of length one or more denoting the
INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Postnatal Growth standard(s) in use. Each
element should be one of:
\itemize{
\item \code{"wfa"} (weight-for-age)
\item \code{"lfa"} (length-for-age)
\item \code{"hcfa"} (head circumference-for-age)
\item \code{"wfl"} (weight-for-length)
}

This argument is case-sensitive. By default, gigs will replace elements in
\code{acronym} which are not one of the above values with \code{NA} and warn you.
This behaviour can be customised using the functions in \link{gigs_options}.}
}
\value{
Numeric vector of z-scores/centiles with length equal to the
outputs from \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}.
}
\description{
Convert z-scores/centiles to values in the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
Postnatal Growth Standards for preterm infants
}
\note{
Input vectors are recycled by \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}.
}
\examples{
# Convert centiles to values
p <- 0.25 # 25th centile
ig_png_centile2value(p = p, x = 55, sex = "M", acronym = "wfa") |>
  round(digits = 2)

# Or z-scores to values
z <- qnorm(p)
ig_png_zscore2value(z = z, x = 55, sex = "M", acronym = "wfa") |>
  round(digits = 2)

# Specify which standard to use with the acronym parameter...
ig_png_zscore2value(z = z, x = 55, sex = "M", acronym = "lfa") |>
  round(digits = 2)

# ... or by using a standard-specific function
ig_png_lfa_zscore2value(z = z, pma_weeks = 55, sex = "M") |>
  round(digits = 2)

# Inputs are recycled to the input of the longest length
ig_png_lfa_zscore2value(z = seq(0.1, 0.9, by = 0.2),
                        pma_weeks = 40,
                        sex = "M") |>
  round(digits = 2)

# Bad inputs will not stop the function but will instead return NA - here 25
# weeks post-menstrual age is outside the bounds of the INTERGROWTH-21st
# postnatal growth standards for preterm infants
ig_png_hcfa_zscore2value(z = z,
                         pma_weeks = c(25, 27, 46, 64),
                         sex = "M") |>
  round(digits = 2)
}
\references{
Villar J, Giuliani F, Bhutta ZA, Bertino E, Ohuma EO, Ismail LC et al.
\strong{Postnatal growth standards for preterm infants: the Preterm Postnatal
Follow-up Study of the INTERGROWTH-21st Project.} \emph{Lancet Glob Health} 2015,
\emph{3(11):e681-e691.} \doi{10.1016/S2214-109X(15)00163-1}
}
