% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_png.R
\name{ig_png_zscore2value}
\alias{ig_png_zscore2value}
\alias{ig_png_wfa_zscore2value}
\alias{ig_png_lfa_zscore2value}
\alias{ig_png_hcfa_zscore2value}
\alias{ig_png_percentile2value}
\alias{ig_png_wfa_percentile2value}
\alias{ig_png_lfa_percentile2value}
\alias{ig_png_hcfa_percentile2value}
\title{Convert z-score/percentiles to values in the INTERGROWTH-21st Postnatal growth standards
for preterm infants measurements}
\usage{
ig_png_zscore2value(z, pma_weeks, sex, acronym)

ig_png_wfa_zscore2value(z, pma_weeks, sex)

ig_png_lfa_zscore2value(z, pma_weeks, sex)

ig_png_hcfa_zscore2value(z, pma_weeks, sex)

ig_png_percentile2value(p, pma_weeks, sex, acronym)

ig_png_wfa_percentile2value(p, pma_weeks, sex)

ig_png_lfa_percentile2value(p, pma_weeks, sex)

ig_png_hcfa_percentile2value(p, pma_weeks, sex)
}
\arguments{
\item{z, p}{Z-score(s)/percentile(s) to convert to a value.}

\item{pma_weeks}{Post-menstrual age(s) in exact weeks. Must be between \code{27}
and \code{64}.}

\item{sex}{Sex(es), either \code{"M"} (male) or \code{"F"} (female).}

\item{acronym}{Acronym(s) denoting an INTERGROWTH-21st standard for
post-natal growth in preterm infants. Should be one of \code{"wfa"}, \code{"lhfa"},
\code{"hcfa"}.}
}
\value{
Expected measurements for each combination of z-score/centile,
post-menstrual age, sex, and acronym provided to the function.
}
\description{
Convert z-score/percentiles to values in the INTERGROWTH-21st Postnatal growth standards
for preterm infants measurements
}
\examples{
# Convert percentiles to values
p <- 0.25 # 25th percentile
ig_png_percentile2value(p = p, pma_weeks = 55, sex = "M", acronym = "wfa") |>
  round(digits = 2)

# Or z-scores to values
z <- qnorm(p)
ig_png_zscore2value(z = z, pma_weeks = 55, sex = "M", acronym = "wfa") |>
  round(digits = 2)

# Specify which standard to use with the acronym parameter...
ig_png_zscore2value(z = z, pma_weeks = 55, sex = "M", acronym = "lfa") |>
  round(digits = 2)

# ... or by using a standard-specific function
ig_png_lfa_zscore2value(z = z, pma_weeks = 55, sex = "M") |>
  round(digits = 2)

# Inputs are recycled to the input of the longest length
ig_png_lfa_zscore2value(z = seq(0.1, 0.9, by = 0.2),
                        pma_weeks = 40,
                        sex = "M") |>
  round(digits = 2)

# Bad inputs will not stop the function but will instead return NA - here 25
# weeks post-menstrual age is outside the bounds of the INTERGROWTH-21st
# postnatal growth standards for preterm infants
ig_png_hcfa_zscore2value(z = z,
                         pma_weeks = c(25, 27, 46, 64),
                         sex = "M") |>
  round(digits = 2)
}
\references{
Villar J, Giuliani F, Bhutta ZA, Bertino E, Ohuma EO, Ismail LC et al.
\strong{Postnatal growth standards for preterm infants: the Preterm Postnatal
Follow-up Study of the INTERGROWTH-21st Project.} \emph{Lancet Glob Health} 2015,
\emph{3(11):e681-e691.} \doi{10.1016/S2214-109X(15)00163-1}
}
