% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life6mo}
\alias{life6mo}
\title{Data extract from the Low birthweight Infant Feeding Exploration (LIFE) study}
\format{
A data frame with 2,191 rows and 10 variables:
\describe{
\item{\code{id}:}{  Unique ID for each infant in the dataset (1--300).}
\item{\code{visitweek}:}{  Chronological age in weeks (±1) when study visit
occurred (0--26).}
\item{\code{sex}:}{  Sex of the infant as a factor (\code{"M"} = Male; \code{"F"} =
Female).}
\item{\code{gestage}:}{  Best obstetric estimate of gestational age in days
(181--291).}
\item{\code{age_days}:}{  Chronological age in days at each visit; equal to
\code{pma - gestage} (0--242).}
\item{\code{pma}:}{  Post-menstrual age in days (182--528).}
\item{\code{wt_kg}:}{  Mean weight in kg (1.24--9.40667).}
\item{\code{len_cm}:}{  Mean length in cm (37.37--72.93).}
\item{\code{headcirc_cm}:}{  Mean head circumference in cm (23.20--44.87).}
\item{\code{muac_cm}:}{  Mean mid-upper arm circumference in cm (6.30--16.83).}
}
}
\description{
A subset of anthropometric data for 300 singleton infants
enrolled in the Low birthweight Infant Feeding Exploration (LIFE) study. The
variables are as follows:
}
\note{
We subsetted the full LIFE 6 month dataset for \link{gigs}. As such, this
extract only includes data from 300 singleton pregnancies where the best
estimate of gestational age was >168 days. We also removed rows corresponding
to visit weeks where no measurement data was taken due to non-attendance of
the visit.
}
\examples{
head(gigs::life6mo)
}
\references{
Vesel L, Bellad RM, Manji K, Saidi F, Velasquez E, Sudfeld C, et al.
\strong{Feeding practices and growth patterns of moderately low birthweight infants
in resource-limited settings: results from a multisite, longitudinal
observational study.} \emph{BMJ Open} 2023, \strong{13(2):e067316.}
\doi{10.1136/BMJOPEN-2022-067316}
}
\keyword{data}
