% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{shared_roxygen_params}
\alias{shared_roxygen_params}
\title{Shared roxygen parameters}
\arguments{
\item{y}{Numeric vector of measurement(s) to convert to centiles/z-scores.
Units depend on which \code{acronym}(s) are in use.}

\item{z, p}{Numeric vector of centiles/z-scores to convert to values.}

\item{sex}{Character vector of sex(es), either \code{"M"} (male) or \code{"F"}
(female).}
}
\description{
Many of the parameters in \verb{[gigs]} are shared between different functions.
This roxygen block is designed to make inheriting these parameters as simple
as possible.
}
\note{
Input vectors are recycled by \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}.
}
\keyword{internal}
