% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{shared_zscore2value_returns}
\alias{shared_zscore2value_returns}
\title{Shared zscore/centile roxygen parameters}
\value{
Numeric vector of z-scores/centiles with length equal to the
outputs from \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}.
}
\description{
Much of the roxygen content in \verb{[gigs]} is shared between different
functions. This roxygen block is designed to make inheriting these parameters
more simple.
}
\keyword{internal}
