% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_gs.R
\name{who_gs_lms}
\alias{who_gs_lms}
\title{Retrieve LMS values for WHO Child Growth Standards}
\usage{
who_gs_lms(x, sex, acronym)
}
\arguments{
\item{x}{X value(s) at which to retrieve LMS values. Must be within bounds of
available x values for given acronym.}

\item{sex}{Sex(es), either \code{"M"} (male) or \code{"F"} (female).}

\item{acronym}{Valid acronym for WHO Growth Standards datasets: one of
\code{"wfa"} (weight-for-age), \code{"bfa"} (bmi-for-age), \code{"lhfa"}
(length/height-for-age), \code{"wfl"} (weight-for-length), or \code{"wfh"}
(weight-for-height), \code{"hcfa"} (head circumference-for-age), \code{"acfa"}
(arm circumference-for-age), \code{"ssfa"} (subscapular skinfold-for-age), or
\code{"tsfa"} (triceps skinfold-for-age).}
}
\value{
A dataframe with lambda, mu and sigma values for the WHO growth
standard(s) specified by the \code{acronym} parameter, for the supplied \code{x} and
\code{sex} values.
}
\description{
Retrieve LMS values for WHO Child Growth Standards
}
\references{
World Health Organisation. \strong{WHO child growth standards:
length/height-for-age, weight-for-age, weight-for-length, weight-for-height
and body mass index-for-age: methods and development.} \emph{Technical report,
WHO, Geneva}, 2006.

World Health Organisation. \strong{WHO child growth standards: head
circumference-for-age, arm circumference-for-age, triceps skinfold-for-age
and subscapular skinfold-for-age: methods and development.} \emph{Technical
report, WHO, Geneva}, 2007.

Cole TJ. \strong{The LMS method for constructing normalized growth standards} \emph{Eur
J Clin Nutr.} 1990, \strong{44(1):45-60.} PMID:
\href{https://pubmed.ncbi.nlm.nih.gov/2354692/}{2354692}
}
\keyword{internal}
