% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_gs.R
\name{who_gs_lms2sd}
\alias{who_gs_lms2sd}
\title{Get standard deviations for WHO Growth Standards data from LMS values}
\usage{
who_gs_lms2sd(l, m, s, n_sd)
}
\arguments{
\item{l}{Lambda value as provided by \link{who_gs_lms}}

\item{m}{Mu value as provided by \link{who_gs_lms}}

\item{s}{Sigma value as provided by \link{who_gs_lms}}

\item{n_sd}{Number of SD from the median to be computed}
}
\description{
Get standard deviations for WHO Growth Standards data from LMS values
}
\references{
World Health Organisation. \strong{WHO child growth standards:
length/height-for-age, weight-for-age, weight-for-length, weight-for-height
and body mass index-for-age: methods and development.} \emph{Technical report,
WHO, Geneva}, 2006.

World Health Organisation. \strong{WHO child growth standards: head
circumference-for-age, arm circumference-for-age, triceps skinfold-for-age
and subscapular skinfold-for-age: methods and development.} \emph{Technical
report, WHO, Geneva}, 2007.

Cole TJ. \strong{The LMS method for constructing normalized growth standards} \emph{Eur
J Clin Nutr.} 1990, \strong{44(1):45-60.} PMID:
\href{https://pubmed.ncbi.nlm.nih.gov/2354692/}{2354692}
}
\keyword{internal}
