% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_gs.R
\name{who_gs_value2zscore}
\alias{who_gs_value2zscore}
\alias{who_gs_wfa_value2zscore}
\alias{who_gs_bfa_value2zscore}
\alias{who_gs_lhfa_value2zscore}
\alias{who_gs_wfl_value2zscore}
\alias{who_gs_wfh_value2zscore}
\alias{who_gs_hcfa_value2zscore}
\alias{who_gs_acfa_value2zscore}
\alias{who_gs_ssfa_value2zscore}
\alias{who_gs_tsfa_value2zscore}
\alias{who_gs_value2centile}
\alias{who_gs_wfa_value2centile}
\alias{who_gs_bfa_value2centile}
\alias{who_gs_lhfa_value2centile}
\alias{who_gs_wfl_value2centile}
\alias{who_gs_wfh_value2centile}
\alias{who_gs_hcfa_value2centile}
\alias{who_gs_acfa_value2centile}
\alias{who_gs_ssfa_value2centile}
\alias{who_gs_tsfa_value2centile}
\title{Convert values to z-scores/centiles in the WHO Child Growth Standards}
\usage{
who_gs_value2zscore(y, x, sex, acronym)

who_gs_wfa_value2zscore(weight_kg, age_days, sex)

who_gs_bfa_value2zscore(bmi, age_days, sex)

who_gs_lhfa_value2zscore(lenht_cm, age_days, sex)

who_gs_wfl_value2zscore(weight_kg, length_cm, sex)

who_gs_wfh_value2zscore(weight_kg, height_cm, sex)

who_gs_hcfa_value2zscore(headcirc_cm, age_days, sex)

who_gs_acfa_value2zscore(armcirc_cm, age_days, sex)

who_gs_ssfa_value2zscore(subscap_sf_mm, age_days, sex)

who_gs_tsfa_value2zscore(triceps_sf_mm, age_days, sex)

who_gs_value2centile(y, x, sex, acronym)

who_gs_wfa_value2centile(weight_kg, age_days, sex)

who_gs_bfa_value2centile(bmi, age_days, sex)

who_gs_lhfa_value2centile(lenht_cm, age_days, sex)

who_gs_wfl_value2centile(weight_kg, length_cm, sex)

who_gs_wfh_value2centile(weight_kg, height_cm, sex)

who_gs_hcfa_value2centile(headcirc_cm, age_days, sex)

who_gs_acfa_value2centile(armcirc_cm, age_days, sex)

who_gs_ssfa_value2centile(subscap_sf_mm, age_days, sex)

who_gs_tsfa_value2centile(triceps_sf_mm, age_days, sex)
}
\arguments{
\item{y}{Numeric vector of length one or more with anthropometric
measurement(s) to convert to centiles/z-scores. Units depend on which
\code{acronym}(s) are in use.}

\item{x, age_days, length_cm, height_cm}{Numeric vector of length one or more
with x values. Elements in \code{x} or its standard-specific equivalents
(\code{age_days}, \code{length_cm}, and \code{height_cm}) should have specific units and
be between certain values depending on the standard in use (defined by
\code{acronym}). These are:
\itemize{
\item Between 0 and 1856 days for \code{"wfa"}, \code{"bfa"}, \code{"lhfa"}, and \code{"hcfa"}.
\item Between 45 and 110 days for \code{"wfl"}.
\item Between 65 and 120 days for \code{"wfh"}.
\item Between 91 and 1856 days for \code{"acfa"}, \code{"ssfa"}, \code{"tsfa"}.
}

By default, gigs will replace out-of-bounds elements in \code{x} with \code{NA} and
warn you. You can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}

\item{sex}{Character vector of length one or more with sex(es), either \code{"M"}
(male) or \code{"F"} (female). This argument is case-sensitive. By default, gigs
will replace elements of \code{sex} which are not \code{"M"} or \code{"F"} with \code{NA}
and warn you. You can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}

\item{acronym}{A single-length character vector denoting the WHO Child Growth
standard in use. Should be one of:
\itemize{
\item \code{"wfa"} (weight-for-age)
\item \code{"bfa"} (BMI-for-age)
\item \code{"lhfa"} (length/height-for-age)
\item \code{"wfl"} (weight-for-length)
\item \code{"wfh"} (weight-for-height)
\item \code{"hcfa"} (head circumference-for-age)
\item \code{"acfa"} (arm circumference-for-age)
\item \code{"ssfa"} (subscapular skinfold-for-age)
\item \code{"tsfa"} (triceps skinfold-for-age)
}

This argument is case-sensitive. If \code{acronym} is not one of the values
listed above the function will throw an error.}

\item{weight_kg}{Numeric vector of length one or more with weight
measurement(s) in kg.}

\item{bmi}{Numeric vector of length one or more with body mass index
measurement(s) in kg/m\if{html}{\out{<sup>}}2\if{html}{\out{</sup>}}.}

\item{lenht_cm}{Numeric vector of length one or more with length/height
measurement(s) in cm.}

\item{headcirc_cm}{Numeric vector of length one or more with head
circumference measurement(s) in cm.}

\item{armcirc_cm}{Numeric vector of length one or more with arm circumference
measurement(s) in cm.}

\item{subscap_sf_mm}{Numeric vector of length one or more with subscapular
skinfold measurement(s) in mm.}

\item{triceps_sf_mm}{Numeric vector of length one or more with triceps
skinfold measurement(s) in mm.}
}
\value{
Numeric vector of z-scores/centiles with length equal to the longest
input vector.
}
\description{
Convert values to z-scores/centiles in the WHO Child Growth Standards
}
\note{
Input vectors other than \code{acronym} are recycled by
\code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}, and must adhere to the \link{vctrs} recycling
rules.
}
\examples{
# Convert values to centiles
who_gs_value2centile(y = c(10.1, 10.2, 10.3),
                     x = 504:506,
                     sex = c("M", "F", "M"),
                     acronym = "wfa") |>
  round(digits = 2)

# Or values to z-scores
who_gs_value2zscore(y = c(10.1, 10.2, 10.3),
                    x = 504:506,
                    sex = c("M", "F", "M"),
                    acronym = "wfa") |>
  round(digits = 2)

# Specify which standard to use with the acronym parameter...
who_gs_value2centile(y = c(73.9, 67.9, 71.4),
                     x = 299:301,
                     sex = c("M", "F", "M"),
                     acronym = "lhfa") |>
  round(digits = 2)

# ... or by using a standard-specific function
who_gs_lhfa_value2centile(lenht_cm = c(73.9, 67.9, 71.4),
                          age_days = 299:301,
                          sex = c("M", "F", "M")) |>
  round(digits = 2)

# Inputs are recycled to the input of the longest length
who_gs_ssfa_value2centile(subscap_sf_mm = 6.1,
                          age_days = seq(450, 650, by = 50),
                          sex = "M") |>
  round(digits = 2)

# Bad inputs will not stop the function but will instead produce `NA`s in the
# output - by default gigs will issue useful warnings
who_gs_wfh_value2zscore(weight_kg = c(7.4, 14.2, 21.8, 22.4),
                        height_cm = c(65, 95, 120, NA),
                        sex = c("M", "X", "F", NA)) |>
    round(digits = 2)
}
\references{
de Onis M, Garza C, Victora CG, Onyango AW, Frongillo EA, Martines J. \strong{The
WHO Multicentre Growth Reference Study: planning, study design, and
methodology} \emph{Food Nutr Bull.} 2004, \strong{25(1 Suppl):S15-26.}
\doi{10.1177/15648265040251s104}

World Health Organisation. \strong{WHO child growth standards:
length/height-for-age, weight-for-age, weight-for-length, weight-for-height
and body mass index-for-age: methods and development.} \emph{Technical report,
WHO, Geneva}, 2006.

World Health Organisation. \strong{WHO child growth standards: head
circumference-for-age, arm circumference-for-age, triceps skinfold-for-age
and subscapular skinfold-for-age: methods and development.} \emph{Technical
report, WHO, Geneva}, 2007.

Cole TJ. \strong{The LMS method for constructing normalized growth standards} \emph{Eur
J Clin Nutr.} 1990, \strong{44(1):45-60.} PMID:
\href{https://pubmed.ncbi.nlm.nih.gov/2354692/}{2354692}
}
