% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_gs.R
\name{who_gs_value2zscore}
\alias{who_gs_value2zscore}
\alias{who_gs_wfa_value2zscore}
\alias{who_gs_bfa_value2zscore}
\alias{who_gs_lhfa_value2zscore}
\alias{who_gs_wfl_value2zscore}
\alias{who_gs_wfh_value2zscore}
\alias{who_gs_hcfa_value2zscore}
\alias{who_gs_acfa_value2zscore}
\alias{who_gs_ssfa_value2zscore}
\alias{who_gs_tsfa_value2zscore}
\alias{who_gs_value2percentile}
\alias{who_gs_wfa_value2percentile}
\alias{who_gs_bfa_value2percentile}
\alias{who_gs_lhfa_value2percentile}
\alias{who_gs_wfl_value2percentile}
\alias{who_gs_wfh_value2percentile}
\alias{who_gs_hcfa_value2percentile}
\alias{who_gs_acfa_value2percentile}
\alias{who_gs_ssfa_value2percentile}
\alias{who_gs_tsfa_value2percentile}
\title{Convert values to z-score/percentiles in the WHO Child Growth Standards}
\usage{
who_gs_value2zscore(y, x, sex, acronym)

who_gs_wfa_value2zscore(weight_kg, age_days, sex)

who_gs_bfa_value2zscore(bmi, age_days, sex)

who_gs_lhfa_value2zscore(lenht_cm, age_days, sex)

who_gs_wfl_value2zscore(weight_kg, length_cm, sex)

who_gs_wfh_value2zscore(weight_kg, height_cm, sex)

who_gs_hcfa_value2zscore(headcirc_cm, age_days, sex)

who_gs_acfa_value2zscore(armcirc_cm, age_days, sex)

who_gs_ssfa_value2zscore(subscap_sf_mm, age_days, sex)

who_gs_tsfa_value2zscore(triceps_sf_mm, age_days, sex)

who_gs_value2percentile(y, x, sex, acronym)

who_gs_wfa_value2percentile(weight_kg, age_days, sex)

who_gs_bfa_value2percentile(bmi, age_days, sex)

who_gs_lhfa_value2percentile(lenht_cm, age_days, sex)

who_gs_wfl_value2percentile(weight_kg, length_cm, sex)

who_gs_wfh_value2percentile(weight_kg, height_cm, sex)

who_gs_hcfa_value2percentile(headcirc_cm, age_days, sex)

who_gs_acfa_value2percentile(armcirc_cm, age_days, sex)

who_gs_ssfa_value2percentile(subscap_sf_mm, age_days, sex)

who_gs_tsfa_value2percentile(triceps_sf_mm, age_days, sex)
}
\arguments{
\item{y}{Y value(s) for growth standard.}

\item{x, age_days, length_cm, height_cm}{The \code{x} value at which to convert a
value to a z-score/percentile. Must be within bounds of available \code{x} values
for given acronym. The standard-specific versions of each function specify a
specific type of \code{x}, either \code{age_days} (age since birth in days),
\code{length_cm} (recumbent length measurement(s) in cm) or \code{height_cm}
(standing height measurement(s) in cm).}

\item{sex}{Sex(es), either \code{"M"} (male) or \code{"F"}  (female).}

\item{acronym}{Valid acronym for WHO Growth Standards datasets: one of
\code{"wfa"} (weight-for-age), \code{"bfa"} (BMI-for-age), \code{"lhfa"}
(length/height-for-age), \code{"wfl"} (weight-for-length), or \code{"wfh"}
(weight-for-height), \code{"hcfa"} (head circumference-for-age), \code{"acfa"}
(arm circumference-for-age), \code{"ssfa"} (subscapular skinfold-for-age), or
\code{"tsfa"} (triceps skinfold-for-age).}

\item{weight_kg}{Weight measurement(s) in kg.}

\item{bmi}{Body mass index measurement(s).}

\item{lenht_cm}{Length/height measurement(s) in cm.}

\item{headcirc_cm}{Head circumference measurement(s) in cm.}

\item{armcirc_cm}{Arm circumference measurement(s) in cm.}

\item{subscap_sf_mm}{Subscapular skinfold measurement(s) in mm.}

\item{triceps_sf_mm}{Triceps skinfold measurement(s) in mm.}
}
\value{
Z-scores/percentiles for each combination of measurement, x
variable, sex, and acronym provided to the function.
}
\description{
Convert values to z-score/percentiles in the WHO Child Growth Standards
}
\examples{
# Convert values to percentiles
who_gs_value2percentile(y = 10.1, x = 505, sex = "M", acronym = "wfa") |>
  round(digits = 2)

# Or values to z-scores
who_gs_value2zscore(y = 10.1, x = 505, sex = "M", acronym = "wfa") |>
  round(digits = 2)

# Specify which standard to use with the acronym parameter...
who_gs_value2percentile(y = 75.2, x = 300, sex = "M", acronym = "lhfa") |>
  round(digits = 2)

# ... or by using a standard-specific function
who_gs_lhfa_value2percentile(lenht_cm = 75.2, age_days = 300, sex = "M") |>
  round(digits = 2)

# Inputs are recycled to the input of the longest length
who_gs_ssfa_value2percentile(subscap_sf_mm = 6.1,
                             age_days = seq(450, 650, by = 50),
                             sex = "M") |>
  round(digits = 2)

# Bad inputs will not stop the function but will instead return NA - here
# 121 cm for height_cm is incompatible with the WHO Growth Standard for
# weight-for-height
who_gs_wfh_value2zscore(weight_kg = c(7.4, 14.2, 21.8, 22.4),
                        height_cm = c(65, 95, 120, 121),
                        sex = "M") |>
  round(digits = 2)
}
\references{
de Onis M, Garza C, Victora CG, Onyango AW, Frongillo EA, Martines J. \strong{The
WHO Multicentre Growth Reference Study: planning, study design, and
methodology} \emph{Food Nutr Bull.} 2004, \strong{25(1 Suppl):S15-26.} doi:
\href{https://journals.sagepub.com/doi/10.1177/15648265040251S104}{10.1177/15648265040251s104}

World Health Organisation. \strong{WHO child growth standards:
length/height-for-age, weight-for-age, weight-for-length, weight-for-height
and body mass index-for-age: methods and development.} \emph{Technical report,
WHO, Geneva}, 2006.

World Health Organisation. \strong{WHO child growth standards: head
circumference-for-age, arm circumference-for-age, triceps skinfold-for-age
and subscapular skinfold-for-age: methods and development.} \emph{Technical
report, WHO, Geneva}, 2007.

Cole TJ. \strong{The LMS method for constructing normalized growth standards} \emph{Eur
J Clin Nutr.} 1990, \strong{44(1):45-60.} PMID:
\href{https://pubmed.ncbi.nlm.nih.gov/2354692/}{2354692}
}
