% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_gs.R
\name{who_gs_zscore2value}
\alias{who_gs_zscore2value}
\alias{who_gs_wfa_zscore2value}
\alias{who_gs_bfa_zscore2value}
\alias{who_gs_lhfa_zscore2value}
\alias{who_gs_wfl_zscore2value}
\alias{who_gs_wfh_zscore2value}
\alias{who_gs_hcfa_zscore2value}
\alias{who_gs_acfa_zscore2value}
\alias{who_gs_ssfa_zscore2value}
\alias{who_gs_tsfa_zscore2value}
\alias{who_gs_centile2value}
\alias{who_gs_wfa_centile2value}
\alias{who_gs_bfa_centile2value}
\alias{who_gs_lhfa_centile2value}
\alias{who_gs_wfl_centile2value}
\alias{who_gs_wfh_centile2value}
\alias{who_gs_hcfa_centile2value}
\alias{who_gs_acfa_centile2value}
\alias{who_gs_ssfa_centile2value}
\alias{who_gs_tsfa_centile2value}
\title{Convert z-scores/centiles to values in the WHO Child Growth Standards}
\usage{
who_gs_zscore2value(z, x, sex, acronym)

who_gs_wfa_zscore2value(z, age_days, sex)

who_gs_bfa_zscore2value(z, age_days, sex)

who_gs_lhfa_zscore2value(z, age_days, sex)

who_gs_wfl_zscore2value(z, length_cm, sex)

who_gs_wfh_zscore2value(z, height_cm, sex)

who_gs_hcfa_zscore2value(z, age_days, sex)

who_gs_acfa_zscore2value(z, age_days, sex)

who_gs_ssfa_zscore2value(z, age_days, sex)

who_gs_tsfa_zscore2value(z, age_days, sex)

who_gs_centile2value(p, x, sex, acronym)

who_gs_wfa_centile2value(p, age_days, sex)

who_gs_bfa_centile2value(p, age_days, sex)

who_gs_lhfa_centile2value(p, age_days, sex)

who_gs_wfl_centile2value(p, length_cm, sex)

who_gs_wfh_centile2value(p, height_cm, sex)

who_gs_hcfa_centile2value(p, age_days, sex)

who_gs_acfa_centile2value(p, age_days, sex)

who_gs_ssfa_centile2value(p, age_days, sex)

who_gs_tsfa_centile2value(p, age_days, sex)
}
\arguments{
\item{z, p}{Numeric vector of length one or more with centiles/z-scores to
convert to values. For \code{p}, if an element of \code{p} is not between \code{0} and
\code{1}, gigs will replace it with \code{NA} and warn you. This behaviour can be
customised using the functions in \link{gigs_options}.}

\item{x, age_days, length_cm, height_cm}{Numeric vector of length one or more
with x values. Elements in \code{x} or its standard-specific equivalents
(\code{age_days}, \code{length_cm}, and \code{height_cm}) should have specific units and
be between certain values depending on the standard in use (defined by
\code{acronym}). These are:
\itemize{
\item Between 0 and 1856 days for \code{"wfa"}, \code{"bfa"}, \code{"lhfa"}, and \code{"hcfa"}.
\item Between 45 and 110 days for \code{"wfl"}.
\item Between 65 and 120 days for \code{"wfh"}.
\item Between 91 and 1856 days for \code{"acfa"}, \code{"ssfa"}, \code{"tsfa"}.
}

By default, gigs will replace out-of-bounds elements in \code{x} with \code{NA} and
warn you. This behaviour can be customised using the functions in
\link{gigs_options}.}

\item{sex}{Character vector of length one or more with sex(es), either \code{"M"}
(male) or \code{"F"} (female). This argument is case-sensitive. By default, gigs
will replace elements of \code{sex} which are not \code{"M"} or \code{"F"} with \code{NA}
and warn you. This behaviour can be customised using the functions in
\link{gigs_options}.}

\item{acronym}{Character vector of length one or more denoting the WHO Child
Growth standard(s) in use. Each element should be one of:
\itemize{
\item \code{"wfa"} (weight-for-age)
\item \code{"bfa"} (BMI-for-age)
\item \code{"lhfa"} (length/height-for-age)
\item \code{"wfl"} (weight-for-length)
\item \code{"wfh"} (weight-for-height)
\item \code{"hcfa"} (head circumference-for-age)
\item \code{"acfa"} (arm circumference-for-age)
\item \code{"ssfa"} (subscapular skinfold-for-age)
\item \code{"tsfa"} (triceps skinfold-for-age)
}

This argument is case-sensitive. By default, gigs will replace elements in
\code{acronym} which are not one of the above values with \code{NA} and warn you.
This behaviour can be customised using the functions in \link{gigs_options}. If
all elements in \code{acronym} are not one of the above values, gigs will throw
an error.}
}
\value{
Numeric vector of expected measurements with length equal to the
longest input vector.
}
\description{
Convert z-scores/centiles to values in the WHO Child Growth Standards
}
\note{
Input vectors are recycled by \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}, and must
adhere to the \link{vctrs} recycling rules.
}
\examples{
# Convert centiles to values
p <- 0.25 # 25th centile
who_gs_centile2value(p = p, x = 501, sex = "M", acronym = "wfa") |>
  round(digits = 2)

# Or z-scores to values
z <- qnorm(p) # Z-score equivalent of 25th centile
who_gs_zscore2value(z = z, x = 501, sex = "M", acronym = "wfa") |>
  round(digits = 2)

# Specify which standard to use with the acronym parameter...
who_gs_zscore2value(z = z, x = 300, sex = "M", acronym = "lhfa") |>
  round(digits = 2)

# ... or by using a standard-specific function
who_gs_lhfa_zscore2value(z = z, age_days = 300, sex = "M") |>
  round(digits = 2)

# Inputs are recycled to the input of the longest length
who_gs_ssfa_centile2value(p = seq(0.1, 0.9, by = 0.2),
                          age_days = 670,
                          sex = "M") |>
  round(digits = 2)

# Bad inputs will not stop the function but will instead return NA - here 60
# cm for height_cm is incompatible with the WHO Growth Standard for
# weight-for-height
who_gs_wfh_zscore2value(z = 0,
                        height_cm = c(60, 85, 105, 120),
                        sex = "M") |>
  round(digits = 2)
}
\references{
de Onis M, Garza C, Victora CG, Onyango AW, Frongillo EA, Martines J. \strong{The
WHO Multicentre Growth Reference Study: planning, study design, and
methodology} \emph{Food Nutr Bull.} 2004, \strong{25(1 Suppl):S15-26.}
\doi{10.1177/15648265040251s104}

World Health Organisation. \strong{WHO child growth standards:
length/height-for-age, weight-for-age, weight-for-length, weight-for-height
and body mass index-for-age: methods and development.} \emph{Technical report,
WHO, Geneva}, 2006.

World Health Organisation. \strong{WHO child growth standards: head
circumference-for-age, arm circumference-for-age, triceps skinfold-for-age
and subscapular skinfold-for-age: methods and development.} \emph{Technical
report, WHO, Geneva}, 2007.

Cole TJ. \strong{The LMS method for constructing normalized growth standards} \emph{Eur
J Clin Nutr.} 1990, \strong{44(1):45-60.} PMID:
\href{https://pubmed.ncbi.nlm.nih.gov/2354692/}{2354692}
}
