% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_gs.R
\name{who_gs_zscore2value}
\alias{who_gs_zscore2value}
\alias{who_gs_wfa_zscore2value}
\alias{who_gs_bfa_zscore2value}
\alias{who_gs_lhfa_zscore2value}
\alias{who_gs_wfl_zscore2value}
\alias{who_gs_wfh_zscore2value}
\alias{who_gs_hcfa_zscore2value}
\alias{who_gs_acfa_zscore2value}
\alias{who_gs_ssfa_zscore2value}
\alias{who_gs_tsfa_zscore2value}
\alias{who_gs_centile2value}
\alias{who_gs_wfa_centile2value}
\alias{who_gs_bfa_centile2value}
\alias{who_gs_lhfa_centile2value}
\alias{who_gs_wfl_centile2value}
\alias{who_gs_wfh_centile2value}
\alias{who_gs_hcfa_centile2value}
\alias{who_gs_acfa_centile2value}
\alias{who_gs_ssfa_centile2value}
\alias{who_gs_tsfa_centile2value}
\title{Convert z-scores/centiles to values in the WHO Child Growth Standards}
\usage{
who_gs_zscore2value(z, x, sex, acronym)

who_gs_wfa_zscore2value(z, age_days, sex)

who_gs_bfa_zscore2value(z, age_days, sex)

who_gs_lhfa_zscore2value(z, age_days, sex)

who_gs_wfl_zscore2value(z, length_cm, sex)

who_gs_wfh_zscore2value(z, height_cm, sex)

who_gs_hcfa_zscore2value(z, age_days, sex)

who_gs_acfa_zscore2value(z, age_days, sex)

who_gs_ssfa_zscore2value(z, age_days, sex)

who_gs_tsfa_zscore2value(z, age_days, sex)

who_gs_centile2value(p, x, sex, acronym)

who_gs_wfa_centile2value(p, age_days, sex)

who_gs_bfa_centile2value(p, age_days, sex)

who_gs_lhfa_centile2value(p, age_days, sex)

who_gs_wfl_centile2value(p, length_cm, sex)

who_gs_wfh_centile2value(p, height_cm, sex)

who_gs_hcfa_centile2value(p, age_days, sex)

who_gs_acfa_centile2value(p, age_days, sex)

who_gs_ssfa_centile2value(p, age_days, sex)

who_gs_tsfa_centile2value(p, age_days, sex)
}
\arguments{
\item{z, p}{Numeric vector of centiles/z-scores to convert to values.}

\item{x, age_days, length_cm, height_cm}{Numeric vector with \code{x} values at which
to convert inputs. Must be within bounds of available \code{x} values for given
acronym. The standard-specific versions of each function specify \code{x}:
either \code{age_days} (age since birth in days), \code{length_cm} (recumbent length
measurement(s) in cm) or \code{height_cm} (standing height measurement(s) in
cm).}

\item{sex}{Character vector of sex(es), either \code{"M"} (male) or \code{"F"}
(female).}

\item{acronym}{Valid acronym(s) for the WHO Child Growth standards: either
\code{"wfa"} (weight-for-age), \code{"bfa"} (BMI-for-age), \code{"lhfa"}
(length/height-for-age), \code{"wfl"} (weight-for-length), \code{"wfh"}
(weight-for-height), \code{"hcfa"} (head circumference-for-age), \code{"acfa"}
(arm circumference-for-age), \code{"ssfa"} (subscapular skinfold-for-age), or
\code{"tsfa"} (triceps skinfold-for-age).}
}
\value{
Numeric vector of expected measurements with length equal to the
outputs from \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}.
}
\description{
Convert z-scores/centiles to values in the WHO Child Growth Standards
}
\note{
Input vectors are recycled by \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}.
}
\examples{
# Convert centiles to values
p <- 0.25 # 25th centile
who_gs_centile2value(p = p, x = 501, sex = "M", acronym = "wfa") |>
  round(digits = 2)

# Or z-scores to values
z <- qnorm(p) # Z-score equivalent of 25th centile
who_gs_zscore2value(z = z, x = 501, sex = "M", acronym = "wfa") |>
  round(digits = 2)

# Specify which standard to use with the acronym parameter...
who_gs_zscore2value(z = z, x = 300, sex = "M", acronym = "lhfa") |>
  round(digits = 2)

# ... or by using a standard-specific function
who_gs_lhfa_zscore2value(z = z, age_days = 300, sex = "M") |>
  round(digits = 2)

# Inputs are recycled to the input of the longest length
who_gs_ssfa_centile2value(p = seq(0.1, 0.9, by = 0.2),
                          age_days = 670,
                          sex = "M") |>
  round(digits = 2)

# Bad inputs will not stop the function but will instead return NA - here 60
# cm for height_cm is incompatible with the WHO Growth Standard for
# weight-for-height
who_gs_wfh_zscore2value(z = 0,
                        height_cm = c(60, 85, 105, 120),
                        sex = "M") |>
  round(digits = 2)
}
\references{
de Onis M, Garza C, Victora CG, Onyango AW, Frongillo EA, Martines J. \strong{The
WHO Multicentre Growth Reference Study: planning, study design, and
methodology} \emph{Food Nutr Bull.} 2004, \strong{25(1 Suppl):S15-26.}
\doi{10.1177/15648265040251s104}

World Health Organisation. \strong{WHO child growth standards:
length/height-for-age, weight-for-age, weight-for-length, weight-for-height
and body mass index-for-age: methods and development.} \emph{Technical report,
WHO, Geneva}, 2006.

World Health Organisation. \strong{WHO child growth standards: head
circumference-for-age, arm circumference-for-age, triceps skinfold-for-age
and subscapular skinfold-for-age: methods and development.} \emph{Technical
report, WHO, Geneva}, 2007.

Cole TJ. \strong{The LMS method for constructing normalized growth standards} \emph{Eur
J Clin Nutr.} 1990, \strong{44(1):45-60.} PMID:
\href{https://pubmed.ncbi.nlm.nih.gov/2354692/}{2354692}
}
